Torch.GraduallyTyped.DType

Documentation

data DType

Constructors

Bool
Bool

UInt8
Byte

Int8
Char

Int16
Short

Int32
Int

Int64
Long

Half
Half

Float
Float

Double
Double

ComplexHalf
ComplexHalf

ComplexFloat
ComplexFloat

ComplexDouble
ComplexDouble

QInt8
QInt8

QUInt8
QUInt8

QInt32
QInt32

BFloat16
BFloat16

type BoolSym0 = 'Bool :: DType

type UInt8Sym0 = 'UInt8 :: DType

type Int8Sym0 = 'Int8 :: DType

type Int16Sym0 = 'Int16 :: DType

type Int32Sym0 = 'Int32 :: DType

type Int64Sym0 = 'Int64 :: DType

type HalfSym0 = 'Half :: DType

type FloatSym0 = 'Float :: DType

type DoubleSym0 = 'Double :: DType

type ComplexHalfSym0 = 'ComplexHalf :: DType

type ComplexFloatSym0 = 'ComplexFloat :: DType

type ComplexDoubleSym0 = 'ComplexDouble :: DType

type QInt8Sym0 = 'QInt8 :: DType

type QUInt8Sym0 = 'QUInt8 :: DType

type QInt32Sym0 = 'QInt32 :: DType

type BFloat16Sym0 = 'BFloat16 :: DType

data SDType z where

Constructors

SBool :: SDType ('Bool :: DType)

SUInt8 :: SDType ('UInt8 :: DType)

SInt8 :: SDType ('Int8 :: DType)

SInt16 :: SDType ('Int16 :: DType)

SInt32 :: SDType ('Int32 :: DType)

SInt64 :: SDType ('Int64 :: DType)

SHalf :: SDType ('Half :: DType)

SFloat :: SDType ('Float :: DType)

SDouble :: SDType ('Double :: DType)

SComplexHalf :: SDType ('ComplexHalf :: DType)

SComplexFloat :: SDType ('ComplexFloat :: DType)

SComplexDouble :: SDType ('ComplexDouble :: DType)

SQInt8 :: SDType ('QInt8 :: DType)

SQUInt8 :: SDType ('QUInt8 :: DType)

SQInt32 :: SDType ('QInt32 :: DType)

SBFloat16 :: SDType ('BFloat16 :: DType)

class KnownDType (dType :: DType) where

Methods

dTypeVal :: DType

data DataType (dType :: Type) where

Data type to represent whether or not the tensor data type is checked, that is, known to the compiler.

Constructors

UncheckedDataType :: forall dType. DataType dType
The tensor data type is unknown to the compiler.

DataType :: forall dType. dType -> DataType dType
The tensor data type is known to the compiler.

data SDataType (dataType :: DataType DType) where

Constructors

SUncheckedDataType :: DType -> SDataType 'UncheckedDataType

SDataType :: forall dType. SDType dType -> SDataType ('DataType dType)

class KnownDataType (dataType :: DataType DType) where

Methods

dataTypeVal :: DataType DType

type family GetDataTypes f where ...

Equations

GetDataTypes (a :: DataType DType) = '[a]

GetDataTypes (f g) = Concat (GetDataTypes f) (GetDataTypes g)

GetDataTypes _ = '[]