Torch.GraduallyTyped.Examples.TwoLayerNetwork

Documentation

sinc :: Floating a => a -> a

Compute the sine cardinal (sinc) function, see https://mathworld.wolfram.com/SincFunction.html.

noisySinc :: (Floating a, Random a, RandomGen g) => a -> a -> g -> (a, g)

Compute the sine cardinal (sinc) function and add normally distributed noise of strength epsilon. We use the normal function from Normal which requires a random number generator with RandomGen instance.

data SincData

Datatype to represent a dataset of sine cardinal (sinc) inputs and outputs. The name field is used to identify the split of the dataset.

Constructors

SincData

Fields

name :: Text

unSincData :: [(Float, Float)]

mkSincData

Arguments

:: (RandomGen g, Monad m)

=> Text
name of the dataset

-> Int
number of samples

-> StateT g m SincData
dataset in the state monad over the random generator

Create a dataset of noisy sine cardinal (sinc) values of a desired size.

data TwoLayerNetwork fstLayer activation dropout sndLayer

Data type to represent a simple two-layer neural network. It is a product type of two layer types, fstLayer and sndLayer, an activation function, activation, and a dropout layer, dropout.

Constructors

TwoLayerNetwork

Fields

tlnFstLayer :: fstLayer

tlnActivation :: activation

tlnDropout :: dropout

tlnSndLayer :: sndLayer

type TwoLayerNetworkF gradient hasDropout device dataType inputDim outputDim hiddenDim = NamedModel (TwoLayerNetwork (TLNFstLayerF gradient device dataType inputDim hiddenDim) TLNActivationF (TLNDropoutF hasDropout) (TLNSndLayerF gradient device dataType outputDim hiddenDim))

Specifies the type of the two-layer network.

type TLNFstLayerF gradient device dataType inputDim hiddenDim = NamedModel (GLinear (NamedModel (LinearWeightF gradient device dataType inputDim hiddenDim)) (NamedModel (LinearBiasF 'WithBias gradient device dataType hiddenDim)))

Specifies the type of the first layer of the neural network.

type TLNActivationF = Tanh

Specifies the type of the activation function

type family TLNDropoutF hasDropout where ...

Specifies the type of the dropout layer

Equations

TLNDropoutF 'WithDropout = Dropout

TLNDropoutF 'WithoutDropout = ()

type TLNSndLayerF gradient device dataType outputDim hiddenDim = NamedModel (GLinear (NamedModel (LinearWeightF gradient device dataType hiddenDim outputDim)) (NamedModel (LinearBiasF 'WithBias gradient device dataType outputDim)))

Specifies the type of the second layer of the neural network.

twoLayerNetworkSpec

Arguments

:: forall gradient hasDropout device dataType inputDim outputDim hiddenDim. SGradient gradient
whether or not to compute gradients for the parametrs

-> SHasDropout hasDropout
whether or not to use dropout

-> SDevice device
which device to use

-> SDataType dataType
which data type to use

-> SDim inputDim
input dimension

-> SDim outputDim
output dimension

-> SDim hiddenDim
hidden dimension

-> Double
dropout rate

-> ModelSpec (TwoLayerNetworkF gradient hasDropout device dataType inputDim outputDim hiddenDim)
specification for the network

Creates a value that specifies the parameters of a two-layer neural network.

data AdamOptions

Options for the Adam optimizer.

Constructors

AdamOptions

Fields

learningRate :: Double

learning rate

beta1 :: Double

beta1

beta2 :: Double

beta2

epsilon :: Double

epsilon

weightDecay :: Double

weight decay

amsgrad :: Bool

use amsgrad

defaultAdamOptions :: AdamOptions

Default Adam options.

data Optimizer model where

Optimizer data type.

Constructors

UnsafeOptimizer

Fields

:: forall model. { optimizerStateDictKeys :: [StateDictKey]

, optimizerPtr :: ForeignPtr Optimizer

} -> Optimizer model

getStateDict :: forall model. Optimizer model -> IO StateDict

Get the model state dictionary from an optimizer.

getModel :: forall model. HasStateDict model => ModelSpec model -> Optimizer model -> IO model

Extract a model from an optimizer.

mkAdam

Arguments

:: forall model. HasStateDict model

=> AdamOptions
Adam options

-> model
initial model

-> IO (Optimizer model)
Adam optimizer

Create a new Adam optimizer from a model.

mseLoss

Arguments

:: forall m gradient layout device dataType shape gradient' layout' device' dataType' shape'. (MonadThrow m, Catch (shape <+> shape'))

=> Tensor gradient layout device dataType shape
prediction tensor

-> Tensor gradient' layout' device' dataType' shape'
target tensor

-> m (Tensor (gradient <|> gradient') (layout <+> layout') (device <+> device') (dataType <+> dataType') ('Shape '[]))
output tensor

Compute the mean squared error between two tensors.

stepWithGenerator

Arguments

:: (HasStateDict model, SGetGeneratorDevice generatorDevice, SGetGeneratorDevice generatorOutputDevice, Catch (lossShape <+> 'Shape '[]), Catch (lossGradient <+> 'Gradient 'WithGradient))

=> Optimizer model
optimizer for the model

-> ModelSpec model
model specification

-> (model -> Generator generatorDevice -> IO (Tensor lossGradient lossLayout lossDataType lossDevice lossShape, Generator generatorOutputDevice))
loss function to minimize

-> Generator generatorDevice
random generator

-> IO (Tensor lossGradient lossLayout lossDataType lossDevice lossShape, Generator generatorOutputDevice)
loss and updated generator

Perform one step of optimization.

train

Arguments

:: forall m model input generatorDevice lossGradient lossLayout lossDataType lossDevice lossShape generatorOutputDevice. (MonadIO m, HasStateDict model, HasForward model input generatorDevice (Tensor lossGradient lossLayout lossDataType lossDevice lossShape) generatorOutputDevice, HasForward model input generatorOutputDevice (Tensor lossGradient lossLayout lossDataType lossDevice lossShape) generatorOutputDevice, SGetGeneratorDevice generatorDevice, SGetGeneratorDevice generatorOutputDevice, SGetGradient lossGradient, SGetShape lossShape, Catch (lossShape <+> 'Shape '[]), Catch (lossGradient <+> 'Gradient 'WithGradient))

=> Optimizer model
optimizer for the model

-> ModelSpec model
model specification

-> ListT m input
stream of training examples

-> Generator generatorDevice
random generator

-> m (Either (Generator generatorDevice) (Tensor ('Gradient 'WithoutGradient) lossLayout lossDataType lossDevice ('Shape '[]), Generator generatorOutputDevice))
returned is either the original generator or the average training loss and a new generator

Train the model for one epoch.

eval

Arguments

:: (MonadIO m, HasStateDict model, HasForward model input generatorDevice (Tensor lossGradient lossLayout lossDataType lossDevice lossShape) generatorOutputDevice, HasForward model input generatorOutputDevice (Tensor lossGradient lossLayout lossDataType lossDevice lossShape) generatorOutputDevice, SGetGradient lossGradient, SGetShape lossShape, Catch (lossShape <+> 'Shape '[]), Catch (lossGradient <+> 'Gradient 'WithoutGradient))

=> model
model

-> ListT m input
stream of examples

-> Generator generatorDevice
random generator

-> m (Either (Generator generatorDevice) (Tensor ('Gradient 'WithoutGradient) lossLayout lossDataType lossDevice ('Shape '[]), Generator generatorOutputDevice))
returned is either the original generator or the average evaluation loss and a new generator

Evaluate the model on the given examples.

singleCycleLearningRateSchedule

Arguments

:: Double
peak learning rate after warmup

-> Double
learning rate at the end of the schedule

-> Int
total number of epochs

-> Int
number of warm-up epochs

-> Int
number of cool-down epochs

-> Int
current epoch

-> Double
current learning rate

Single-cycle learning rate schedule. See, for instance, https://arxiv.org/abs/1803.09820.

This is a simple schedule that is a stepwise linear interpolation between the initial, maximum, and final learning rates. The initial learning rate is zero.

data Monitor

Data type for monitoring the training and evaluation losses.

Constructors

TrainingMonitor
monitor for training loss

Fields

mtLoss :: Float

mtEpoch :: Int

EvaluationMonitor
monitor for evaluation loss

Fields

meLoss :: Float

meEpoch :: Int

monitor :: MonadIO m => Consumer Monitor m r

A simple monitor that prints the training and evaluation losses to stdout.

collate :: SDevice device -> Int -> ListT IO (Float, Float) -> _

Collate a stream of examples into batches. The returned batches are of the form (input, target), where (input, target) is a pair of tensors. input and target are both of shape:

Dim (SName @"*") (UncheckedSize batchSize) :|: Dim (SName @"*") (SSize @1) :|: SNil

where batchSize is the number of examples in the batch.

runTwoLayerNetworkExample :: IO ()

Run the two-layer network training loop on a toy dataset.