Torch.GraduallyTyped.LearningRateSchedules

Documentation

singleCycleLearningRateSchedule

Arguments

:: Double
peak learning rate after warmup

-> Double
learning rate at the end of the schedule

-> Int
total number of epochs

-> Int
number of warm-up epochs

-> Int
number of cool-down epochs

-> Int
current epoch

-> Double
current learning rate

Single-cycle learning rate schedule. See, for instance, https://arxiv.org/abs/1803.09820.

This is a simple schedule that is a stepwise linear interpolation between the initial, maximum, and final learning rates. The initial learning rate is zero.