Torch.GraduallyTyped.NN.Class

Documentation

class HasForward model input generatorDevice output generatorOutputDevice | model input generatorDevice -> output, model input generatorDevice -> generatorOutputDevice where

Methods

forward :: forall m. MonadThrow m => model -> input -> Generator generatorDevice -> m (output, Generator generatorOutputDevice)

forward m i g for a model m, an input i, and a generator g returns the tuple (o, g') where o is the output of the model applied to the input and g' is the updated generator. forward m i g may throw an exception if the input i or the generator g are not compatible with the model m.

defaultOptions {genQuotedDecs = True, genSingKindInsts = False} $ genSingletons [''SomeModelSpecs])

type family ModelSpec model = (spec :: Type) | spec -> model

class HasInitialize (model :: Type) (generatorDevice :: Device (DeviceType Nat)) (output :: Type) (generatorOutputDevice :: Device (DeviceType Nat)) | model generatorDevice -> output, model generatorDevice -> generatorOutputDevice where

Methods

initialize :: forall m. MonadThrow m => ModelSpec model -> Generator generatorDevice -> m (output, Generator generatorOutputDevice)

data VectorSpec (n :: Nat) (a :: Type) where

Constructors

VectorSpec :: forall n a. SNat n -> Vector n (ModelSpec a) -> VectorSpec n a

type StateDictKey = String

type StateDict = Map StateDictKey (ForeignPtr Tensor)

newtype FromStateDictError

Constructors

FromStateDictKeyNotFoundError

Fields

fsdeExpectedKey :: StateDictKey

newtype ToStateDictError

Constructors

ToStateDictKeyAlreadyInUseError

Fields

fsdeTakenKey :: StateDictKey

class HasStateDict model where

Methods

fromStateDict :: forall m. (MonadThrow m, MonadState StateDict m) => ModelSpec model -> StateDictKey -> m model

toStateDict :: forall m. (MonadThrow m, MonadState StateDict m) => StateDictKey -> model -> m ()

stateDictFromPretrained :: FilePath -> IO StateDict