Torch.GraduallyTyped.NN.Dropout

Documentation

newtype Dropout where

Given a random generator, randomly zeroes some of the elements of the input tensor with probability p using samples from a Bernoulli distribution. Each channel will be zeroed out independently on every forward call.

Constructors

Dropout

Fields

:: Double

probability of an element to be zeroed

-> Dropout