Torch.GraduallyTyped.NN.Functional.Dropout

Documentation

>>> import Data.Singletons.Prelude.List (SList (..))
>>> import Torch.GraduallyTyped

dropout

Arguments

:: forall gradient layout device dataType shape generatorDevice m. MonadThrow m

=> Double
probability of an element to be zeroed

-> Tensor gradient layout device dataType shape
input

-> Generator generatorDevice
generator

-> m (Tensor gradient layout (device <+> generatorDevice) dataType shape, Generator (device <+> generatorDevice))
output

Dropout randomly zeroes some of the elements of the input tensor with probability p using samples from a Bernoulli distribution.