Torch.GraduallyTyped.NN.Transformer.Block

Documentation

data TransformerBlock (style :: TransformerStyle) (gradient :: Gradient RequiresGradient) (device :: Device (DeviceType Nat)) (dataType :: DataType DType) (headDim :: Dim (Name Symbol) (Size Nat)) (headEmbedDim :: Dim (Name Symbol) (Size Nat)) (embedDim :: Dim (Name Symbol) (Size Nat)) (queryEmbedDim :: Dim (Name Symbol) (Size Nat)) (ffnDim :: Dim (Name Symbol) (Size Nat)) (dropoutP :: Type) where

Transformer encoder block consisting of self-attention and a feed-forward network.

TODO: Some transformers use LayerDrop, see https://arxiv.org/abs/1909.11556, during training. To support this, we will need a layer wrapper that is either the identity function or the wrapped layer based on a uniformly random draw from a supplied generator. Complications will arise due to the gradual typing...

Constructors

TransformerBlock

Fields

:: forall style gradient device dataType headDim headEmbedDim embedDim queryEmbedDim ffnDim dropoutP. { tbSelfAttention :: SelfAttention style gradient device dataType headDim headEmbedDim embedDim queryEmbedDim dropoutP

self-attention layer

, tbFeedForwardNetwork :: TransformerFeedForwardNetwork style gradient device dataType queryEmbedDim ffnDim dropoutP

feed-forward network

} -> TransformerBlock style gradient device dataType headDim headEmbedDim embedDim queryEmbedDim ffnDim dropoutP

testBlock :: IO (Tensor ('Gradient 'WithGradient) ('Layout 'Dense) ('Device ('CPU :: DeviceType Nat)) ('DataType 'Float) ('Shape '['Dim ('Name "*") ('Size 3), 'Dim ('Name "*") ('Size 17), 'Dim ('Name "*") ('Size 512)]))