Torch.GraduallyTyped.NN.Transformer.GBlock

Documentation

data GTransformerBlock (selfAttention :: Type) (crossAttention :: Type) (feedForwardNetwork :: Type) where

Generic transformer encoder block consisting of self-attention, cross-attention, and a feed-forward network.

selfAttention is a self-attention layer.

crossAttention is a cross-attention layer.

feedForwardNetwork is a feed-forward layer.

TODO: Some transformers use LayerDrop, see https://arxiv.org/abs/1909.11556, during training. To support this, we will need a layer wrapper that is either the identity function or the wrapped layer based on a uniformly random draw from a supplied generator.

Constructors

GTransformerBlock

Fields

:: forall selfAttention crossAttention feedForwardNetwork. { tbSelfAttention :: selfAttention

self-attention layer

, tbCrossAttention :: crossAttention

cross-attention layer

, tbFeedForwardNetwork :: feedForwardNetwork

feed-forward network

} -> GTransformerBlock selfAttention crossAttention feedForwardNetwork

type family EncoderBlockSelfAttentionF (style :: TransformerStyle) (gradient :: Gradient RequiresGradient) (device :: Device (DeviceType Nat)) (dataType :: DataType DType) (headDim :: Dim (Name Symbol) (Size Nat)) (headEmbedDim :: Dim (Name Symbol) (Size Nat)) (embedDim :: Dim (Name Symbol) (Size Nat)) (queryEmbedDim :: Dim (Name Symbol) (Size Nat)) :: Type where ...

Equations

EncoderBlockSelfAttentionF style gradient device dataType headDim headEmbedDim embedDim queryEmbedDim = NamedModel (GSelfAttention (SAInitialLayerNormF style gradient device dataType queryEmbedDim) (SAMultiheadAttentionF style gradient device dataType headDim headEmbedDim embedDim queryEmbedDim) (SADropoutF style) (SAFinalLayerNormF style gradient device dataType queryEmbedDim))

type family EncoderBlockCrossAttentionF :: Type where ...

Equations

EncoderBlockCrossAttentionF = ()

type family EncoderBlockFeedForwardNetworkF (style :: TransformerStyle) (gradient :: Gradient RequiresGradient) (device :: Device (DeviceType Nat)) (dataType :: DataType DType) (queryEmbedDim :: Dim (Name Symbol) (Size Nat)) (ffnDim :: Dim (Name Symbol) (Size Nat)) :: Type where ...

Equations

EncoderBlockFeedForwardNetworkF style gradient device dataType queryEmbedDim ffnDim = NamedModel (GTransformerFeedForwardNetwork (FFNInputLayerNormF style gradient device dataType queryEmbedDim) (FFNInputTransformationF style gradient device dataType queryEmbedDim ffnDim) (FFNActivationF style) (FFNActivationDropoutF style) (FFNOutputProjectionF style gradient device dataType queryEmbedDim ffnDim) (FFNOutputDropoutF style) (FFNOutputLayerNormF style gradient device dataType queryEmbedDim))

encoderBlockSpec :: forall style gradient device dataType headDim headEmbedDim embedDim queryEmbedDim ffnDim. STransformerStyle style -> SGradient gradient -> SDevice device -> SDataType dataType -> SDim headDim -> SDim headEmbedDim -> SDim embedDim -> SDim queryEmbedDim -> SDim ffnDim -> Double -> Double -> ModelSpec (GTransformerBlock (EncoderBlockSelfAttentionF style gradient device dataType headDim headEmbedDim embedDim queryEmbedDim) EncoderBlockCrossAttentionF (EncoderBlockFeedForwardNetworkF style gradient device dataType queryEmbedDim ffnDim))

type family DecoderBlockSelfAttentionF (style :: TransformerStyle) (gradient :: Gradient RequiresGradient) (device :: Device (DeviceType Nat)) (dataType :: DataType DType) (headDim :: Dim (Name Symbol) (Size Nat)) (headEmbedDim :: Dim (Name Symbol) (Size Nat)) (embedDim :: Dim (Name Symbol) (Size Nat)) (queryEmbedDim :: Dim (Name Symbol) (Size Nat)) :: Type where ...

Equations

DecoderBlockSelfAttentionF style gradient device dataType headDim headEmbedDim embedDim queryEmbedDim = NamedModel (GSelfAttention (SAInitialLayerNormF style gradient device dataType queryEmbedDim) (SAMultiheadAttentionF style gradient device dataType headDim headEmbedDim embedDim queryEmbedDim) (SADropoutF style) (SAFinalLayerNormF style gradient device dataType queryEmbedDim))

type family DecoderBlockCrossAttentionF (style :: TransformerStyle) (gradient :: Gradient RequiresGradient) (device :: Device (DeviceType Nat)) (dataType :: DataType DType) (headDim :: Dim (Name Symbol) (Size Nat)) (headEmbedDim :: Dim (Name Symbol) (Size Nat)) (embedDim :: Dim (Name Symbol) (Size Nat)) (queryEmbedDim :: Dim (Name Symbol) (Size Nat)) (keyEmbedDim :: Dim (Name Symbol) (Size Nat)) :: Type where ...

Equations

DecoderBlockCrossAttentionF style gradient device dataType headDim headEmbedDim embedDim queryEmbedDim keyEmbedDim = NamedModel (GCrossAttention (CAInitialLayerNormF style gradient device dataType queryEmbedDim) (CAMultiheadAttentionF style gradient device dataType headDim headEmbedDim embedDim queryEmbedDim keyEmbedDim) (CADropoutF style) (CAFinalLayerNormF style gradient device dataType queryEmbedDim))

type family DecoderBlockFeedForwardNetworkF (style :: TransformerStyle) (gradient :: Gradient RequiresGradient) (device :: Device (DeviceType Nat)) (dataType :: DataType DType) (queryEmbedDim :: Dim (Name Symbol) (Size Nat)) (ffnDim :: Dim (Name Symbol) (Size Nat)) :: Type where ...

Equations

DecoderBlockFeedForwardNetworkF style gradient device dataType queryEmbedDim ffnDim = NamedModel (GTransformerFeedForwardNetwork (FFNInputLayerNormF style gradient device dataType queryEmbedDim) (FFNInputTransformationF style gradient device dataType queryEmbedDim ffnDim) (FFNActivationF style) (FFNActivationDropoutF style) (FFNOutputProjectionF style gradient device dataType queryEmbedDim ffnDim) (FFNOutputDropoutF style) (FFNOutputLayerNormF style gradient device dataType queryEmbedDim))

decoderBlockSpec :: forall style gradient device dataType headDim headEmbedDim embedDim queryEmbedDim keyEmbedDim ffnDim. STransformerStyle style -> SGradient gradient -> SDevice device -> SDataType dataType -> SDim headDim -> SDim headEmbedDim -> SDim embedDim -> SDim queryEmbedDim -> SDim keyEmbedDim -> SDim ffnDim -> Double -> Double -> ModelSpec (GTransformerBlock (DecoderBlockSelfAttentionF style gradient device dataType headDim headEmbedDim embedDim queryEmbedDim) (DecoderBlockCrossAttentionF style gradient device dataType headDim headEmbedDim embedDim queryEmbedDim keyEmbedDim) (DecoderBlockFeedForwardNetworkF style gradient device dataType queryEmbedDim ffnDim))