Torch.GraduallyTyped.NN.Transformer.GCrossAttention

Documentation

data GCrossAttention (initialLayerNorm :: Type) (mha :: Type) (dropout :: Type) (finalLayerNorm :: Type) where

Generic cross-attention layer data type.

initialLayerNorm: the initial layer normalization

mha: the multi-headed attention layer

dropout: the dropout layer

finalLayerNorm: the final layer normalization

Constructors

GCrossAttention

Fields

:: forall initialLayerNorm mha dropout finalLayerNorm. { caInitialLayerNorm :: initialLayerNorm

initial layer normalization of the cross-attention layer.

, caMultiHeadAttention :: mha

multi-headed attention layer specialized for cross-attention.

, caDropout :: dropout

dropout

, caFinalLayerNorm :: finalLayerNorm

final layer normalization of the cross-attention layer.

} -> GCrossAttention initialLayerNorm mha dropout finalLayerNorm

type family CAInitialLayerNormF (style :: TransformerStyle) (gradient :: Gradient RequiresGradient) (device :: Device (DeviceType Nat)) (dataType :: DataType DType) (queryEmbedDim :: Dim (Name Symbol) (Size Nat)) :: Type where ...

Specifies the initial layer normalization of the cross-attention layer.

Equations

CAInitialLayerNormF 'T5 gradient device dataType queryEmbedDim = NamedModel (LayerNorm 'WithoutBias gradient device dataType ('Shape '[queryEmbedDim]))

CAInitialLayerNormF 'ByT5 gradient device dataType queryEmbedDim = CAInitialLayerNormF 'T5 gradient device dataType queryEmbedDim

CAInitialLayerNormF 'BART _ _ _ _ = ()

CAInitialLayerNormF 'MBART gradient device dataType queryEmbedDim = CAInitialLayerNormF 'BART gradient device dataType queryEmbedDim

CAInitialLayerNormF 'Pegasus gradient device dataType queryEmbedDim = NamedModel (LayerNorm 'WithBias gradient device dataType ('Shape '[queryEmbedDim]))

type family CAMultiheadAttentionF (style :: TransformerStyle) (gradient :: Gradient RequiresGradient) (device :: Device (DeviceType Nat)) (dataType :: DataType DType) (headDim :: Dim (Name Symbol) (Size Nat)) (headEmbedDim :: Dim (Name Symbol) (Size Nat)) (embedDim :: Dim (Name Symbol) (Size Nat)) (queryEmbedDim :: Dim (Name Symbol) (Size Nat)) (keyEmbedDim :: Dim (Name Symbol) (Size Nat)) :: Type where ...

Specifies the multi-headed attention layer specialized for cross-attention.

Equations

CAMultiheadAttentionF style gradient device dataType headDim headEmbedDim embedDim queryEmbedDim keyEmbedDim = NamedModel (GMultiHeadAttention headDim headEmbedDim embedDim (QInProjF style gradient device dataType queryEmbedDim embedDim) (KInProjF style gradient device dataType keyEmbedDim embedDim) (VInProjF style gradient device dataType keyEmbedDim embedDim) (OutProjF style gradient device dataType embedDim queryEmbedDim) (DropoutF style))

type family CADropoutF (style :: TransformerStyle) :: Type where ...

Specifies the dropout layer of the cross-attention layer.

Equations

CADropoutF _ = Dropout

type family CAFinalLayerNormF (style :: TransformerStyle) (gradient :: Gradient RequiresGradient) (device :: Device (DeviceType Nat)) (dataType :: DataType DType) (queryEmbedDim :: Dim (Name Symbol) (Size Nat)) :: Type where ...

Specifies the final layer normalization of the cross-attention layer.

Equations

CAFinalLayerNormF 'T5 _ _ _ _ = ()

CAFinalLayerNormF 'ByT5 gradient device dataType queryEmbedDim = CAFinalLayerNormF 'T5 gradient device dataType queryEmbedDim

CAFinalLayerNormF 'BART gradient device dataType queryEmbedDim = NamedModel (LayerNorm 'WithBias gradient device dataType ('Shape '[queryEmbedDim]))

CAFinalLayerNormF 'MBART gradient device dataType queryEmbedDim = CAFinalLayerNormF 'BART gradient device dataType queryEmbedDim

CAFinalLayerNormF 'Pegasus gradient device dataType queryEmbedDim = ()

crossAttentionSpec :: forall style gradient device dataType headDim headEmbedDim embedDim queryEmbedDim keyEmbedDim. STransformerStyle style -> SGradient gradient -> SDevice device -> SDataType dataType -> SDim headDim -> SDim headEmbedDim -> SDim embedDim -> SDim queryEmbedDim -> SDim keyEmbedDim -> Double -> Double -> ModelSpec (GCrossAttention (CAInitialLayerNormF style gradient device dataType queryEmbedDim) (CAMultiheadAttentionF style gradient device dataType headDim headEmbedDim embedDim queryEmbedDim keyEmbedDim) (CADropoutF style) (CAFinalLayerNormF style gradient device dataType queryEmbedDim))

Specifies the parameters of a cross-attention layer.

style: the style of the transformer stack, e.g. ST5, SByT5, etc.

gradient: whether to compute the gradient of the stack's parameters.

device: the computational device on which the stack is allocated.

dataType: the data type of the stack's parameters.

headDim: the dimension of all transformer heads in the stack.

headEmbedDim: the dimension of the transformer head embeddings.

embedDim: the dimension of the transformer embeddings.

queryEmbedDim: the dimension of the transformer query embeddings.

keyEmbedDim: the dimension of the transformer key embeddings.

dropoutP: the dropout rate.

eps: the epsilon value for numerical stability of the layer normalization.