Torch.GraduallyTyped.NN.Transformer.GEncoderDecoder

Documentation

data EncoderDecoderTransformerHasEmbedScaling

Data type that is used to represent whether the encoder-decoder transformer model has a scaled embedding.

Constructors

EncoderDecoderTransformerWithEmbedScaling

EncoderDecoderTransformerWithoutEmbedScaling

data GEncoderDecoderTransformer (inputEmbedDim :: Dim (Name Symbol) (Size Nat)) (encoder :: Type) (decoder :: Type) (sharedEmbedding :: Type) (head :: Type) where

Generic encoder-decoder transformer model. This is a model that can be used to encode and decode sequences of variable length.

inputEmbedDim: the dimension of the input embedding.

encoder: a transformer encoder.

decoder: a transformer decoder.

sharedEmbedding: a shared embedding layer.

head: a head layer for the output.

Constructors

GEncoderDecoderTransformer

Fields

:: forall inputEmbedDim encoder decoder sharedEmbedding head. { edtInputEmbedDim :: SDim inputEmbedDim

input embedding dim for scaling

, edtEncoder :: encoder

encoder

, edtDecoder :: decoder

decoder

, edtSharedEmbedding :: sharedEmbedding

embedding shared between encoder and decoder

, edtHead :: head

transformer head

, edtEmbedScaling :: EncoderDecoderTransformerHasEmbedScaling

embedding scaling

} -> GEncoderDecoderTransformer inputEmbedDim encoder decoder sharedEmbedding head

type family EDTEncoderF (style :: TransformerStyle) (numEncoderLayers :: Nat) (gradient :: Gradient RequiresGradient) (device :: Device (DeviceType Nat)) (dataType :: DataType DType) (headDim :: Dim (Name Symbol) (Size Nat)) (headEmbedDim :: Dim (Name Symbol) (Size Nat)) (embedDim :: Dim (Name Symbol) (Size Nat)) (inputEmbedDim :: Dim (Name Symbol) (Size Nat)) (ffnDim :: Dim (Name Symbol) (Size Nat)) (posEncDim :: Dim (Name Symbol) (Size Nat)) where ...

Specifies the encoder of the encoder-decoder transformer model.

Equations

EDTEncoderF style numEncoderLayers gradient device dataType headDim headEmbedDim embedDim inputEmbedDim ffnDim posEncDim = NamedModel (GTransformer (TEPosEncF style gradient device dataType inputEmbedDim posEncDim) (TERelPosEncF style gradient device dataType headDim posEncDim) (TEInitialLayerNormF style gradient device dataType inputEmbedDim) (TEInitialDropoutF style) (TEStackF style numEncoderLayers gradient device dataType headDim headEmbedDim embedDim inputEmbedDim ffnDim) (TEFinalLayerNormF style gradient device dataType inputEmbedDim) (TEFinalDropoutF style))

type family EDTDecoderF (style :: TransformerStyle) (numDecoderLayers :: Nat) (gradient :: Gradient RequiresGradient) (device :: Device (DeviceType Nat)) (dataType :: DataType DType) (headDim :: Dim (Name Symbol) (Size Nat)) (headEmbedDim :: Dim (Name Symbol) (Size Nat)) (embedDim :: Dim (Name Symbol) (Size Nat)) (inputEmbedDim :: Dim (Name Symbol) (Size Nat)) (ffnDim :: Dim (Name Symbol) (Size Nat)) (posEncDim :: Dim (Name Symbol) (Size Nat)) where ...

Specifies the decoder of the encoder-decoder transformer model.

Equations

EDTDecoderF style numDecoderLayers gradient device dataType headDim headEmbedDim embedDim inputEmbedDim ffnDim posEncDim = NamedModel (GTransformer (TDPosEncF style gradient device dataType inputEmbedDim posEncDim) (TDRelPosEncF style gradient device dataType headDim posEncDim) (TDInitialLayerNormF style gradient device dataType inputEmbedDim) (TDInitialDropoutF style) (TDStackF style numDecoderLayers gradient device dataType headDim headEmbedDim embedDim inputEmbedDim inputEmbedDim ffnDim) (TDFinalLayerNormF style gradient device dataType inputEmbedDim) (TDFinalDropoutF style))

type family EDTSharedEmbeddingF (style :: TransformerStyle) (gradient :: Gradient RequiresGradient) (device :: Device (DeviceType Nat)) (dataType :: DataType DType) (inputEmbedDim :: Dim (Name Symbol) (Size Nat)) (vocabDim :: Dim (Name Symbol) (Size Nat)) where ...

Specifies the shared embedding layer of the encoder-decoder transformer model.

Equations

EDTSharedEmbeddingF _ gradient device dataType inputEmbedDim vocabDim = NamedModel (Embedding gradient ('Layout 'Dense) device dataType vocabDim inputEmbedDim 'Nothing)

type family EDTHeadF (style :: TransformerStyle) (transformerHead :: TransformerHead) (gradient :: Gradient RequiresGradient) (device :: Device (DeviceType Nat)) (dataType :: DataType DType) (inputEmbedDim :: Dim (Name Symbol) (Size Nat)) (vocabDim :: Dim (Name Symbol) (Size Nat)) :: Type where ...

Specifies the head of the encoder-decoder transformer model.

Equations

EDTHeadF style 'WithoutHead gradient device dataType inputEmbedDim vocabDim = ()

EDTHeadF style 'WithLMHead gradient device dataType inputEmbedDim vocabDim = NamedModel (GLMHead inputEmbedDim (LMHeadDenseF style gradient device dataType inputEmbedDim) (LMHeadActivationF style) (LMHeadLayerNormF style gradient device dataType inputEmbedDim) (LMHeadDecoderF style gradient device dataType inputEmbedDim vocabDim) (LMHeadBiasF style gradient device dataType vocabDim))

encoderDecoderTransformerSpec :: forall style transformerHead numEncoderLayers numDecoderLayers gradient device dataType headDim headEmbedDim embedDim inputEmbedDim ffnDim posEncDim vocabDim. STransformerStyle style -> STransformerHead transformerHead -> SNat numEncoderLayers -> SNat numDecoderLayers -> SGradient gradient -> SDevice device -> SDataType dataType -> SDim headDim -> SDim headEmbedDim -> SDim embedDim -> SDim inputEmbedDim -> SDim ffnDim -> SDim posEncDim -> SDim vocabDim -> Double -> Double -> ModelSpec (GEncoderDecoderTransformer inputEmbedDim (EDTEncoderF style numEncoderLayers gradient device dataType headDim headEmbedDim embedDim inputEmbedDim ffnDim posEncDim) (EDTDecoderF style numDecoderLayers gradient device dataType headDim headEmbedDim embedDim inputEmbedDim ffnDim posEncDim) (EDTSharedEmbeddingF style gradient device dataType inputEmbedDim vocabDim) (EDTHeadF style transformerHead gradient device dataType inputEmbedDim vocabDim))

Specifies the parameters of an encoder-decoder transformer model.

style: the style of the encoder-decoder transformer model, e.g. ST5, SBART, etc.

transformerHead: the head of the encoder-decoder transformer model.

numEncoderLayers: the number of encoder layers of the encoder-decoder transformer model.

numDecoderLayers: the number of decoder layers of the encoder-decoder transformer model.

gradient: whether to compute the gradient of the model parameters

device: the computational device on which the model is allocated.

dataType: the data type of the model parameters.

headDim: the dimension of all transformer heads in the encoder-decoder transformer model.

headEmbedDim: the dimension of the transformer head embeddings.

embedDim: the dimension of the transformer embeddings.

inputEmbedDim: the dimension of the input embeddings for both the encoder and the decoder.

ffnDim: the dimension of the feed-forward network.

posEncDim: the dimension of the positional embeddings.

vocabDim: the dimension of the vocabulary.

typeVocabDim: the dimension of the type vocabulary.

dropoutP: the dropout rate.

eps: the epsilon value for numerical stability of the layer normalization.

data GSimplifiedEncoderDecoderTransformer (model :: Type) (mkPos :: Type) (mkDecoderPos :: Type) (mkPaddingMask :: Type) (mkAttentionMask :: Type) (mkCrossAttentionMask :: Type) (mkDecoderAttentionMask :: Type) where

Constructors

GSimplifiedEncoderDecoderTransformer

Fields

:: forall model mkPos mkDecoderPos mkPaddingMask mkAttentionMask mkCrossAttentionMask mkDecoderAttentionMask. { sedtModel :: model

encoder-decoder model

, sedtDecoderInputShift :: ShiftRight Int

shift for decoder input

, sedtPaddingMaskShift :: ShiftRight Int

shift for padding mask

, sedtMkPos :: mkPos

make encoder input positions

, sedtMkDecoderPos :: mkDecoderPos

make decoder input position

, sedtMkPaddingMask :: mkPaddingMask

make padding mask

, sedtMkAttentionMask :: mkAttentionMask

make attention mask

, sedtMkCrossAttentionMask :: mkCrossAttentionMask

make cross-attention mask

, sedtMkDecoderAttentionMask :: mkDecoderAttentionMask

make decoder attention mask

} -> GSimplifiedEncoderDecoderTransformer model mkPos mkDecoderPos mkPaddingMask mkAttentionMask mkCrossAttentionMask mkDecoderAttentionMask

data EncoderDecoderTransformerInput input decoderInput pos decoderPos attentionMask decoderAttentionMask crossAttentionMask where

Input data type for use with an encoder-decoder transformer. Use this for training.

Constructors

EncoderDecoderTransformerInput

Fields

:: forall input decoderInput pos decoderPos attentionMask decoderAttentionMask crossAttentionMask. { edtInput :: input

, edtDecoderInput :: decoderInput

, edtPos :: pos

, edtDecoderPos :: decoderPos

, edtAttentionMask :: attentionMask

, edtDecoderAttentionMask :: decoderAttentionMask

, edtCrossAttentionMask :: crossAttentionMask

} -> EncoderDecoderTransformerInput input decoderInput pos decoderPos attentionMask decoderAttentionMask crossAttentionMask

data SimplifiedEncoderDecoderTransformerInput input decoderInput where

Constructors

SimplifiedEncoderDecoderTransformerInput

Fields

:: forall input decoderInput. { sedtInput :: input

, sedtDecoderInput :: decoderInput

} -> SimplifiedEncoderDecoderTransformerInput input decoderInput

data EncoderDecoderTransformerOutput decoderOutput encoderOutput where

Output data type for use with an encoder-decoder transformer.

Constructors

EncoderDecoderTransformerOutput

Fields

:: forall decoderOutput encoderOutput. { edtDecoderOutput :: decoderOutput

, edtEncoderOutput :: encoderOutput

} -> EncoderDecoderTransformerOutput decoderOutput encoderOutput

data SimplifiedEncoderDecoderTransformerOutput decoderOutput encoderOutput inputPaddingMask where

Constructors

SimplifiedEncoderDecoderTransformerOutput

Fields

:: forall decoderOutput encoderOutput inputPaddingMask. { sedtDecoderOutput :: decoderOutput

, sedtEncoderOutput :: encoderOutput

, sedtInputPaddingMask :: inputPaddingMask

} -> SimplifiedEncoderDecoderTransformerOutput decoderOutput encoderOutput inputPaddingMask

data EncoderDecoderTransformerGenerationInput decoderInput encoderOutput decoderPos decoderAttentionMask crossAttentionMask where

Input data type for use with an encoder-decoder transformer. Use this for inference.

Constructors

EncoderDecoderTransformerGenerationInput

Fields

:: forall decoderInput encoderOutput decoderPos decoderAttentionMask crossAttentionMask. { edtGenerationDecoderInput :: decoderInput

, edtGenerationEncoderOutput :: encoderOutput

, edtGenerationDecoderPos :: decoderPos

, edtGenerationDecoderAttentionMask :: decoderAttentionMask

, edtGenerationCrossAttentionMask :: crossAttentionMask

} -> EncoderDecoderTransformerGenerationInput decoderInput encoderOutput decoderPos decoderAttentionMask crossAttentionMask

data SimplifiedEncoderDecoderTransformerGenerationInput decoderInput encoderOutput inputPaddingMask where

Constructors

SimplifiedEncoderDecoderTransformerGenerationInput

Fields

:: forall decoderInput encoderOutput inputPaddingMask. { sedtGenerationDecoderInput :: decoderInput

, sedtGenerationEncoderOutput :: encoderOutput

, sedtGenerationInputPaddingMask :: inputPaddingMask

} -> SimplifiedEncoderDecoderTransformerGenerationInput decoderInput encoderOutput inputPaddingMask

testEncoderDecoderTransformer :: IO _