Torch.GraduallyTyped.NN.Transformer.GFeedForwardNetwork

Documentation

data GGate (layer0 :: Type) (activation :: Type) (layer1 :: Type) where

Generic two-layer gate with activation function.

layer0 is the first layer.

activation is the activation function.

layer1 is the second layer.

Constructors

GGate

Fields

:: forall layer0 activation layer1. { gateLayer0 :: layer0

first gate layer

, gateActivation :: activation

gate activation

, gateLayer1 :: layer1

second gate layer

} -> GGate layer0 activation layer1

data GTransformerFeedForwardNetwork (inputLayerNorm :: Type) (inputTransformation :: Type) (activation :: Type) (activationDropout :: Type) (outputProjection :: Type) (outputDropout :: Type) (outputLayerNorm :: Type) where

Generic transformer feed-forward network.

inputLayerNorm is the layer normalization for the input.

inputTransformation is the input transformation.

activation is the activation function.

activationDropout is the activation dropout layer.

outputProjection is the output projection.

outputDropout is the dropout layer for the output.

outputLayerNorm is the layer normalization for the output.

Constructors

GTransformerFeedForwardNetwork

Fields

:: forall inputLayerNorm inputTransformation activation activationDropout outputProjection outputDropout outputLayerNorm. { ffnInputLayerNorm :: inputLayerNorm

input layer norm

, ffnInputTransformation :: inputTransformation

input transformation

, ffnActivation :: activation

activation

, ffnActivationDropout :: activationDropout

activation dropout

, ffnOutputProjection :: outputProjection

output projection

, ffnOutputDropout :: outputDropout

output dropout

, ffnOutputLayerNorm :: outputLayerNorm

output layer norm

} -> GTransformerFeedForwardNetwork inputLayerNorm inputTransformation activation activationDropout outputProjection outputDropout outputLayerNorm

type family GTransformerFeedForwardNetworkF (style :: TransformerStyle) (gradient :: Gradient RequiresGradient) (device :: Device (DeviceType Nat)) (dataType :: DataType DType) (queryEmbedDim :: Dim (Name Symbol) (Size Nat)) (ffnDim :: Dim (Name Symbol) (Size Nat)) (hasDropout :: HasDropout) :: Type where ...

Equations

GTransformerFeedForwardNetworkF style gradient device dataType queryEmbedDim ffnDim hasDropout = GTransformerFeedForwardNetwork (FFNInputLayerNormF style gradient device dataType queryEmbedDim) (FFNInputTransformationF style gradient device dataType queryEmbedDim ffnDim) (FFNActivationF style) (FFNActivationDropoutF style) (FFNOutputProjectionF style gradient device dataType queryEmbedDim ffnDim) (FFNOutputDropoutF style hasDropout) (FFNOutputLayerNormF style gradient device dataType queryEmbedDim)

type family FFNInputLayerNormF (style :: TransformerStyle) (gradient :: Gradient RequiresGradient) (device :: Device (DeviceType Nat)) (dataType :: DataType DType) (queryEmbedDim :: Dim (Name Symbol) (Size Nat)) :: Type where ...

Specifies the layer normalization for the input.

Equations

FFNInputLayerNormF 'T5 gradient device dataType queryEmbedDim = NamedModel (LayerNorm 'WithoutBias gradient device dataType ('Shape '[queryEmbedDim]))

FFNInputLayerNormF 'ByT5 gradient device dataType queryEmbedDim = FFNInputLayerNormF 'T5 gradient device dataType queryEmbedDim

FFNInputLayerNormF 'BART _ _ _ _ = ()

FFNInputLayerNormF 'MBART gradient device dataType queryEmbedDim = FFNInputLayerNormF 'BART gradient device dataType queryEmbedDim

FFNInputLayerNormF 'Pegasus gradient device dataType queryEmbedDim = NamedModel (LayerNorm 'WithBias gradient device dataType ('Shape '[queryEmbedDim]))

FFNInputLayerNormF 'BERT _ _ _ _ = ()

FFNInputLayerNormF 'RoBERTa gradient device dataType queryEmbedDim = FFNInputLayerNormF 'BERT gradient device dataType queryEmbedDim

type family FFNInputTransformationF (style :: TransformerStyle) (gradient :: Gradient RequiresGradient) (device :: Device (DeviceType Nat)) (dataType :: DataType DType) (queryEmbedDim :: Dim (Name Symbol) (Size Nat)) (ffnDim :: Dim (Name Symbol) (Size Nat)) :: Type where ...

Specifies the first input projection.

Equations

FFNInputTransformationF 'T5 gradient device dataType queryEmbedDim ffnDim = NamedModel (GLinearF 'WithoutBias gradient device dataType queryEmbedDim ffnDim)

FFNInputTransformationF 'ByT5 gradient device dataType queryEmbedDim ffnDim = GGate (NamedModel (GLinearF 'WithoutBias gradient device dataType queryEmbedDim ffnDim)) GeluNew (NamedModel (GLinearF 'WithoutBias gradient device dataType queryEmbedDim ffnDim))

FFNInputTransformationF _ gradient device dataType queryEmbedDim ffnDim = NamedModel (GLinearF 'WithBias gradient device dataType queryEmbedDim ffnDim)

type family FFNActivationF (style :: TransformerStyle) :: Type where ...

Specifies the activation.

Equations

FFNActivationF 'T5 = Relu

FFNActivationF 'ByT5 = GeluNew

FFNActivationF 'BART = Gelu

FFNActivationF 'MBART = Gelu

FFNActivationF 'Pegasus = Relu

FFNActivationF 'BERT = Gelu

FFNActivationF 'RoBERTa = Gelu

type family FFNActivationDropoutF (style :: TransformerStyle) :: Type where ...

Specifies the activation dropout.

Equations

FFNActivationDropoutF 'T5 = Dropout

FFNActivationDropoutF 'ByT5 = FFNActivationDropoutF 'T5

FFNActivationDropoutF 'BART = Dropout

FFNActivationDropoutF 'MBART = FFNActivationDropoutF 'BART

FFNActivationDropoutF 'Pegasus = FFNActivationDropoutF 'BART

FFNActivationDropoutF 'BERT = ()

FFNActivationDropoutF 'RoBERTa = FFNActivationDropoutF 'BERT

type family FFNOutputProjectionF (style :: TransformerStyle) (gradient :: Gradient RequiresGradient) (device :: Device (DeviceType Nat)) (dataType :: DataType DType) (queryEmbedDim :: Dim (Name Symbol) (Size Nat)) (ffnDim :: Dim (Name Symbol) (Size Nat)) :: Type where ...

Specifies the output projection.

Equations

FFNOutputProjectionF 'T5 gradient device dataType queryEmbedDim ffnDim = NamedModel (GLinearF 'WithoutBias gradient device dataType ffnDim queryEmbedDim)

FFNOutputProjectionF 'ByT5 gradient device dataType queryEmbedDim ffnDim = FFNOutputProjectionF 'T5 gradient device dataType queryEmbedDim ffnDim

FFNOutputProjectionF 'BART gradient device dataType queryEmbedDim ffnDim = NamedModel (GLinearF 'WithBias gradient device dataType ffnDim queryEmbedDim)

FFNOutputProjectionF 'MBART gradient device dataType queryEmbedDim ffnDim = FFNOutputProjectionF 'BART gradient device dataType queryEmbedDim ffnDim

FFNOutputProjectionF 'Pegasus gradient device dataType queryEmbedDim ffnDim = FFNOutputProjectionF 'BART gradient device dataType queryEmbedDim ffnDim

FFNOutputProjectionF 'BERT gradient device dataType queryEmbedDim ffnDim = NamedModel (GLinearF 'WithBias gradient device dataType ffnDim queryEmbedDim)

FFNOutputProjectionF 'RoBERTa gradient device dataType queryEmbedDim ffnDim = FFNOutputProjectionF 'BERT gradient device dataType queryEmbedDim ffnDim

type family FFNOutputDropoutF (style :: TransformerStyle) (hasDropout :: HasDropout) :: Type where ...

Specifies the dropout for the output.

Equations

FFNOutputDropoutF _ 'WithDropout = Dropout

FFNOutputDropoutF _ 'WithoutDropout = ()

type family FFNOutputLayerNormF (style :: TransformerStyle) (gradient :: Gradient RequiresGradient) (device :: Device (DeviceType Nat)) (dataType :: DataType DType) (queryEmbedDim :: Dim (Name Symbol) (Size Nat)) :: Type where ...

Specifies the layer normalization for the output.

Equations

FFNOutputLayerNormF 'T5 _ _ _ _ = ()

FFNOutputLayerNormF 'ByT5 gradient device dataType queryEmbedDim = FFNOutputLayerNormF 'T5 gradient device dataType queryEmbedDim

FFNOutputLayerNormF 'BART gradient device dataType queryEmbedDim = NamedModel (LayerNorm 'WithBias gradient device dataType ('Shape '[queryEmbedDim]))

FFNOutputLayerNormF 'MBART gradient device dataType queryEmbedDim = FFNOutputLayerNormF 'BART gradient device dataType queryEmbedDim

FFNOutputLayerNormF 'Pegasus _ _ _ _ = ()

FFNOutputLayerNormF 'BERT gradient device dataType queryEmbedDim = NamedModel (LayerNorm 'WithBias gradient device dataType ('Shape '[queryEmbedDim]))

FFNOutputLayerNormF 'RoBERTa gradient device dataType queryEmbedDim = FFNOutputLayerNormF 'BERT gradient device dataType queryEmbedDim

transformerFeedForwardNetworkSpec :: forall style gradient device dataType queryEmbedDim ffnDim hasDropout. STransformerStyle style -> SGradient gradient -> SDevice device -> SDataType dataType -> SDim queryEmbedDim -> SDim ffnDim -> SHasDropout hasDropout -> Double -> Double -> ModelSpec (GTransformerFeedForwardNetworkF style gradient device dataType queryEmbedDim ffnDim hasDropout)

Specifies the parameters of the transformer feed forward network.

style: the style of the transformer feed forward network, e.g. ST5, SByT5, etc.

gradient: whether to compute the gradient of the network's parameters.

device: the computational device on which the parameters are allocated.

dataType: the data type of the parameters.

queryEmbedDim: the dimension of the query embedding.

ffnDim: the dimension of the feed forward network's hidden state.

dropoutP: the dropout rate.

eps: the epsilon value for numerical stability of the layer normalization.