Torch.GraduallyTyped.NN.Transformer.GLMHead

Documentation

data LMHeadHasScaling

A data type that represents whether or not the language modelling head has a scaled decoder output.

Constructors

LMHeadWithScaling

LMHeadWithoutScaling

data GLMHead (inputEmbedDim :: Dim (Name Symbol) (Size Nat)) (dense :: Type) (activation :: Type) (layerNorm :: Type) (decoder :: Type) (bias :: Type) where

Generic language modelling head for transformer encoders and decoders.

inputEmbedDim is the dimension of the input embedding.

dense is a dense layer.

activation is an activation function.

layerNorm is a layer normalization layer.

decoder is a decoder layer.

bias is a bias layer.

Constructors

GLMHead

Fields

:: forall inputEmbedDim dense activation layerNorm decoder bias. { lmHeadInputEmbedDim :: SDim inputEmbedDim

the dimension of the input embedding.

, lmHeadDense :: dense

the dense layer.

, lmHeadActivation :: activation

the activation function.

, lmHeadLayerNorm :: layerNorm

the layer normalization layer.

, lmHeadDecoder :: decoder

the decoder layer.

, lmHeadBias :: bias

the bias layer.

, lmHeadHasScaling :: LMHeadHasScaling

whether or not the head has a scaled decoder output.

} -> GLMHead inputEmbedDim dense activation layerNorm decoder bias

data GBias (bias :: Type) where

Generic data type for biasing the language model head.

Constructors

GBias :: forall bias. bias -> GBias bias

type family LMHeadDenseF (style :: TransformerStyle) (gradient :: Gradient RequiresGradient) (device :: Device (DeviceType Nat)) (dataType :: DataType DType) (inputEmbedDim :: Dim (Name Symbol) (Size Nat)) :: Type where ...

Specifies the dense layer of the language model head.

Equations

LMHeadDenseF 'T5 _ _ _ _ = ()

LMHeadDenseF 'ByT5 gradient device dataType inputEmbedDim = LMHeadDenseF 'T5 gradient device dataType inputEmbedDim

LMHeadDenseF 'BART _ _ _ _ = ()

LMHeadDenseF 'MBART gradient device dataType inputEmbedDim = LMHeadDenseF 'BART gradient device dataType inputEmbedDim

LMHeadDenseF 'Pegasus gradient device dataType inputEmbedDim = LMHeadDenseF 'BART gradient device dataType inputEmbedDim

LMHeadDenseF 'BERT gradient device dataType inputEmbedDim = NamedModel (GLinear (NamedModel (LinearWeightF gradient device dataType inputEmbedDim inputEmbedDim)) (NamedModel (LinearBiasF 'WithBias gradient device dataType inputEmbedDim)))

LMHeadDenseF 'RoBERTa gradient device dataType inputEmbedDim = LMHeadDenseF 'BERT gradient device dataType inputEmbedDim

type family LMHeadActivationF (style :: TransformerStyle) :: Type where ...

Specifies the activation function of the language model head.

Equations

LMHeadActivationF 'T5 = ()

LMHeadActivationF 'ByT5 = LMHeadActivationF 'T5

LMHeadActivationF 'BART = ()

LMHeadActivationF 'MBART = LMHeadActivationF 'BART

LMHeadActivationF 'Pegasus = LMHeadActivationF 'BART

LMHeadActivationF 'BERT = Gelu

LMHeadActivationF 'RoBERTa = LMHeadActivationF 'BERT

type family LMHeadLayerNormF (style :: TransformerStyle) (gradient :: Gradient RequiresGradient) (device :: Device (DeviceType Nat)) (dataType :: DataType DType) (inputEmbedDim :: Dim (Name Symbol) (Size Nat)) :: Type where ...

Specifies the layer normalization layer of the language model head.

Equations

LMHeadLayerNormF 'T5 _ _ _ _ = ()

LMHeadLayerNormF 'ByT5 gradient device dataType inputEmbedDim = LMHeadLayerNormF 'T5 gradient device dataType inputEmbedDim

LMHeadLayerNormF 'BART _ _ _ _ = ()

LMHeadLayerNormF 'MBART gradient device dataType inputEmbedDim = LMHeadLayerNormF 'BART gradient device dataType inputEmbedDim

LMHeadLayerNormF 'Pegasus gradient device dataType inputEmbedDim = LMHeadLayerNormF 'BART gradient device dataType inputEmbedDim

LMHeadLayerNormF 'BERT gradient device dataType inputEmbedDim = NamedModel (LayerNorm 'WithBias gradient device dataType ('Shape '[inputEmbedDim]))

LMHeadLayerNormF 'RoBERTa gradient device dataType inputEmbedDim = LMHeadLayerNormF 'BERT gradient device dataType inputEmbedDim

type family LMHeadDecoderF (style :: TransformerStyle) (gradient :: Gradient RequiresGradient) (device :: Device (DeviceType Nat)) (dataType :: DataType DType) (inputEmbedDim :: Dim (Name Symbol) (Size Nat)) (vocabDim :: Dim (Name Symbol) (Size Nat)) :: Type where ...

Specifies the decoder layer of the language model head.

Equations

LMHeadDecoderF 'T5 gradient device dataType inputEmbedDim vocabDim = NamedModel (GLinear (NamedModel (LinearWeightF gradient device dataType inputEmbedDim vocabDim)) (NamedModel (LinearBiasF 'WithoutBias gradient device dataType vocabDim)))

LMHeadDecoderF 'ByT5 gradient device dataType inputEmbedDim vocabDim = LMHeadDecoderF 'T5 gradient device dataType inputEmbedDim vocabDim

LMHeadDecoderF 'BART gradient device dataType inputEmbedDim vocabDim = NamedModel (GLinear (NamedModel (LinearWeightF gradient device dataType inputEmbedDim vocabDim)) (NamedModel (LinearBiasF 'WithoutBias gradient device dataType vocabDim)))

LMHeadDecoderF 'MBART gradient device dataType inputEmbedDim vocabDim = LMHeadDecoderF 'BART gradient device dataType inputEmbedDim vocabDim

LMHeadDecoderF 'Pegasus gradient device dataType inputEmbedDim vocabDim = LMHeadDecoderF 'BART gradient device dataType inputEmbedDim vocabDim

LMHeadDecoderF 'BERT gradient device dataType inputEmbedDim vocabDim = NamedModel (GLinear (NamedModel (LinearWeightF gradient device dataType inputEmbedDim vocabDim)) (NamedModel (LinearBiasF 'WithBias gradient device dataType vocabDim)))

LMHeadDecoderF 'RoBERTa gradient device dataType inputEmbedDim vocabDim = LMHeadDecoderF 'BERT gradient device dataType inputEmbedDim vocabDim

type family LMHeadBiasF (style :: TransformerStyle) (gradient :: Gradient RequiresGradient) (device :: Device (DeviceType Nat)) (dataType :: DataType DType) (vocabDim :: Dim (Name Symbol) (Size Nat)) :: Type where ...

Specifies the bias layer of the language model head.

Equations

LMHeadBiasF 'T5 _ _ _ _ = GBias ()

LMHeadBiasF 'ByT5 gradient device dataType vocabDim = LMHeadBiasF 'T5 gradient device dataType vocabDim

LMHeadBiasF 'BART gradient device dataType vocabDim = GBias (NamedModel (Tensor gradient ('Layout 'Dense) device dataType ('Shape '['Dim ('Name "*") ('Size 1), vocabDim])))

LMHeadBiasF 'MBART gradient device dataType vocabDim = LMHeadBiasF 'BART gradient device dataType vocabDim

LMHeadBiasF 'Pegasus gradient device dataType vocabDim = LMHeadBiasF 'BART gradient device dataType vocabDim

LMHeadBiasF 'BERT _ _ _ _ = GBias ()

LMHeadBiasF 'RoBERTa gradient device dataType vocabDim = LMHeadBiasF 'BERT gradient device dataType vocabDim

lmHeadSpec :: forall style gradient device dataType inputEmbedDim vocabDim. STransformerStyle style -> SGradient gradient -> SDevice device -> SDataType dataType -> SDim inputEmbedDim -> SDim vocabDim -> Double -> ModelSpec (GLMHead inputEmbedDim (LMHeadDenseF style gradient device dataType inputEmbedDim) (LMHeadActivationF style) (LMHeadLayerNormF style gradient device dataType inputEmbedDim) (LMHeadDecoderF style gradient device dataType inputEmbedDim vocabDim) (LMHeadBiasF style gradient device dataType vocabDim))

type family LMHeadOutputF (style :: TransformerStyle) (decoderOutput :: Type) (gradient :: Gradient RequiresGradient) (device :: Device (DeviceType Nat)) (dataType :: DataType DType) (vocabDim :: Dim (Name Symbol) (Size Nat)) :: Type where ...

Equations

LMHeadOutputF 'T5 decoderOutput _ _ _ _ = decoderOutput

LMHeadOutputF 'ByT5 decoderOutput gradient device dataType vocabDim = LMHeadOutputF 'T5 decoderOutput gradient device dataType vocabDim

LMHeadOutputF 'BART (Tensor gradient' layout' device' dataType' shape') gradient device dataType vocabDim = Tensor (gradient' <|> gradient) (layout' <+> 'Layout 'Dense) (device' <+> device) (dataType' <+> dataType) (BroadcastShapesF shape' ('Shape '['Dim ('Name "*") ('Size 1), vocabDim]))

LMHeadOutputF 'MBART decoderOutput gradient device dataType vocabDim = LMHeadOutputF 'BART decoderOutput gradient device dataType vocabDim

LMHeadOutputF 'Pegasus decoderOutput gradient device dataType vocabDim = LMHeadOutputF 'BART decoderOutput gradient device dataType vocabDim

LMHeadOutputF 'RoBERTa decoderOutput _ _ _ _ = decoderOutput

LMHeadOutputF 'BERT decoderOutput _ _ _ _ = decoderOutput

testLMHead :: IO _