Torch.GraduallyTyped.NN.Transformer.GSelfAttention

Documentation

data GSelfAttention (initialLayerNorm :: Type) (mha :: Type) (dropout :: Type) (finalLayerNorm :: Type) where

Generic self-attention layer data type.

initialLayerNorm: the initial layer normalization

mha: the multi-headed attention layer

dropout: the dropout layer

finalLayerNorm: the final layer normalization

Constructors

GSelfAttention

Fields

:: forall initialLayerNorm mha dropout finalLayerNorm. { saInitialLayerNorm :: initialLayerNorm

initial layer normalization of the self-attention layer.

, saMultiHeadAttention :: mha

multi-headed attention layer specialized for self-attention.

, saDropout :: dropout

dropout

, saFinalLayerNorm :: finalLayerNorm

final layer normalization of the self-attention layer.

} -> GSelfAttention initialLayerNorm mha dropout finalLayerNorm

type family GSelfAttentionF (style :: TransformerStyle) (gradient :: Gradient RequiresGradient) (device :: Device (DeviceType Nat)) (dataType :: DataType DType) (headDim :: Dim (Name Symbol) (Size Nat)) (headEmbedDim :: Dim (Name Symbol) (Size Nat)) (embedDim :: Dim (Name Symbol) (Size Nat)) (queryEmbedDim :: Dim (Name Symbol) (Size Nat)) (hasDropout :: HasDropout) :: Type where ...

Equations

GSelfAttentionF style gradient device dataType headDim headEmbedDim embedDim queryEmbedDim hasDropout = GSelfAttention (SAInitialLayerNormF style gradient device dataType queryEmbedDim) (SAMultiheadAttentionF style gradient device dataType headDim headEmbedDim embedDim queryEmbedDim hasDropout) (SADropoutF style hasDropout) (SAFinalLayerNormF style gradient device dataType queryEmbedDim)

type family SAInitialLayerNormF (style :: TransformerStyle) (gradient :: Gradient RequiresGradient) (device :: Device (DeviceType Nat)) (dataType :: DataType DType) (queryEmbedDim :: Dim (Name Symbol) (Size Nat)) :: Type where ...

Specifies the initial layer normalization of the self-attention layer.

Equations

SAInitialLayerNormF 'T5 gradient device dataType queryEmbedDim = NamedModel (LayerNorm 'WithoutBias gradient device dataType ('Shape '[queryEmbedDim]))

SAInitialLayerNormF 'ByT5 gradient device dataType queryEmbedDim = SAInitialLayerNormF 'T5 gradient device dataType queryEmbedDim

SAInitialLayerNormF 'BART _ _ _ _ = ()

SAInitialLayerNormF 'MBART gradient device dataType queryEmbedDim = SAInitialLayerNormF 'BART gradient device dataType queryEmbedDim

SAInitialLayerNormF 'Pegasus gradient device dataType queryEmbedDim = NamedModel (LayerNorm 'WithBias gradient device dataType ('Shape '[queryEmbedDim]))

SAInitialLayerNormF 'BERT _ _ _ _ = ()

SAInitialLayerNormF 'RoBERTa gradient device dataType queryEmbedDim = SAInitialLayerNormF 'BERT gradient device dataType queryEmbedDim

type family SAMultiheadAttentionF (style :: TransformerStyle) (gradient :: Gradient RequiresGradient) (device :: Device (DeviceType Nat)) (dataType :: DataType DType) (headDim :: Dim (Name Symbol) (Size Nat)) (headEmbedDim :: Dim (Name Symbol) (Size Nat)) (embedDim :: Dim (Name Symbol) (Size Nat)) (queryEmbedDim :: Dim (Name Symbol) (Size Nat)) (hasDropout :: HasDropout) :: Type where ...

Specifies the multi-headed attention layer of the self-attention layer.

Equations

SAMultiheadAttentionF style gradient device dataType headDim headEmbedDim embedDim queryEmbedDim hasDropout = NamedModel (GMultiHeadAttentionF style gradient device dataType headDim headEmbedDim embedDim queryEmbedDim queryEmbedDim queryEmbedDim hasDropout)

type family SADropoutF (style :: TransformerStyle) (hasDropout :: HasDropout) :: Type where ...

Specifies the dropout layer of the self-attention layer.

Equations

SADropoutF _ 'WithDropout = Dropout

SADropoutF _ 'WithoutDropout = ()

type family SAFinalLayerNormF (style :: TransformerStyle) (gradient :: Gradient RequiresGradient) (device :: Device (DeviceType Nat)) (dataType :: DataType DType) (queryEmbedDim :: Dim (Name Symbol) (Size Nat)) :: Type where ...

Specifies the final layer normalization of the self-attention layer.

Equations

SAFinalLayerNormF 'T5 _ _ _ _ = ()

SAFinalLayerNormF 'ByT5 gradient device dataType queryEmbedDim = SAFinalLayerNormF 'T5 gradient device dataType queryEmbedDim

SAFinalLayerNormF 'BART gradient device dataType queryEmbedDim = NamedModel (LayerNorm 'WithBias gradient device dataType ('Shape '[queryEmbedDim]))

SAFinalLayerNormF 'MBART gradient device dataType queryEmbedDim = SAFinalLayerNormF 'BART gradient device dataType queryEmbedDim

SAFinalLayerNormF 'Pegasus gradient device dataType queryEmbedDim = ()

SAFinalLayerNormF 'BERT gradient device dataType queryEmbedDim = NamedModel (LayerNorm 'WithBias gradient device dataType ('Shape '[queryEmbedDim]))

SAFinalLayerNormF 'RoBERTa gradient device dataType queryEmbedDim = SAFinalLayerNormF 'BERT gradient device dataType queryEmbedDim

selfAttentionSpec :: forall style gradient device dataType headDim headEmbedDim embedDim queryEmbedDim hasDropout. STransformerStyle style -> SGradient gradient -> SDevice device -> SDataType dataType -> SDim headDim -> SDim headEmbedDim -> SDim embedDim -> SDim queryEmbedDim -> SHasDropout hasDropout -> Double -> Double -> ModelSpec (GSelfAttentionF style gradient device dataType headDim headEmbedDim embedDim queryEmbedDim hasDropout)

Specifies the parameters of a self-attention layer.

style: the style of the transformer stack, e.g. ST5, SByT5, etc.

gradient: whether to compute the gradient of the stack's parameters.

device: the computational device on which the stack is allocated.

dataType: the data type of the stack's parameters.

headDim: the dimension of all transformer heads in the stack.

headEmbedDim: the dimension of the transformer head embeddings.

embedDim: the dimension of the transformer embeddings.

queryEmbedDim: the dimension of the transformer query embeddings.

dropoutP: the dropout rate.

eps: the epsilon value for numerical stability of the layer normalization.