Torch.GraduallyTyped.NN.Transformer.GStack

Documentation

newtype GTransformerStack (stack :: Type) where

Generic transformer stack.

stack is a stack of tranformer blocks.

Constructors

GTransformerStack :: forall stack. stack -> GTransformerStack stack

encoderStackSpec :: forall style numLayers gradient device dataType headDim headEmbedDim embedDim queryEmbedDim ffnDim. STransformerStyle style -> SNat numLayers -> SGradient gradient -> SDevice device -> SDataType dataType -> SDim headDim -> SDim headEmbedDim -> SDim embedDim -> SDim queryEmbedDim -> SDim ffnDim -> Double -> Double -> ModelSpec (GTransformerStack (Vector numLayers (GTransformerBlock (EncoderBlockSelfAttentionF style gradient device dataType headDim headEmbedDim embedDim queryEmbedDim) EncoderBlockCrossAttentionF (EncoderBlockFeedForwardNetworkF style gradient device dataType queryEmbedDim ffnDim))))

Specifies the parameters of a transformer stack in an encoder configuration.

style: the style of the transformer stack, e.g. ST5, SByT5, etc.

gradient: whether to compute the gradient of the stack's parameters.

device: the computational device on which the stack is allocated.

dataType: the data type of the stack's parameters.

headDim: the dimension of all transformer heads in the stack.

headEmbedDim: the dimension of the transformer head embeddings.

embedDim: the dimension of the transformer embeddings.

queryEmbedDim: the dimension of the transformer query embeddings.

ffnDim: the dimension of the feed-forward network.

dropoutP: the dropout rate.

eps: the epsilon value for numerical stability of the layer normalization.

decoderStackSpec :: forall style numLayers gradient device dataType headDim headEmbedDim embedDim queryEmbedDim keyEmbedDim ffnDim. STransformerStyle style -> SNat numLayers -> SGradient gradient -> SDevice device -> SDataType dataType -> SDim headDim -> SDim headEmbedDim -> SDim embedDim -> SDim queryEmbedDim -> SDim keyEmbedDim -> SDim ffnDim -> Double -> Double -> ModelSpec (GTransformerStack (Vector numLayers (GTransformerBlock (DecoderBlockSelfAttentionF style gradient device dataType headDim headEmbedDim embedDim queryEmbedDim) (DecoderBlockCrossAttentionF style gradient device dataType headDim headEmbedDim embedDim queryEmbedDim keyEmbedDim) (DecoderBlockFeedForwardNetworkF style gradient device dataType queryEmbedDim ffnDim))))

Specifies the parameters of a transformer stack in a decoder configuration.

style: the style of the transformer stack, e.g. ST5, SByT5, etc.

gradient: whether to compute the gradient of the stack's parameters.

device: the computational device on which the stack is allocated.

dataType: the data type of the stack's parameters.

headDim: the dimension of all transformer heads in the stack.

headEmbedDim: the dimension of the transformer head embeddings.

embedDim: the dimension of the transformer embeddings.

queryEmbedDim: the dimension of the transformer query embeddings.

keyEmbedDim: the dimension of the transformer key embeddings.

ffnDim: the dimension of the feed-forward network.

dropoutP: the dropout rate.

eps: the epsilon value for numerical stability of the layer normalization.

testEncoderStack :: IO _

testDecoderStack :: IO _