Torch.GraduallyTyped.NN.Transformer.GTransformer

Documentation

data GTransformer (posEnc :: Type) (relPosEnc :: Type) (initialLayerNorm :: Type) (initialDropout :: Type) (stack :: Type) (finalLayerNorm :: Type) (finalDropout :: Type) where

Generic transformer. Can specialize to either encoder or decoder.

posEnc: an absolute positional encoding layer as used by, e.g., BERT.

relPosEnc: a relative positional encoding layer as used by, e.g., T5.

initialLayerNorm: a layer normalization layer for the embeddings.

initialDropout: a dropout layer for the embeddings.

stack: a stack of transformer blocks.

finalLayerNorm: the final layer normalization layer.

finalDropout: the final dropout layer.

Constructors

GTransformer

Fields

:: forall posEnc relPosEnc initialLayerNorm initialDropout stack finalLayerNorm finalDropout. { tPosEnc :: posEnc

absolute positional encoding

, tRelPosEnc :: relPosEnc

relative positional encoding

, tInitialLayerNorm :: initialLayerNorm

initial layer norm

, tInitialDropout :: initialDropout

initial dropout

, tStack :: stack

transformer block stack

, tFinalLayerNorm :: finalLayerNorm

final layer norm

, tFinalDropout :: finalDropout

final dropout

} -> GTransformer posEnc relPosEnc initialLayerNorm initialDropout stack finalLayerNorm finalDropout

type family TEPosEncF (style :: TransformerStyle) (gradient :: Gradient RequiresGradient) (device :: Device (DeviceType Nat)) (dataType :: DataType DType) (inputEmbedDim :: Dim (Name Symbol) (Size Nat)) (posEncDim :: Dim (Name Symbol) (Size Nat)) :: Type where ...

Specifies the absolute positional encoding layer of a transformer encoder.

Equations

TEPosEncF 'T5 _ _ _ _ _ = ()

TEPosEncF 'ByT5 gradient device dataType inputEmbedDim posEncDim = TEPosEncF 'T5 gradient device dataType inputEmbedDim posEncDim

TEPosEncF 'BART gradient device dataType inputEmbedDim posEncDim = NamedModel (Embedding gradient ('Layout 'Dense) device dataType posEncDim inputEmbedDim 'Nothing)

TEPosEncF 'MBART gradient device dataType inputEmbedDim posEncDim = TEPosEncF 'BART gradient device dataType inputEmbedDim posEncDim

TEPosEncF 'Pegasus gradient device dataType inputEmbedDim posEncDim = TEPosEncF 'BART gradient device dataType inputEmbedDim posEncDim

TEPosEncF 'BERT gradient device dataType inputEmbedDim posEncDim = NamedModel (Embedding gradient ('Layout 'Dense) device dataType posEncDim inputEmbedDim 'Nothing)

TEPosEncF 'RoBERTa gradient device dataType inputEmbedDim posEncDim = TEPosEncF 'BERT gradient device dataType inputEmbedDim posEncDim

type family TERelPosEncF (style :: TransformerStyle) (gradient :: Gradient RequiresGradient) (device :: Device (DeviceType Nat)) (dataType :: DataType DType) (headDim :: Dim (Name Symbol) (Size Nat)) (posEncDim :: Dim (Name Symbol) (Size Nat)) :: Type where ...

Specifies the relative positional encoding layer of a transformer encoder.

Equations

TERelPosEncF 'T5 gradient device dataType headDim posEncDim = NamedModel (Embedding gradient ('Layout 'Dense) device dataType posEncDim headDim 'Nothing)

TERelPosEncF 'ByT5 gradient device dataType headDim posEncDim = TERelPosEncF 'T5 gradient device dataType headDim posEncDim

TERelPosEncF 'BART _ _ _ _ _ = ()

TERelPosEncF 'MBART gradient device dataType headDim posEncDim = TERelPosEncF 'BART gradient device dataType headDim posEncDim

TERelPosEncF 'Pegasus gradient device dataType headDim posEncDim = TERelPosEncF 'BART gradient device dataType headDim posEncDim

TERelPosEncF 'BERT _ _ _ _ _ = ()

TERelPosEncF 'RoBERTa gradient device dataType headDim posEncDim = TERelPosEncF 'BERT gradient device dataType headDim posEncDim

type family TEInitialLayerNormF (style :: TransformerStyle) (gradient :: Gradient RequiresGradient) (device :: Device (DeviceType Nat)) (dataType :: DataType DType) (inputEmbedDim :: Dim (Name Symbol) (Size Nat)) :: Type where ...

Specifies the initial layer normalization layer of a transformer encoder.

Equations

TEInitialLayerNormF 'T5 _ _ _ _ = ()

TEInitialLayerNormF 'ByT5 gradient device dataType inputEmbedDim = TEInitialLayerNormF 'T5 gradient device dataType inputEmbedDim

TEInitialLayerNormF 'BART gradient device dataType inputEmbedDim = NamedModel (LayerNorm 'WithBias gradient device dataType ('Shape '[inputEmbedDim]))

TEInitialLayerNormF 'MBART gradient device dataType inputEmbedDim = TEInitialLayerNormF 'BART gradient device dataType inputEmbedDim

TEInitialLayerNormF 'Pegasus _ _ _ _ = ()

TEInitialLayerNormF 'BERT gradient device dataType inputEmbedDim = NamedModel (LayerNorm 'WithBias gradient device dataType ('Shape '[inputEmbedDim]))

TEInitialLayerNormF 'RoBERTa gradient device dataType inputEmbedDim = TEInitialLayerNormF 'BERT gradient device dataType inputEmbedDim

type family TEInitialDropoutF (style :: TransformerStyle) :: Type where ...

Specifies the initial dropout layer of a transformer encoder.

Equations

TEInitialDropoutF 'T5 = Dropout

TEInitialDropoutF 'ByT5 = Dropout

TEInitialDropoutF 'BART = Dropout

TEInitialDropoutF 'MBART = Dropout

TEInitialDropoutF 'Pegasus = Dropout

TEInitialDropoutF 'BERT = Dropout

TEInitialDropoutF 'RoBERTa = Dropout

type family TEStackF (style :: TransformerStyle) (numLayers :: Nat) (gradient :: Gradient RequiresGradient) (device :: Device (DeviceType Nat)) (dataType :: DataType DType) (headDim :: Dim (Name Symbol) (Size Nat)) (headEmbedDim :: Dim (Name Symbol) (Size Nat)) (embedDim :: Dim (Name Symbol) (Size Nat)) (inputEmbedDim :: Dim (Name Symbol) (Size Nat)) (ffnDim :: Dim (Name Symbol) (Size Nat)) :: Type where ...

Specifies the transformer block stack of a transformer encoder.

Equations

TEStackF style numLayers gradient device dataType headDim headEmbedDim embedDim inputEmbedDim ffnDim = NamedModel (GTransformerStack (Vector numLayers (GTransformerBlock (EncoderBlockSelfAttentionF style gradient device dataType headDim headEmbedDim embedDim inputEmbedDim) EncoderBlockCrossAttentionF (EncoderBlockFeedForwardNetworkF style gradient device dataType inputEmbedDim ffnDim))))

type family TEFinalLayerNormF (style :: TransformerStyle) (gradient :: Gradient RequiresGradient) (device :: Device (DeviceType Nat)) (dataType :: DataType DType) (inputEmbedDim :: Dim (Name Symbol) (Size Nat)) :: Type where ...

Specifies the final layer normalization layer of a transformer encoder.

Equations

TEFinalLayerNormF 'T5 gradient device dataType inputEmbedDim = NamedModel (LayerNorm 'WithoutBias gradient device dataType ('Shape '[inputEmbedDim]))

TEFinalLayerNormF 'ByT5 gradient device dataType inputEmbedDim = TEFinalLayerNormF 'T5 gradient device dataType inputEmbedDim

TEFinalLayerNormF 'BART _ _ _ _ = ()

TEFinalLayerNormF 'MBART gradient device dataType inputEmbedDim = TEFinalLayerNormF 'BART gradient device dataType inputEmbedDim

TEFinalLayerNormF 'Pegasus gradient device dataType inputEmbedDim = NamedModel (LayerNorm 'WithBias gradient device dataType ('Shape '[inputEmbedDim]))

TEFinalLayerNormF 'BERT _ _ _ _ = ()

TEFinalLayerNormF 'RoBERTa gradient device dataType inputEmbedDim = TEFinalLayerNormF 'BERT gradient device dataType inputEmbedDim

type family TEFinalDropoutF (style :: TransformerStyle) :: Type where ...

Specifies the final dropout layer of a transformer encoder.

Equations

TEFinalDropoutF 'T5 = Dropout

TEFinalDropoutF 'ByT5 = Dropout

TEFinalDropoutF 'BART = ()

TEFinalDropoutF 'MBART = ()

TEFinalDropoutF 'Pegasus = ()

TEFinalDropoutF 'BERT = ()

TEFinalDropoutF 'RoBERTa = ()

transformerEncoderSpec :: forall style numLayers gradient device dataType headDim headEmbedDim embedDim inputEmbedDim ffnDim posEncDim. STransformerStyle style -> SNat numLayers -> SGradient gradient -> SDevice device -> SDataType dataType -> SDim headDim -> SDim headEmbedDim -> SDim embedDim -> SDim inputEmbedDim -> SDim ffnDim -> SDim posEncDim -> Double -> Double -> ModelSpec (GTransformer (TEPosEncF style gradient device dataType inputEmbedDim posEncDim) (TERelPosEncF style gradient device dataType headDim posEncDim) (TEInitialLayerNormF style gradient device dataType inputEmbedDim) (TEInitialDropoutF style) (TEStackF style numLayers gradient device dataType headDim headEmbedDim embedDim inputEmbedDim ffnDim) (TEFinalLayerNormF style gradient device dataType inputEmbedDim) (TEFinalDropoutF style))

Specifies the parameters of a transformer in an encoder configuration.

style: the style of the transformer stack, e.g. ST5, SByT5, etc.

gradient: whether to compute the gradient of the stack's parameters.

device: the computational device on which the stack is allocated.

dataType: the data type of the stack's parameters.

headDim: the dimension of all transformer heads in the stack.

headEmbedDim: the dimension of the transformer head embeddings.

embedDim: the dimension of the transformer embeddings.

inputEmbedDim: the dimension of the transformer query embeddings.

ffnDim: the dimension of the feed-forward network.

posEncDim: the dimension of the positional encoding.

dropoutP: the dropout rate.

eps: the epsilon value for numerical stability of the layer normalization.

type family TDPosEncF (style :: TransformerStyle) (gradient :: Gradient RequiresGradient) (device :: Device (DeviceType Nat)) (dataType :: DataType DType) (inputEmbedDim :: Dim (Name Symbol) (Size Nat)) (posEncDim :: Dim (Name Symbol) (Size Nat)) :: Type where ...

Specifies the absolute positional encoding layer of a transformer decoder.

Equations

TDPosEncF 'T5 _ _ _ _ _ = ()

TDPosEncF 'ByT5 gradient device dataType inputEmbedDim posEncDim = TDPosEncF 'T5 gradient device dataType inputEmbedDim posEncDim

TDPosEncF 'BART gradient device dataType inputEmbedDim posEncDim = NamedModel (Embedding gradient ('Layout 'Dense) device dataType posEncDim inputEmbedDim 'Nothing)

TDPosEncF 'MBART gradient device dataType inputEmbedDim posEncDim = TDPosEncF 'BART gradient device dataType inputEmbedDim posEncDim

TDPosEncF 'Pegasus gradient device dataType inputEmbedDim posEncDim = TDPosEncF 'BART gradient device dataType inputEmbedDim posEncDim

type family TDRelPosEncF (style :: TransformerStyle) (gradient :: Gradient RequiresGradient) (device :: Device (DeviceType Nat)) (dataType :: DataType DType) (headDim :: Dim (Name Symbol) (Size Nat)) (posEncDim :: Dim (Name Symbol) (Size Nat)) :: Type where ...

Specifies the relative positional encoding layer of a transformer decoder.

Equations

TDRelPosEncF 'T5 gradient device dataType headDim posEncDim = NamedModel (Embedding gradient ('Layout 'Dense) device dataType posEncDim headDim 'Nothing)

TDRelPosEncF 'ByT5 gradient device dataType headDim posEncDim = TDRelPosEncF 'T5 gradient device dataType headDim posEncDim

TDRelPosEncF 'BART _ _ _ _ _ = ()

TDRelPosEncF 'MBART gradient device dataType headDim posEncDim = TDRelPosEncF 'BART gradient device dataType headDim posEncDim

TDRelPosEncF 'Pegasus gradient device dataType headDim posEncDim = TDRelPosEncF 'BART gradient device dataType headDim posEncDim

type family TDInitialLayerNormF (style :: TransformerStyle) (gradient :: Gradient RequiresGradient) (device :: Device (DeviceType Nat)) (dataType :: DataType DType) (inputEmbedDim :: Dim (Name Symbol) (Size Nat)) :: Type where ...

Specifies the initial layer normalization layer of a transformer decoder.

Equations

TDInitialLayerNormF 'T5 _ _ _ _ = ()

TDInitialLayerNormF 'ByT5 gradient device dataType inputEmbedDim = TDInitialLayerNormF 'T5 gradient device dataType inputEmbedDim

TDInitialLayerNormF 'BART gradient device dataType inputEmbedDim = NamedModel (LayerNorm 'WithBias gradient device dataType ('Shape '[inputEmbedDim]))

TDInitialLayerNormF 'MBART gradient device dataType inputEmbedDim = TDInitialLayerNormF 'BART gradient device dataType inputEmbedDim

TDInitialLayerNormF 'Pegasus _ _ _ _ = ()

type family TDInitialDropoutF (style :: TransformerStyle) :: Type where ...

Specifies the initial dropout layer of a transformer decoder.

Equations

TDInitialDropoutF 'T5 = Dropout

TDInitialDropoutF 'ByT5 = Dropout

TDInitialDropoutF 'BART = Dropout

TDInitialDropoutF 'MBART = Dropout

TDInitialDropoutF 'Pegasus = Dropout

type family TDStackF (style :: TransformerStyle) (numLayers :: Nat) (gradient :: Gradient RequiresGradient) (device :: Device (DeviceType Nat)) (dataType :: DataType DType) (headDim :: Dim (Name Symbol) (Size Nat)) (headEmbedDim :: Dim (Name Symbol) (Size Nat)) (embedDim :: Dim (Name Symbol) (Size Nat)) (decoderInputEmbedDim :: Dim (Name Symbol) (Size Nat)) (encoderOutputEmbedDim :: Dim (Name Symbol) (Size Nat)) (ffnDim :: Dim (Name Symbol) (Size Nat)) :: Type where ...

Specifies the transformer block stack of a transformer decoder.

Equations

TDStackF style numLayers gradient device dataType headDim headEmbedDim embedDim decoderInputEmbedDim encoderOutputEmbedDim ffnDim = NamedModel (GTransformerStack (Vector numLayers (GTransformerBlock (DecoderBlockSelfAttentionF style gradient device dataType headDim headEmbedDim embedDim decoderInputEmbedDim) (DecoderBlockCrossAttentionF style gradient device dataType headDim headEmbedDim embedDim decoderInputEmbedDim encoderOutputEmbedDim) (DecoderBlockFeedForwardNetworkF style gradient device dataType decoderInputEmbedDim ffnDim))))

type family TDFinalLayerNormF (style :: TransformerStyle) (gradient :: Gradient RequiresGradient) (device :: Device (DeviceType Nat)) (dataType :: DataType DType) (inputEmbedDim :: Dim (Name Symbol) (Size Nat)) :: Type where ...

Specifies the final layer normalization layer of a transformer decoder.

Equations

TDFinalLayerNormF 'T5 gradient device dataType inputEmbedDim = NamedModel (LayerNorm 'WithoutBias gradient device dataType ('Shape '[inputEmbedDim]))

TDFinalLayerNormF 'ByT5 gradient device dataType inputEmbedDim = TDFinalLayerNormF 'T5 gradient device dataType inputEmbedDim

TDFinalLayerNormF 'BART _ _ _ _ = ()

TDFinalLayerNormF 'MBART gradient device dataType inputEmbedDim = TDFinalLayerNormF 'BART gradient device dataType inputEmbedDim

TDFinalLayerNormF 'Pegasus gradient device dataType inputEmbedDim = NamedModel (LayerNorm 'WithBias gradient device dataType ('Shape '[inputEmbedDim]))

type family TDFinalDropoutF (style :: TransformerStyle) :: Type where ...

Specifies the final dropout layer of a transformer decoder.

Equations

TDFinalDropoutF 'T5 = Dropout

TDFinalDropoutF 'ByT5 = Dropout

TDFinalDropoutF 'BART = ()

TDFinalDropoutF 'MBART = ()

TDFinalDropoutF 'Pegasus = ()

transformerDecoderSpec :: forall style numLayers gradient device dataType headDim headEmbedDim embedDim decoderInputEmbedDim encoderOutputEmbedDim ffnDim posEncDim. STransformerStyle style -> SNat numLayers -> SGradient gradient -> SDevice device -> SDataType dataType -> SDim headDim -> SDim headEmbedDim -> SDim embedDim -> SDim decoderInputEmbedDim -> SDim encoderOutputEmbedDim -> SDim ffnDim -> SDim posEncDim -> Double -> Double -> ModelSpec (GTransformer (TDPosEncF style gradient device dataType decoderInputEmbedDim posEncDim) (TDRelPosEncF style gradient device dataType headDim posEncDim) (TDInitialLayerNormF style gradient device dataType decoderInputEmbedDim) (TDInitialDropoutF style) (TDStackF style numLayers gradient device dataType headDim headEmbedDim embedDim decoderInputEmbedDim encoderOutputEmbedDim ffnDim) (TDFinalLayerNormF style gradient device dataType decoderInputEmbedDim) (TDFinalDropoutF style))

Specifies the parameters of a transformer in a decoder configuration.

style: the style of the transformer stack, e.g. ST5, SByT5, etc.

gradient: whether to compute the gradient of the stack's parameters.

device: the computational device on which the stack is allocated.

dataType: the data type of the stack's parameters.

headDim: the dimension of all transformer heads in the stack.

headEmbedDim: the dimension of the transformer head embeddings.

embedDim: the dimension of the transformer embeddings.

decoderInputEmbedDim: the dimension of the decoder input embeddings.

encoderOutputEmbedDim: the dimension of the encoder output embeddings.

ffnDim: the dimension of the feed-forward network.

posEncDim: the dimension of the positional encoding.

dropoutP: the dropout rate.

eps: the epsilon value for numerical stability of the layer normalization.