Torch.GraduallyTyped.NN.Transformer.MultiHeadAttention

Documentation

data GMultiHeadAttention (headDim :: Dim (Name Symbol) (Size Nat)) (headEmbedDim :: Dim (Name Symbol) (Size Nat)) (embedDim :: Dim (Name Symbol) (Size Nat)) (qInProj :: Type) (kInProj :: Type) (vInProj :: Type) (outProj :: Type) (dropout :: Type) where

Generic multi-headed attention layer. Needs to be specialized to a given transformer type, e.g. T5. See MultiHeadAttention.

Constructors

GMultiHeadAttention

Fields

:: forall headDim headEmbedDim embedDim qInProj kInProj vInProj outProj dropout. { mhaHeadDim :: SDim headDim

head dim

, mhaHeadEmbedDim :: SDim headEmbedDim

head embed dim

, mhaEmbedDim :: SDim embedDim

embed dim

, mhaQInProj :: qInProj

in-projection for query

, mhaKInProj :: kInProj

in-projection for key

, mhaVInProj :: vInProj

in-projection for value

, mhaOutProj :: outProj

out-projection

, mhaDropout :: dropout

dropout

} -> GMultiHeadAttention headDim headEmbedDim embedDim qInProj kInProj vInProj outProj dropout

newtype MultiHeadAttention (style :: TransformerStyle) (gradient :: Gradient RequiresGradient) (device :: Device (DeviceType Nat)) (dataType :: DataType DType) (headDim :: Dim (Name Symbol) (Size Nat)) (headEmbedDim :: Dim (Name Symbol) (Size Nat)) (embedDim :: Dim (Name Symbol) (Size Nat)) (queryEmbedDim :: Dim (Name Symbol) (Size Nat)) (keyEmbedDim :: Dim (Name Symbol) (Size Nat)) (valueEmbedDim :: Dim (Name Symbol) (Size Nat)) where

Multi-headed attention layer.

Constructors

MultiHeadAttention :: forall style gradient device dataType headDim headEmbedDim embedDim queryEmbedDim keyEmbedDim valueEmbedDim. GMultiHeadAttention headDim headEmbedDim embedDim (QInProjF style gradient device dataType queryEmbedDim embedDim) (KInProjF style gradient device dataType keyEmbedDim embedDim) (VInProjF style gradient device dataType valueEmbedDim embedDim) (OutProjF style gradient device dataType embedDim queryEmbedDim) (DropoutF style) -> MultiHeadAttention style gradient device dataType headDim headEmbedDim embedDim queryEmbedDim keyEmbedDim valueEmbedDim

data MultiHeadAttentionSpec (style :: TransformerStyle) (gradient :: Gradient RequiresGradient) (device :: Device (DeviceType Nat)) (dataType :: DataType DType) (headDim :: Dim (Name Symbol) (Size Nat)) (headEmbedDim :: Dim (Name Symbol) (Size Nat)) (embedDim :: Dim (Name Symbol) (Size Nat)) (queryEmbedDim :: Dim (Name Symbol) (Size Nat)) (keyEmbedDim :: Dim (Name Symbol) (Size Nat)) (valueEmbedDim :: Dim (Name Symbol) (Size Nat)) where

Constructors

MultiHeadAttentionSpec :: forall style gradient device dataType headDim headEmbedDim embedDim queryEmbedDim keyEmbedDim valueEmbedDim. STransformerStyle style -> SGradient gradient -> SDevice device -> SDataType dataType -> SDim headDim -> SDim headEmbedDim -> SDim embedDim -> SDim queryEmbedDim -> SDim keyEmbedDim -> SDim valueEmbedDim -> Double -> MultiHeadAttentionSpec style gradient device dataType headDim headEmbedDim embedDim queryEmbedDim keyEmbedDim valueEmbedDim

type family QInProjF (style :: TransformerStyle) (gradient :: Gradient RequiresGradient) (device :: Device (DeviceType Nat)) (dataType :: DataType DType) (queryEmbedDim :: Dim (Name Symbol) (Size Nat)) (embedDim :: Dim (Name Symbol) (Size Nat)) :: Type where ...

Equations

QInProjF 'T5 gradient device dataType queryEmbedDim embedDim = Linear 'WithoutBias gradient device dataType queryEmbedDim embedDim

QInProjF 'ByT5 gradient device dataType queryEmbedDim embedDim = QInProjF 'T5 gradient device dataType queryEmbedDim embedDim

QInProjF _ gradient device dataType queryEmbedDim embedDim = Linear 'WithBias gradient device dataType queryEmbedDim embedDim

type family KInProjF (style :: TransformerStyle) (gradient :: Gradient RequiresGradient) (device :: Device (DeviceType Nat)) (dataType :: DataType DType) (keyEmbedDim :: Dim (Name Symbol) (Size Nat)) (embedDim :: Dim (Name Symbol) (Size Nat)) :: Type where ...

Equations

KInProjF 'T5 gradient device dataType keyEmbedDim embedDim = Linear 'WithoutBias gradient device dataType keyEmbedDim embedDim

KInProjF 'ByT5 gradient device dataType keyEmbedDim embedDim = KInProjF 'T5 gradient device dataType keyEmbedDim embedDim

KInProjF _ gradient device dataType keyEmbedDim embedDim = Linear 'WithBias gradient device dataType keyEmbedDim embedDim

type family VInProjF (style :: TransformerStyle) (gradient :: Gradient RequiresGradient) (device :: Device (DeviceType Nat)) (dataType :: DataType DType) (valueEmbedDim :: Dim (Name Symbol) (Size Nat)) (embedDim :: Dim (Name Symbol) (Size Nat)) :: Type where ...

Equations

VInProjF 'T5 gradient device dataType valueEmbedDim embedDim = Linear 'WithoutBias gradient device dataType valueEmbedDim embedDim

VInProjF 'ByT5 gradient device dataType valueEmbedDim embedDim = VInProjF 'T5 gradient device dataType valueEmbedDim embedDim

VInProjF _ gradient device dataType valueEmbedDim embedDim = Linear 'WithBias gradient device dataType valueEmbedDim embedDim

type family OutProjF (style :: TransformerStyle) (gradient :: Gradient RequiresGradient) (device :: Device (DeviceType Nat)) (dataType :: DataType DType) (embedDim :: Dim (Name Symbol) (Size Nat)) (queryEmbedDim :: Dim (Name Symbol) (Size Nat)) :: Type where ...

Equations

OutProjF 'T5 gradient device dataType embedDim queryEmbedDim = Linear 'WithoutBias gradient device dataType embedDim queryEmbedDim

OutProjF 'ByT5 gradient device dataType embedDim queryEmbedDim = OutProjF 'T5 gradient device dataType embedDim queryEmbedDim

OutProjF _ gradient device dataType embedDim queryEmbedDim = Linear 'WithBias gradient device dataType embedDim queryEmbedDim

type family DropoutF (style :: TransformerStyle) :: Type where ...

Equations

DropoutF _ = Dropout

data Scaling

Whether or not scaling is applied in the multi-headed attention layer.

Constructors

NoScaling
Scaling is not done.

QueryScaling Double
Scaling is applied to the query after in the in-projection.

WeightScaling Double
Scaling is applied to the attention weights.

data OutProj

Whether or not out-projection is applied in the multi-headed attention layer.

Constructors

NoOutProj
Out-projection is absent.

OutProj
Out-projection is applied.

type BatchDim queryShape keyShape valueShape = (queryShape ! 0) <+> ((keyShape ! 0) <+> (valueShape ! 0))

getBatchDim :: forall m queryShape keyShape valueShape batchDim. (MonadThrow m, batchDim ~ BatchDim queryShape keyShape valueShape) => SShape queryShape -> SShape keyShape -> SShape valueShape -> m (SDim batchDim)

type QuerySeqDim queryShape = queryShape ! 1

getQuerySeqDim :: forall m queryShape querySeqDim. (MonadThrow m, querySeqDim ~ QuerySeqDim queryShape) => SShape queryShape -> m (SDim querySeqDim)

type KeySeqDim keyShape valueShape = (keyShape ! 1) <+> (valueShape ! 1)

getKeySeqDim :: forall m keyShape valueShape keySeqDim. (MonadThrow m, keySeqDim ~ KeySeqDim keyShape valueShape) => SShape keyShape -> SShape valueShape -> m (SDim keySeqDim)

testMHA :: IO _