Torch.GraduallyTyped.NN.Transformer.Type

Documentation

data TransformerStyle

A data type representing the style of a transformer. Every supported transformer has a constructor of this type.

Constructors

T5
T5 transformer style, see https://ai.googleblog.com/2020/02/exploring-transfer-learning-with-t5.html

ByT5
ByT5 transformer style, see https://arxiv.org/abs/2105.13626

BART
BART transformer style, see https://arxiv.org/abs/1910.13461

MBART
MBART transformer style, see https://arxiv.org/abs/2001.08210

Pegasus
Pegasus transformer style, see https://ai.googleblog.com/2020/06/pegasus-state-of-art-model-for.html

BERT
BERT transformer style, see https://arxiv.org/abs/1810.04805

RoBERTa
RoBERTa transformer style, see https://arxiv.org/abs/1907.11692

GPT2
GPT2 transformer style, see https://openai.com/blog/better-language-models/

type T5Sym0 = 'T5 :: TransformerStyle

type ByT5Sym0 = 'ByT5 :: TransformerStyle

type BARTSym0 = 'BART :: TransformerStyle

type MBARTSym0 = 'MBART :: TransformerStyle

type PegasusSym0 = 'Pegasus :: TransformerStyle

type BERTSym0 = 'BERT :: TransformerStyle

type RoBERTaSym0 = 'RoBERTa :: TransformerStyle

type GPT2Sym0 = 'GPT2 :: TransformerStyle

data STransformerStyle z where

Constructors

ST5 :: STransformerStyle ('T5 :: TransformerStyle)

SByT5 :: STransformerStyle ('ByT5 :: TransformerStyle)

SBART :: STransformerStyle ('BART :: TransformerStyle)

SMBART :: STransformerStyle ('MBART :: TransformerStyle)

SPegasus :: STransformerStyle ('Pegasus :: TransformerStyle)

SBERT :: STransformerStyle ('BERT :: TransformerStyle)

SRoBERTa :: STransformerStyle ('RoBERTa :: TransformerStyle)

SGPT2 :: STransformerStyle ('GPT2 :: TransformerStyle)

data TransformerHead

A data type representing the type of head used in a transformer.

Constructors

WithoutHead

WithLMHead

type WithoutHeadSym0 = 'WithoutHead :: TransformerHead

type WithLMHeadSym0 = 'WithLMHead :: TransformerHead

data STransformerHead z where

Constructors

SWithoutHead :: STransformerHead ('WithoutHead :: TransformerHead)

SWithLMHead :: STransformerHead ('WithLMHead :: TransformerHead)

padded :: Integral n => n -> a -> [a] -> [a]

mkTransformerInput

Arguments

:: forall batchDim seqDim device m output. (MonadThrow m, SGetDim batchDim, SGetDim seqDim, Catch ('Shape '['Dim ('Name "*") 'UncheckedSize, 'Dim ('Name "*") 'UncheckedSize] <+> 'Shape '[batchDim, seqDim]), output ~ Tensor ('Gradient 'WithoutGradient) ('Layout 'Dense) device ('DataType 'Int64) ('Shape '[batchDim, seqDim]))

=> Int
padding token id

-> SDim batchDim
batch dimension singleton

-> SDim seqDim
sequence dimension singleton

-> SDevice device
device for the tensor

-> [[Int]]
batch of input ids

-> m output
input tensor

Converts a doubly-nested list of input ids to a batched input tensor. The outer list is over batches, the inner list over sequences. The batch size is inferred from the length of the outer list. The sequence length is inferred from the length of the inner list. The input ids are padded to the maximum sequence length. The output tensor is truncated to the maximum sequence length.

type MkPosC device shape seqDim seqName seqSize output = (SGetDevice device, SGetShape shape, seqDim ~ (shape ! 1), seqDim ~ 'Dim seqName seqSize, output ~ Tensor ('Gradient 'WithoutGradient) ('Layout 'Dense) device ('DataType 'Int64) ('Shape '['Dim ('Name "*") seqSize]))

mkPos

Arguments

:: forall m gradient layout device dataType shape seqDim seqName seqSize output. (MonadThrow m, MkPosC device shape seqDim seqName seqSize output)

=> Tensor gradient layout device dataType shape
input tensor

-> m output
positions of the input tokens

Computes absolute positions of the input tokens. Given an input tensor of shape [batchDim, Dim seqName seqSize], returns a tensor of shape [Dim "*" seqSize].

data MkAbsPos

Constructors

MkAbsPos

MkAbsPosWithOffset

Fields

absPosOffset :: Int

mkRelPos' :: Int -> Int -> Int -> Int -> [[Int]]

Computes relative positions of the input tokens to the encoder.

>>> mkRelPos' 32 128 21 17
[[0,17,18,19,20,21,22,23,24,24,24,24,25,25,25,25,26],[1,0,17,18,19,20,21,22,23,24,24,24,24,25,25,25,25],[2,1,0,17,18,19,20,21,22,23,24,24,24,24,25,25,25],[3,2,1,0,17,18,19,20,21,22,23,24,24,24,24,25,25],[4,3,2,1,0,17,18,19,20,21,22,23,24,24,24,24,25],[5,4,3,2,1,0,17,18,19,20,21,22,23,24,24,24,24],[6,5,4,3,2,1,0,17,18,19,20,21,22,23,24,24,24],[7,6,5,4,3,2,1,0,17,18,19,20,21,22,23,24,24],[8,7,6,5,4,3,2,1,0,17,18,19,20,21,22,23,24],[8,8,7,6,5,4,3,2,1,0,17,18,19,20,21,22,23],[8,8,8,7,6,5,4,3,2,1,0,17,18,19,20,21,22],[8,8,8,8,7,6,5,4,3,2,1,0,17,18,19,20,21],[9,8,8,8,8,7,6,5,4,3,2,1,0,17,18,19,20],[9,9,8,8,8,8,7,6,5,4,3,2,1,0,17,18,19],[9,9,9,8,8,8,8,7,6,5,4,3,2,1,0,17,18],[9,9,9,9,8,8,8,8,7,6,5,4,3,2,1,0,17],[10,9,9,9,9,8,8,8,8,7,6,5,4,3,2,1,0],[10,10,9,9,9,9,8,8,8,8,7,6,5,4,3,2,1],[10,10,10,9,9,9,9,8,8,8,8,7,6,5,4,3,2],[10,10,10,10,9,9,9,9,8,8,8,8,7,6,5,4,3],[10,10,10,10,10,9,9,9,9,8,8,8,8,7,6,5,4]]

type MkRelPosC device shape seqDim seqName seqSize output = (SGetDevice device, SGetShape shape, seqDim ~ (shape ! 1), seqDim ~ 'Dim seqName seqSize, Catch ('['Dim ('Name "*") 'UncheckedSize, 'Dim ('Name "*") 'UncheckedSize] <+> '['Dim ('Name "*") seqSize, 'Dim ('Name "*") seqSize]), output ~ Tensor ('Gradient 'WithoutGradient) ('Layout 'Dense) device ('DataType 'Int64) ('Shape '['Dim ('Name "*") ('Size 1), 'Dim ('Name "*") seqSize, 'Dim ('Name "*") seqSize]))

mkRelPos

Arguments

:: forall m gradient layout device dataType shape relPosEncBucketDim seqDim seqName seqSize output. (MonadThrow m, MkRelPosC device shape seqDim seqName seqSize output)

=> SDim relPosEncBucketDim
bucket dimension

-> Int
maximum distance

-> Tensor gradient layout device dataType shape
input tensor

-> m output
relative positions of the input tokens

Computes relative positions of the input tokens to the encoder. Given an input tensor of shape [batchDim, Dim seqName seqSize], returns a tensor of shape [1, Dim "*" seqSize, Dim "*" seqSize].

mkDecoderRelPos' :: Int -> Int -> Int -> Int -> [[Int]]

Computes relative positions of the input tokens to the decoder.

>>> mkDecoderRelPos' 32 128 21 17
[[0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0],[1,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0],[2,1,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0],[3,2,1,0,0,0,0,0,0,0,0,0,0,0,0,0,0],[4,3,2,1,0,0,0,0,0,0,0,0,0,0,0,0,0],[5,4,3,2,1,0,0,0,0,0,0,0,0,0,0,0,0],[6,5,4,3,2,1,0,0,0,0,0,0,0,0,0,0,0],[7,6,5,4,3,2,1,0,0,0,0,0,0,0,0,0,0],[8,7,6,5,4,3,2,1,0,0,0,0,0,0,0,0,0],[9,8,7,6,5,4,3,2,1,0,0,0,0,0,0,0,0],[10,9,8,7,6,5,4,3,2,1,0,0,0,0,0,0,0],[11,10,9,8,7,6,5,4,3,2,1,0,0,0,0,0,0],[12,11,10,9,8,7,6,5,4,3,2,1,0,0,0,0,0],[13,12,11,10,9,8,7,6,5,4,3,2,1,0,0,0,0],[14,13,12,11,10,9,8,7,6,5,4,3,2,1,0,0,0],[15,14,13,12,11,10,9,8,7,6,5,4,3,2,1,0,0],[16,15,14,13,12,11,10,9,8,7,6,5,4,3,2,1,0],[16,16,15,14,13,12,11,10,9,8,7,6,5,4,3,2,1],[16,16,16,15,14,13,12,11,10,9,8,7,6,5,4,3,2],[17,16,16,16,15,14,13,12,11,10,9,8,7,6,5,4,3],[17,17,16,16,16,15,14,13,12,11,10,9,8,7,6,5,4]]

mkDecoderRelPos

Arguments

:: forall m gradient layout device dataType shape relPosEncBucketDim seqDim seqName seqSize output. (MonadThrow m, MkRelPosC device shape seqDim seqName seqSize output)

=> SDim relPosEncBucketDim
bucket dimension

-> Int
maximum distance

-> Tensor gradient layout device dataType shape
decoder input tensor

-> m output
relative positions of the input tokens

Computes relative positions of the input tokens to the decoder. Given an input tensor of shape [batchDim, Dim seqName seqSize], returns a tensor of shape [1, Dim "*" seqSize, Dim "*" seqSize].

data MkRelPos (relPosEncBucketDim :: Dim (Name Symbol) (Size Nat)) where

Constructors

MkRelPos

Fields

:: forall relPosEncBucketDim. { relPosEncBucketDim :: SDim relPosEncBucketDim

, relPosMaxDistance :: Int

} -> MkRelPos relPosEncBucketDim

MkDecoderRelPos

Fields

:: forall relPosEncBucketDim. { decoderRelPosEncBucketDim :: SDim relPosEncBucketDim

, decoderRelPosMaxDistance :: Int

} -> MkRelPos relPosEncBucketDim

type MkTransformerPaddingMaskC layout device dataType shape output = (SGetDevice device, Catch (dataType <+> 'DataType 'Int64), output ~ Tensor ('Gradient 'WithoutGradient) (layout <+> 'Layout 'Dense) device ('DataType 'Bool) (BroadcastShapesF shape ('Shape '['Dim ('Name "*") ('Size 1)])))

mkTransformerPaddingMask

Arguments

:: forall m gradient layout device dataType shape output. (MonadThrow m, MkTransformerPaddingMaskC layout device dataType shape output)

=> Int
padding token id

-> Tensor gradient layout device dataType shape
input tensor

-> m output
padding mask

Computes the padding mask for a transformer. Given an input tensor of shape [batchDim, Dim seqName seqSize], returns a tensor of shape [batchDim, Dim "*" seqSize].

newtype MkTransformerPaddingMask

Constructors

MkTransformerPaddingMask

Fields

padTokenId :: Int

type MkTransformerAttentionMaskC transformerDataType gradient layout device dataType shape seqDim output = (SGetLayout layout, SGetDevice device, SGetShape shape, seqDim ~ (shape ! 1), Catch (gradient <+> 'Gradient 'WithoutGradient), Catch (dataType <+> 'DataType 'Bool), Catch (UnsqueezeF ('SelectDim ('ByIndex 1)) shape), Catch (BroadcastShapesF (UnsqueezeF ('SelectDim ('ByIndex 1)) shape) ('Shape '['Dim ('Name "*") ('Size 1), seqDim, seqDim])), output ~ Tensor ('Gradient 'WithoutGradient) (layout <+> 'Layout 'Dense) device transformerDataType (BroadcastShapesF (UnsqueezeF ('SelectDim ('ByIndex 1)) shape) ('Shape '['Dim ('Name "*") ('Size 1), seqDim, seqDim])))

mkTransformerAttentionMask

Arguments

:: forall m transformerDataType gradient layout device dataType shape seqDim output. (MonadThrow m, MkTransformerAttentionMaskC transformerDataType gradient layout device dataType shape seqDim output)

=> SDataType transformerDataType
data type singleton of the transformer

-> Double
attention mask bias (typically a large negative number)

-> Tensor gradient layout device dataType shape
encoder padding mask

-> m output

Creates a bidirectional attention mask for a transformer. Given a padding mask of shape [batchDim, seqDim], returns a tensor of shape [batchDim, seqDim, seqDim].

data MkTransformerAttentionMask (dataType :: DataType DType) where

Constructors

MkTransformerAttentionMask

Fields

:: forall dataType. { attentionMaskDataType :: SDataType dataType

, attentionMaskBias :: Double

} -> MkTransformerAttentionMask dataType

type MkTransformerDecoderAttentionMaskC transformerDataType layout device shape seqDim output = (SGetLayout layout, SGetDevice device, SGetShape shape, seqDim ~ (shape ! 1), Catch seqDim, Catch (UnsqueezeF ('SelectDim ('ByIndex 1)) shape), Catch (BroadcastShapesF ('Shape '['Dim ('Name "*") ('Size 1), seqDim, seqDim]) (UnsqueezeF ('SelectDim ('ByIndex 1)) shape)), Catch (BroadcastShapesF (BroadcastShapesF ('Shape '['Dim ('Name "*") ('Size 1), seqDim, seqDim]) (UnsqueezeF ('SelectDim ('ByIndex 1)) shape)) ('Shape '['Dim ('Name "*") ('Size 1), seqDim, seqDim])), output ~ Tensor ('Gradient 'WithoutGradient) (layout <+> 'Layout 'Dense) device transformerDataType (BroadcastShapesF (BroadcastShapesF ('Shape '['Dim ('Name "*") ('Size 1), seqDim, seqDim]) (UnsqueezeF ('SelectDim ('ByIndex 1)) shape)) ('Shape '['Dim ('Name "*") ('Size 1), seqDim, seqDim])))

mkTransformerDecoderAttentionMask

Arguments

:: forall m transformerDataType gradient layout device dataType shape seqDim output. (MonadThrow m, MkTransformerDecoderAttentionMaskC transformerDataType layout device shape seqDim output)

=> SDataType transformerDataType
data type singleton of the transformer

-> Double
attention mask bias (typically a large negative number)

-> Tensor gradient layout device dataType shape
decoder padding mask

-> m output

Creates a causal attention mask for a transformer decoder. Given a padding mask of shape [batchDim, seqDim], returns a tensor of shape [batchDim, seqDim, seqDim].

data MkTransformerDecoderAttentionMask (dataType :: DataType DType) where

Constructors

MkTransformerDecoderAttentionMask

Fields

:: forall dataType. { decoderAttentionMaskDataType :: SDataType dataType

, decoderAttentionMaskBias :: Double

} -> MkTransformerDecoderAttentionMask dataType

type MkTransformerCrossAttentionMaskC transformerDataType decoderInputShape decoderInputSeqDim gradient layout device dataType shape seqDim output = (SGetLayout layout, SGetDevice device, SGetShape shape, seqDim ~ (shape ! 1), SGetShape decoderInputShape, decoderInputSeqDim ~ (decoderInputShape ! 1), Catch (gradient <+> 'Gradient 'WithoutGradient), Catch (dataType <+> 'DataType 'Bool), Catch (UnsqueezeF ('SelectDim ('ByIndex 1)) shape), Catch (BroadcastShapesF (UnsqueezeF ('SelectDim ('ByIndex 1)) shape) ('Shape '['Dim ('Name "*") ('Size 1), decoderInputSeqDim, seqDim])), output ~ Tensor ('Gradient 'WithoutGradient) (layout <+> 'Layout 'Dense) device transformerDataType (BroadcastShapesF (UnsqueezeF ('SelectDim ('ByIndex 1)) shape) ('Shape '['Dim ('Name "*") ('Size 1), decoderInputSeqDim, seqDim])))

mkTransformerCrossAttentionMask

Arguments

:: forall m transformerDataType decoderInputShape decoderInputSeqDim gradient layout device dataType shape seqDim output. (MonadThrow m, MkTransformerCrossAttentionMaskC transformerDataType decoderInputShape decoderInputSeqDim gradient layout device dataType shape seqDim output)

=> SDataType transformerDataType
data type singleton of the transformer

-> SShape decoderInputShape
decoder input shape

-> Double
attention mask bias (typically a large negative number)

-> Tensor gradient layout device dataType shape
encoder padding mask

-> m output

Creates a cross-attention mask for an encoder-decoder transformer. Given an encoder padding mask of shape [batchDim, seqDim], and the shape [batchDim, decoderSeqDim] of the decoder's input, returns a tensor of shape [batchDim, decoderSeqDim, seqDim].

data MkTransformerCrossAttentionMask (dataType :: DataType DType) where

Constructors

MkTransformerCrossAttentionMask

Fields

:: forall dataType. { crossAttentionMaskDataType :: SDataType dataType

, crossAttentionMaskBias :: Double

} -> MkTransformerCrossAttentionMask dataType

data ShiftRight fillValue where

Constructors

ShiftRight

Fields

:: forall fillValue. fillValue

fill value for shift right

-> ShiftRight fillValue