Torch.GraduallyTyped.RequiresGradient

Documentation

data RequiresGradient

Data type to represent whether or not the tensor requires gradient computations.

Constructors

WithGradient
The tensor requires gradient computations.

WithoutGradient
Gradient computations for this tensor are disabled.

type WithGradientSym0 = 'WithGradient :: RequiresGradient

type WithoutGradientSym0 = 'WithoutGradient :: RequiresGradient

data SRequiresGradient z where

Constructors

SWithGradient :: SRequiresGradient ('WithGradient :: RequiresGradient)

SWithoutGradient :: SRequiresGradient ('WithoutGradient :: RequiresGradient)

class KnownRequiresGradient (requiresGradient :: RequiresGradient) where

Methods

requiresGradientVal :: RequiresGradient

data Gradient (requiresGradient :: Type) where

Data type to represent whether or not it is known by the compiler if the tensor requires gradient computations.

Constructors

UncheckedGradient :: forall requiresGradient. Gradient requiresGradient
Whether or not the tensor requires gradient computations is unknown to the compiler.

Gradient :: forall requiresGradient. requiresGradient -> Gradient requiresGradient
Whether or not the tensor requires gradient computations is known to the compiler.

data SGradient (gradient :: Gradient RequiresGradient) where

Constructors

SUncheckedGradient :: RequiresGradient -> SGradient 'UncheckedGradient

SGradient :: forall requiresGradient. SRequiresGradient requiresGradient -> SGradient ('Gradient requiresGradient)

class KnownGradient (gradient :: Gradient RequiresGradient) where

Methods

gradientVal :: Gradient RequiresGradient

type family GetGradients f where ...

Equations

GetGradients (a :: Gradient RequiresGradient) = '[a]

GetGradients (f g) = Concat (GetGradients f) (GetGradients g)

GetGradients _ = '[]