Torch.GraduallyTyped.Shape.Type

Documentation

data Size (size :: Type) where

Constructors

UncheckedSize :: forall size. Size size

Size :: forall size. size -> Size size

data SSize (size :: Size Nat) where

Constructors

SUncheckedSize :: Integer -> SSize 'UncheckedSize

SSize :: forall size. KnownNat size => SSize ('Size size)

type family SizeF (size :: Size Nat) :: Nat where ...

Equations

SizeF ('Size size) = size

class KnownSize (size :: Size Nat) where

Methods

sizeVal :: Size Integer

data Name (name :: Type) where

Constructors

UncheckedName :: forall name. Name name

Name :: forall name. name -> Name name

data SName (name :: Name Symbol) where

Constructors

SUncheckedName :: String -> SName 'UncheckedName

SName :: forall name. KnownSymbol name => SName ('Name name)

type family NameF (name :: Name Symbol) :: Symbol where ...

Equations

NameF ('Name name) = name

class KnownName (name :: Name Symbol) where

Methods

nameVal :: Name String

data Dim (name :: Type) (size :: Type) where

Constructors

Dim

Fields

:: forall name size. { dimName :: name

, dimSize :: size

} -> Dim name size

data SDim (dim :: Dim (Name Symbol) (Size Nat)) where

Constructors

SDim

Fields

:: forall name size. { sDimName :: SName name

, sDimSize :: SSize size

} -> SDim ('Dim name size)

pattern (:&:) :: forall (name :: Name Symbol) (size :: Size Nat). SName name -> SSize size -> SDim ('Dim name size) infix 9

class KnownDim (dim :: Dim (Name Symbol) (Size Nat)) where

Methods

dimVal :: Dim (Name String) (Size Integer)

data By (name :: Type) (index :: Type) where

Data type to select dimensions by name or by index.

Constructors

ByName :: forall name index. name -> By name index
Select a dimension by name.

ByIndex :: forall name index. index -> By name index
Select a dimension by index. Counting starts at zero for the first dimension.

data SBy (by :: By Symbol Nat) where

Constructors

SByName :: forall name. KnownSymbol name => SBy ('ByName name)

SByIndex :: forall index. KnownNat index => SBy ('ByIndex index)

type family ByNameF (by :: By Symbol Nat) :: Symbol where ...

Equations

ByNameF ('ByName name) = name

type family ByIndexF (by :: By Symbol Nat) :: Nat where ...

Equations

ByIndexF ('ByIndex index) = index

class KnownBy (by :: By Symbol Nat) where

Methods

byVal :: By String Integer

data SelectDim (by :: Type) where

Constructors

UncheckedSelectDim :: forall by. SelectDim by
Unknown method of dimension selection.

SelectDim :: forall by. by -> SelectDim by
Known method of dimension selection, that is, either by name or by index.

data SSelectDim (selectDim :: SelectDim (By Symbol Nat)) where

Constructors

SUncheckedSelectDim :: By String Integer -> SSelectDim 'UncheckedSelectDim

SSelectDim :: forall by. SBy by -> SSelectDim ('SelectDim by)

class KnownSelectDim (selectDim :: SelectDim (By Symbol Nat)) where

Methods

selectDimVal :: SelectDim (By String Integer)

data SelectDims (selectDims :: Type) where

Constructors

UncheckedSelectDims :: forall selectDims. SelectDims selectDims

SelectDims :: forall selectDims. selectDims -> SelectDims selectDims

data SSelectDims (selectDims :: SelectDims [By Symbol Nat]) where

Constructors

SUncheckedSelectDims :: [By String Integer] -> SSelectDims 'UncheckedSelectDims

SSelectDims :: forall bys. SList bys -> SSelectDims ('SelectDims bys)

class KnownSelectDims (selectDims :: SelectDims [By Symbol Nat]) where

Methods

selectDimsVal :: SelectDims [By String Integer]

data Shape (dims :: Type) where

Data type to represent tensor shapes, that is, lists of dimensions.

Constructors

UncheckedShape :: forall dims. Shape dims
The shape is fully unchecked. Neither the number of the dimensions nor any dimension properties are known to the compiler.

Shape :: forall dims. dims -> Shape dims
The shape is partially known to the compiler. The list of dimensions has a known length, but may contain UncheckedDim, that is, unknown dimensions.

data SShape (shape :: Shape [Dim (Name Symbol) (Size Nat)]) where

Constructors

SUncheckedShape :: [Dim String Integer] -> SShape 'UncheckedShape

SShape :: forall dims. SList dims -> SShape ('Shape dims)

pattern (:|:) :: forall (dim :: Dim (Name Symbol) (Size Nat)) (dims :: [Dim (Name Symbol) (Size Nat)]). SDim dim -> SList dims -> SList (dim ': dims) infixr 8

class KnownShape (shape :: Shape [Dim (Name Symbol) (Size Nat)]) where

Methods

shapeVal :: Shape [Dim (Name String) (Size Integer)]

type family GetShapes f where ...

Equations

GetShapes (a :: Shape [Dim (Name Symbol) (Size Nat)]) = '[a]

GetShapes (f g) = Concat (GetShapes f) (GetShapes g)

GetShapes _ = '[]

Orphan instances

Castable String (ForeignPtr Dimname)

Instance details

Methods

cast :: String -> (ForeignPtr Dimname -> IO r) -> IO r

uncast :: ForeignPtr Dimname -> (String -> IO r) -> IO r

Castable [Integer] (ForeignPtr IntArray)

Instance details

Methods

cast :: [Integer] -> (ForeignPtr IntArray -> IO r) -> IO r

uncast :: ForeignPtr IntArray -> ([Integer] -> IO r) -> IO r

Castable [String] (ForeignPtr DimnameList)

Instance details

Methods

cast :: [String] -> (ForeignPtr DimnameList -> IO r) -> IO r

uncast :: ForeignPtr DimnameList -> ([String] -> IO r) -> IO r

Castable [ForeignPtr Dimname] (ForeignPtr DimnameList)

Instance details

Methods

cast :: [ForeignPtr Dimname] -> (ForeignPtr DimnameList -> IO r) -> IO r

uncast :: ForeignPtr DimnameList -> ([ForeignPtr Dimname] -> IO r) -> IO r