Torch.GraduallyTyped.Tensor.Indexing

Documentation

data IndexType a

Constructors

None

Ellipsis

SliceAll

SliceAt a

SliceBool Bool

SliceFrom a

SliceUpTo a

SliceFromUpTo a a

SliceFromWithStride a a

SliceUpToWithStride a a

SliceFromUpToWithStride a a a

type NoneSym0 = 'None :: IndexType (a :: Type)

type EllipsisSym0 = 'Ellipsis :: IndexType (a :: Type)

type SliceAllSym0 = 'SliceAll :: IndexType (a :: Type)

type SliceAtSym1 (a6989586621679314323 :: a) = 'SliceAt a6989586621679314323 :: IndexType (a :: Type)

data SliceAtSym0 a6989586621679314323 where

Constructors

SliceAtSym0KindInference :: SameKind (Apply SliceAtSym0 arg) (SliceAtSym1 arg) => SliceAtSym0 a6989586621679314323

type SliceBoolSym1 (a6989586621679314325 :: Bool) = 'SliceBool a6989586621679314325 :: IndexType (a :: Type)

data SliceBoolSym0 a6989586621679314325 where

Constructors

SliceBoolSym0KindInference :: SameKind (Apply SliceBoolSym0 arg) (SliceBoolSym1 arg) => SliceBoolSym0 a6989586621679314325

type SliceFromSym1 (a6989586621679314327 :: a) = 'SliceFrom a6989586621679314327 :: IndexType (a :: Type)

data SliceFromSym0 a6989586621679314327 where

Constructors

SliceFromSym0KindInference :: SameKind (Apply SliceFromSym0 arg) (SliceFromSym1 arg) => SliceFromSym0 a6989586621679314327

type SliceUpToSym1 (a6989586621679314329 :: a) = 'SliceUpTo a6989586621679314329 :: IndexType (a :: Type)

data SliceUpToSym0 a6989586621679314329 where

Constructors

SliceUpToSym0KindInference :: SameKind (Apply SliceUpToSym0 arg) (SliceUpToSym1 arg) => SliceUpToSym0 a6989586621679314329

type SliceFromUpToSym2 (a6989586621679314331 :: a) (a6989586621679314332 :: a) = 'SliceFromUpTo a6989586621679314331 a6989586621679314332 :: IndexType (a :: Type)

data SliceFromUpToSym1 a6989586621679314331 a6989586621679314332 where

Constructors

SliceFromUpToSym1KindInference :: SameKind (Apply (SliceFromUpToSym1 a6989586621679314331) arg) (SliceFromUpToSym2 a6989586621679314331 arg) => SliceFromUpToSym1 a6989586621679314331 a6989586621679314332

data SliceFromUpToSym0 a6989586621679314331 where

Constructors

SliceFromUpToSym0KindInference :: SameKind (Apply SliceFromUpToSym0 arg) (SliceFromUpToSym1 arg) => SliceFromUpToSym0 a6989586621679314331

type SliceFromWithStrideSym2 (a6989586621679314334 :: a) (a6989586621679314335 :: a) = 'SliceFromWithStride a6989586621679314334 a6989586621679314335 :: IndexType (a :: Type)

data SliceFromWithStrideSym1 a6989586621679314334 a6989586621679314335 where

Constructors

SliceFromWithStrideSym1KindInference :: SameKind (Apply (SliceFromWithStrideSym1 a6989586621679314334) arg) (SliceFromWithStrideSym2 a6989586621679314334 arg) => SliceFromWithStrideSym1 a6989586621679314334 a6989586621679314335

data SliceFromWithStrideSym0 a6989586621679314334 where

Constructors

SliceFromWithStrideSym0KindInference :: SameKind (Apply SliceFromWithStrideSym0 arg) (SliceFromWithStrideSym1 arg) => SliceFromWithStrideSym0 a6989586621679314334

type SliceUpToWithStrideSym2 (a6989586621679314337 :: a) (a6989586621679314338 :: a) = 'SliceUpToWithStride a6989586621679314337 a6989586621679314338 :: IndexType (a :: Type)

data SliceUpToWithStrideSym1 a6989586621679314337 a6989586621679314338 where

Constructors

SliceUpToWithStrideSym1KindInference :: SameKind (Apply (SliceUpToWithStrideSym1 a6989586621679314337) arg) (SliceUpToWithStrideSym2 a6989586621679314337 arg) => SliceUpToWithStrideSym1 a6989586621679314337 a6989586621679314338

data SliceUpToWithStrideSym0 a6989586621679314337 where

Constructors

SliceUpToWithStrideSym0KindInference :: SameKind (Apply SliceUpToWithStrideSym0 arg) (SliceUpToWithStrideSym1 arg) => SliceUpToWithStrideSym0 a6989586621679314337

type SliceFromUpToWithStrideSym3 (a6989586621679314340 :: a) (a6989586621679314341 :: a) (a6989586621679314342 :: a) = 'SliceFromUpToWithStride a6989586621679314340 a6989586621679314341 a6989586621679314342 :: IndexType (a :: Type)

data SliceFromUpToWithStrideSym2 a6989586621679314340 a6989586621679314341 a6989586621679314342 where

Constructors

SliceFromUpToWithStrideSym2KindInference :: SameKind (Apply (SliceFromUpToWithStrideSym2 a6989586621679314340 a6989586621679314341) arg) (SliceFromUpToWithStrideSym3 a6989586621679314340 a6989586621679314341 arg) => SliceFromUpToWithStrideSym2 a6989586621679314340 a6989586621679314341 a6989586621679314342

data SliceFromUpToWithStrideSym1 a6989586621679314340 a6989586621679314341 where

Constructors

SliceFromUpToWithStrideSym1KindInference :: SameKind (Apply (SliceFromUpToWithStrideSym1 a6989586621679314340) arg) (SliceFromUpToWithStrideSym2 a6989586621679314340 arg) => SliceFromUpToWithStrideSym1 a6989586621679314340 a6989586621679314341

data SliceFromUpToWithStrideSym0 a6989586621679314340 where

Constructors

SliceFromUpToWithStrideSym0KindInference :: SameKind (Apply SliceFromUpToWithStrideSym0 arg) (SliceFromUpToWithStrideSym1 arg) => SliceFromUpToWithStrideSym0 a6989586621679314340

data SIndexType z where

Constructors

SNone :: forall (a :: Type). SIndexType ('None :: IndexType (a :: Type))

SEllipsis :: forall (a :: Type). SIndexType ('Ellipsis :: IndexType (a :: Type))

SSliceAll :: forall (a :: Type). SIndexType ('SliceAll :: IndexType (a :: Type))

SSliceAt :: forall (a :: Type) (n :: a). (Sing n) -> SIndexType ('SliceAt n :: IndexType (a :: Type))

SSliceBool :: forall (a :: Type) (n :: Bool). (Sing n) -> SIndexType ('SliceBool n :: IndexType (a :: Type))

SSliceFrom :: forall (a :: Type) (n :: a). (Sing n) -> SIndexType ('SliceFrom n :: IndexType (a :: Type))

SSliceUpTo :: forall (a :: Type) (n :: a). (Sing n) -> SIndexType ('SliceUpTo n :: IndexType (a :: Type))

SSliceFromUpTo :: forall (a :: Type) (n :: a) (n :: a). (Sing n) -> (Sing n) -> SIndexType ('SliceFromUpTo n n :: IndexType (a :: Type))

SSliceFromWithStride :: forall (a :: Type) (n :: a) (n :: a). (Sing n) -> (Sing n) -> SIndexType ('SliceFromWithStride n n :: IndexType (a :: Type))

SSliceUpToWithStride :: forall (a :: Type) (n :: a) (n :: a). (Sing n) -> (Sing n) -> SIndexType ('SliceUpToWithStride n n :: IndexType (a :: Type))

SSliceFromUpToWithStride :: forall (a :: Type) (n :: a) (n :: a) (n :: a). (Sing n) -> (Sing n) -> (Sing n) -> SIndexType ('SliceFromUpToWithStride n n n :: IndexType (a :: Type))

type family ReverseShape shape where ...

Equations

ReverseShape 'UncheckedShape = 'UncheckedShape

ReverseShape ('Shape dims) = 'Shape (Reverse dims)

type family ErrorOnEllipsis indices where ...

Equations

ErrorOnEllipsis '[] = '[]

ErrorOnEllipsis ('Ellipsis ': ixs) = TypeError ('Text "Indices can only contain a single ellipsis.")

ErrorOnEllipsis (ix ': ixs) = ix ': ErrorOnEllipsis ixs

type StrideZeroErrorMessage = 'Text "Slice stride cannot be zero"

type family Strided length stride where ...

Calculate the size of the dimension from stride.

>>> :kind! Strided 8 1
Strided 8 1 :: Nat
= 8
>>> :kind! Strided 5 2
Strided 5 2 :: Nat
= 3
>>> :kind! Strided 6 3
Strided 6 3 :: Nat
= 2

Equations

Strided _ 0 = TypeError StrideZeroErrorMessage

Strided 0 _ = 0

Strided length stride = ((length - 1) `Div` stride) + 1

type family CheckUpTo (upTo :: Nat) ok where ...

Equations

CheckUpTo upTo ok = If (upTo == 0) (TypeError ('Text "Slice 'upTo' type parameter must not be equal to zero")) ok

type family CheckFromSize (from :: Nat) (size :: Nat) ok where ...

Equations

CheckFromSize from size ok = If (from <? size) ok (TypeError ("Slice 'from' type parameter must be smaller than the size of the indexed dimension:" % ((" " <> "from < size") % ("but" % (((" " <> from) <> " >= ") <> size)))))

type family CheckUpToSize (upTo :: Nat) (size :: Size Nat) ok where ...

Equations

CheckUpToSize upTo 'UncheckedSize ok = CheckUpTo upTo ok

CheckUpToSize upTo ('Size size) ok = CheckUpTo upTo (If (upTo <=? size) ok (TypeError ("Slice 'upTo' type parameter must be less than or equal to the size of the indexed dimension:" % ((" " <> "upTo <= size") % ("but" % (((" " <> upTo) <> " > ") <> size))))))

type family CheckFromUpTo (from :: Nat) (upTo :: Nat) ok where ...

Equations

CheckFromUpTo from upTo ok = If (from <? upTo) ok (TypeError ("Slice 'from' type parameter must be less than the 'upTo' type parameter:" % ((" " <> "from < upTo") % ("but" % (((" " <> from) <> " >= ") <> upTo)))))

type family CheckFromUpToSize (from :: Nat) (upTo :: Nat) (size :: Size Nat) ok where ...

Equations

CheckFromUpToSize from upTo size ok = CheckFromUpTo from upTo (CheckUpToSize upTo size ok)

type family CheckSliceAt (at :: Nat) (size :: Nat) ok where ...

Equations

CheckSliceAt at size ok = If (at <? size) ok (TypeError ((("Index of 'SliceAt " <> at) <> "' must be less than the dimension size:") % ((" " <> "at < size") % ("but" % (((" " <> at) <> " >= ") <> size)))))

type family IndexDimsImpl indices dims where ...

Equations

IndexDimsImpl '[] dims = 'Shape dims

IndexDimsImpl ('None ': ixs) dims = 'Dim ('Name "*") ('Size 1) `PrependDimF` IndexDimsImpl ixs dims

IndexDimsImpl ('Ellipsis ': _) '[] = 'Shape '[]

IndexDimsImpl ('Ellipsis ': ixs) dims = ReverseShape (IndexDimsImpl (Reverse (ErrorOnEllipsis ixs)) (Reverse dims))

IndexDimsImpl ('SliceAll ': ixs) (dim ': dims) = dim `PrependDimF` IndexDimsImpl ixs dims

IndexDimsImpl ('SliceAt ('Index at) ': ixs) ('Dim name ('Size size) ': dims) = CheckSliceAt at size (IndexDimsImpl ixs dims)

IndexDimsImpl ('SliceAt _ ': ixs) ('Dim name _ ': dims) = IndexDimsImpl ixs dims

IndexDimsImpl ('SliceBool 'False ': ixs) ('Dim name _ ': dims) = 'Dim name ('Size 0) `PrependDimF` IndexDimsImpl ixs dims

IndexDimsImpl ('SliceBool 'True ': ixs) ('Dim name _ ': dims) = 'Dim name ('Size 1) `PrependDimF` IndexDimsImpl ixs dims

IndexDimsImpl ('SliceFrom ('Index from) ': ixs) ('Dim name ('Size size) ': dims) = CheckFromSize from size ('Dim name ('Size (size - from)) `PrependDimF` IndexDimsImpl ixs dims)

IndexDimsImpl ('SliceFrom _ ': ixs) ('Dim name _ ': dims) = 'Dim name 'UncheckedSize `PrependDimF` IndexDimsImpl ixs dims

IndexDimsImpl ('SliceUpTo ('Index upTo) ': ixs) ('Dim name size ': dims) = CheckUpToSize upTo size ('Dim name ('Size upTo) `PrependDimF` IndexDimsImpl ixs dims)

IndexDimsImpl ('SliceUpTo _ ': ixs) ('Dim name _ ': dims) = 'Dim name 'UncheckedSize `PrependDimF` IndexDimsImpl ixs dims

IndexDimsImpl ('SliceFromUpTo ('Index from) ('Index upTo) ': ixs) ('Dim name size ': dims) = CheckFromUpToSize from upTo size ('Dim name ('Size (upTo - from)) `PrependDimF` IndexDimsImpl ixs dims)

IndexDimsImpl ('SliceFromUpTo _ _ ': ixs) ('Dim name _ ': dims) = 'Dim name 'UncheckedSize `PrependDimF` IndexDimsImpl ixs dims

IndexDimsImpl ('SliceFromWithStride ('Index from) ('Index stride) ': ixs) ('Dim name ('Size size) ': dims) = CheckFromSize from size ('Dim name ('Size (Strided (size - from) stride)) `PrependDimF` IndexDimsImpl ixs dims)

IndexDimsImpl ('SliceFromWithStride _ _ ': ixs) ('Dim name _ ': dims) = 'Dim name 'UncheckedSize `PrependDimF` IndexDimsImpl ixs dims

IndexDimsImpl ('SliceUpToWithStride ('Index upTo) ('Index stride) ': ixs) ('Dim name size ': dims) = CheckUpToSize upTo size ('Dim name ('Size (Strided upTo stride)) `PrependDimF` IndexDimsImpl ixs dims)

IndexDimsImpl ('SliceUpToWithStride _ _ ': ixs) ('Dim name _ ': dims) = 'Dim name 'UncheckedSize `PrependDimF` IndexDimsImpl ixs dims

IndexDimsImpl ('SliceFromUpToWithStride ('Index from) ('Index upTo) ('Index stride) ': ixs) ('Dim name size ': dims) = CheckFromUpToSize from upTo size ('Dim name ('Size (Strided (upTo - from) stride)) `PrependDimF` IndexDimsImpl ixs dims)

IndexDimsImpl ('SliceFromUpToWithStride _ _ _ ': ixs) ('Dim name _ ': dims) = 'Dim name 'UncheckedSize `PrependDimF` IndexDimsImpl ixs dims

type family IndexDims indices shape where ...

Equations

IndexDims 'UncheckedIndices _ = 'UncheckedShape

IndexDims _ 'UncheckedShape = 'UncheckedShape

IndexDims ('Indices indices) ('Shape dims) = IndexDimsImpl indices dims

data Indices (dimIndices :: Type) where

Constructors

UncheckedIndices :: forall dimIndices. Indices dimIndices

Indices :: forall dimIndices. dimIndices -> Indices dimIndices

data SIndices (indices :: Indices [IndexType (Index Nat)]) where

Constructors

SUncheckedIndices :: [IndexType Integer] -> SIndices 'UncheckedIndices

SIndices :: forall dimIndices. SList dimIndices -> SIndices ('Indices dimIndices)

(!) :: forall indices requiresGradient layout device dataType shape m. MonadThrow m => Tensor requiresGradient layout device dataType shape -> SIndices indices -> m (Tensor requiresGradient layout device dataType (IndexDims indices shape))