Control.Monad.Fresh

Documentation

newtype Successor a

Constructors

Successor

Fields

suc :: a -> a

newtype FreshT e m a

The monad transformer for generating fresh values.

Constructors

FreshT

Fields

unFreshT :: ReaderT (Successor e) (StateT e m) a

type Fresh e = FreshT e Identity

successor :: forall e. (e -> e) -> Successor e

enumSucc :: forall e. Enum e => Successor e

runFreshT :: forall e m a. (Enum e, Monad m) => FreshT e m a -> m a

Run a FreshT computation starting from the value toEnum 0

runFresh :: forall e a. Enum e => Fresh e a -> a

Run a Fresh computation starting from the value toEnum 0

runFreshTFrom :: forall e m a. (Monad m, Enum e) => e -> FreshT e m a -> m a

Run a FreshT computation starting from a specific value e.

runFreshFrom :: forall e a. Enum e => e -> Fresh e a -> a

Run a Fresh computation starting from a specific value e.

runFreshTWith :: forall e m a. Monad m => Successor e -> e -> FreshT e m a -> m a

Run a FreshT computation starting from a specific value e with a the next fresh value determined by Successor e.

runFreshWith :: forall e a. Successor e -> e -> Fresh e a -> a

Run a FreshT computation starting from a specific value e with a the next fresh value determined by Successor e.

class Monad m => MonadFresh e m | m -> e where

The MTL style class for generating fresh values

Methods

fresh :: m e

Generate a fresh value e, fresh should never produce the same value within a monadic computation.