Torch.GraduallyTyped.DType

Documentation

type BoolSym0 = 'Bool :: DType

type UInt8Sym0 = 'UInt8 :: DType

type Int8Sym0 = 'Int8 :: DType

type Int16Sym0 = 'Int16 :: DType

type Int32Sym0 = 'Int32 :: DType

type Int64Sym0 = 'Int64 :: DType

type HalfSym0 = 'Half :: DType

type FloatSym0 = 'Float :: DType

type DoubleSym0 = 'Double :: DType

type ComplexHalfSym0 = 'ComplexHalf :: DType

type ComplexFloatSym0 = 'ComplexFloat :: DType

type ComplexDoubleSym0 = 'ComplexDouble :: DType

type QInt8Sym0 = 'QInt8 :: DType

type QUInt8Sym0 = 'QUInt8 :: DType

type QInt32Sym0 = 'QInt32 :: DType

type BFloat16Sym0 = 'BFloat16 :: DType

data SDType z where

Constructors

SBool :: SDType ('Bool :: DType)

SUInt8 :: SDType ('UInt8 :: DType)

SInt8 :: SDType ('Int8 :: DType)

SInt16 :: SDType ('Int16 :: DType)

SInt32 :: SDType ('Int32 :: DType)

SInt64 :: SDType ('Int64 :: DType)

SHalf :: SDType ('Half :: DType)

SFloat :: SDType ('Float :: DType)

SDouble :: SDType ('Double :: DType)

SComplexHalf :: SDType ('ComplexHalf :: DType)

SComplexFloat :: SDType ('ComplexFloat :: DType)

SComplexDouble :: SDType ('ComplexDouble :: DType)

SQInt8 :: SDType ('QInt8 :: DType)

SQUInt8 :: SDType ('QUInt8 :: DType)

SQInt32 :: SDType ('QInt32 :: DType)

SBFloat16 :: SDType ('BFloat16 :: DType)

class KnownDType (dType :: DType) where

Methods

dTypeVal :: DType

data DataType (dType :: Type) where

Data type to represent whether or not the tensor data type is checked, that is, known to the compiler.

Constructors

UncheckedDataType :: forall dType. DataType dType
The tensor data type is unknown to the compiler.

DataType :: forall dType. dType -> DataType dType
The tensor data type is known to the compiler.

data SDataType (dataType :: DataType DType) where

Constructors

SUncheckedDataType :: DType -> SDataType 'UncheckedDataType

SDataType :: forall dType. SDType dType -> SDataType ('DataType dType)

class KnownDataType (dataType :: DataType DType) where

Methods

dataTypeVal :: DataType DType

type family GetDataTypes f where ...

Equations

GetDataTypes (a :: DataType DType) = '[a]

GetDataTypes (f g) = Concat (GetDataTypes f) (GetDataTypes g)

GetDataTypes _ = '[]

Orphan instances

SingKind DType

Instance details

Associated Types

type Demote DType = (r :: Type)

Methods

fromSing :: forall (a :: DType). Sing a -> Demote DType

toSing :: Demote DType -> SomeSing DType

SingI 'BFloat16

Instance details

Methods

sing :: Sing 'BFloat16

SingI 'QInt32

Instance details

Methods

sing :: Sing 'QInt32

SingI 'QUInt8

Instance details

Methods

sing :: Sing 'QUInt8

SingI 'QInt8

Instance details

Methods

sing :: Sing 'QInt8

SingI 'ComplexDouble

Instance details

Methods

sing :: Sing 'ComplexDouble

SingI 'ComplexFloat

Instance details

Methods

sing :: Sing 'ComplexFloat

SingI 'ComplexHalf

Instance details

Methods

sing :: Sing 'ComplexHalf

SingI 'Double

Instance details

Methods

sing :: Sing 'Double

SingI 'Float

Instance details

Methods

sing :: Sing 'Float

SingI 'Half

Instance details

Methods

sing :: Sing 'Half

SingI 'Int64

Instance details

Methods

sing :: Sing 'Int64

SingI 'Int32

Instance details

Methods

sing :: Sing 'Int32

SingI 'Int16

Instance details

Methods

sing :: Sing 'Int16

SingI 'Int8

Instance details

Methods

sing :: Sing 'Int8

SingI 'UInt8

Instance details

Methods

sing :: Sing 'UInt8

SingI 'Bool

Instance details

Methods

sing :: Sing 'Bool