Torch.GraduallyTyped.DType

Documentation

data DType

Constructors

Bool
Bool

UInt8
Byte

Int8
Char

Int16
Short

Int32
Int

Int64
Long

Half
Half

Float
Float

Double
Double

ComplexHalf
ComplexHalf

ComplexFloat
ComplexFloat

ComplexDouble
ComplexDouble

QInt8
QInt8

QUInt8
QUInt8

QInt32
QInt32

BFloat16
BFloat16

type family BoolSym0 :: DType where ...

Equations

BoolSym0 = 'Bool

type family UInt8Sym0 :: DType where ...

Equations

UInt8Sym0 = 'UInt8

type family Int8Sym0 :: DType where ...

Equations

Int8Sym0 = 'Int8

type family Int16Sym0 :: DType where ...

Equations

Int16Sym0 = 'Int16

type family Int32Sym0 :: DType where ...

Equations

Int32Sym0 = 'Int32

type family Int64Sym0 :: DType where ...

Equations

Int64Sym0 = 'Int64

type family HalfSym0 :: DType where ...

Equations

HalfSym0 = 'Half

type family FloatSym0 :: DType where ...

Equations

FloatSym0 = 'Float

type family DoubleSym0 :: DType where ...

Equations

DoubleSym0 = 'Double

type family ComplexHalfSym0 :: DType where ...

Equations

ComplexHalfSym0 = 'ComplexHalf

type family ComplexFloatSym0 :: DType where ...

Equations

ComplexFloatSym0 = 'ComplexFloat

type family ComplexDoubleSym0 :: DType where ...

Equations

ComplexDoubleSym0 = 'ComplexDouble

type family QInt8Sym0 :: DType where ...

Equations

QInt8Sym0 = 'QInt8

type family QUInt8Sym0 :: DType where ...

Equations

QUInt8Sym0 = 'QUInt8

type family QInt32Sym0 :: DType where ...

Equations

QInt32Sym0 = 'QInt32

type family BFloat16Sym0 :: DType where ...

Equations

BFloat16Sym0 = 'BFloat16

data SDType :: DType -> Type where

Constructors

SBool :: SDType ('Bool :: DType)

SUInt8 :: SDType ('UInt8 :: DType)

SInt8 :: SDType ('Int8 :: DType)

SInt16 :: SDType ('Int16 :: DType)

SInt32 :: SDType ('Int32 :: DType)

SInt64 :: SDType ('Int64 :: DType)

SHalf :: SDType ('Half :: DType)

SFloat :: SDType ('Float :: DType)

SDouble :: SDType ('Double :: DType)

SComplexHalf :: SDType ('ComplexHalf :: DType)

SComplexFloat :: SDType ('ComplexFloat :: DType)

SComplexDouble :: SDType ('ComplexDouble :: DType)

SQInt8 :: SDType ('QInt8 :: DType)

SQUInt8 :: SDType ('QUInt8 :: DType)

SQInt32 :: SDType ('QInt32 :: DType)

SBFloat16 :: SDType ('BFloat16 :: DType)

class KnownDType (dType :: DType) where

Methods

dTypeVal :: DType

data DataType (dType :: Type) where

Data type to represent whether or not the tensor data type is checked, that is, known to the compiler.

Constructors

UncheckedDataType :: forall dType. DataType dType
The tensor data type is unknown to the compiler.

DataType :: forall dType. dType -> DataType dType
The tensor data type is known to the compiler.

data SDataType (dataType :: DataType DType) where

Constructors

SUncheckedDataType :: DType -> SDataType 'UncheckedDataType

SDataType :: forall dType. SDType dType -> SDataType ('DataType dType)

class KnownDataType (dataType :: DataType DType) where

Methods

dataTypeVal :: DataType DType

type family GetDataTypes f where ...

Equations

GetDataTypes (a :: DataType DType) = '[a]

GetDataTypes (f g) = Concat (GetDataTypes f) (GetDataTypes g)

GetDataTypes _ = '[]