Torch.GraduallyTyped.Device

Documentation

data DeviceType (deviceId :: Type) where

Data type to represent compute devices.

Constructors

CPU :: forall deviceId. DeviceType deviceId
The tensor is stored in the CPU's memory.

CUDA :: forall deviceId. deviceId -> DeviceType deviceId
The tensor is stored the memory of the GPU with ID deviceId.

data SDeviceType (deviceType :: DeviceType Nat) where

Constructors

SCPU :: SDeviceType 'CPU

SCUDA :: forall deviceId. KnownNat deviceId => SDeviceType ('CUDA deviceId)

type family CUDAF (deviceType :: DeviceType Nat) :: Nat where ...

Equations

CUDAF ('CUDA deviceId) = deviceId

class KnownDeviceType (deviceType :: DeviceType Nat) where

Methods

deviceTypeVal :: DeviceType Int16

data Device (deviceType :: Type) where

Data type to represent whether or not the compute device is checked, that is, known to the compiler.

Constructors

UncheckedDevice :: forall deviceType. Device deviceType
The compute device is unknown to the compiler.

Device :: forall deviceType. deviceType -> Device deviceType
The compute device is known to the compiler.

data SDevice (deviceType :: Device (DeviceType Nat)) where

Constructors

SUncheckedDevice :: DeviceType Int16 -> SDevice 'UncheckedDevice

SDevice :: forall deviceType. SDeviceType deviceType -> SDevice ('Device deviceType)

class KnownDevice (device :: Device (DeviceType Nat)) where

Methods

deviceVal :: Device (DeviceType Int16)

type family GetDevices f where ...

Equations

GetDevices (a :: Device (DeviceType Nat)) = '[a]

GetDevices (f g) = Concat (GetDevices f) (GetDevices g)

GetDevices _ = '[]