Torch.GraduallyTyped.Examples.TwoLayerNetwork

Documentation

sinc :: Floating a => a -> a

Compute the sine cardinal (sinc) function, see https://mathworld.wolfram.com/SincFunction.html.

noisySinc :: (Floating a, Random a, RandomGen g) => a -> a -> g -> (a, g)

Compute the sine cardinal (sinc) function and add normally distributed noise of strength epsilon. We use the normal function from Normal which requires a random number generator with RandomGen instance.

data SincData

Datatype to represent a dataset of sine cardinal (sinc) inputs and outputs. The name field is used to identify the split of the dataset.

Constructors

SincData

Fields

name :: Text

unSincData :: [(Float, Float)]

mkSincData

Arguments

:: (RandomGen g, Monad m)

=> Text
name of the dataset

-> Int
number of samples

-> StateT g m SincData
dataset in the state monad over the random generator

Create a dataset of noisy sine cardinal (sinc) values of a desired size.

newtype TwoLayerNetwork fstLayer activation dropout sndLayer

Data type to represent a simple two-layer neural network. It is a product type of two fully connected layers, fstLayer and sndLayer, an activation function, activation, and a dropout layer, dropout.

Constructors

TwoLayerNetwork (ModelStack '[fstLayer, activation, dropout, sndLayer])

data Monitor

Data type for monitoring the training and evaluation losses.

Constructors

TrainingMonitor
monitor for training loss

Fields

mtLoss :: Float

mtEpoch :: Int

EvaluationMonitor
monitor for evaluation loss

Fields

meLoss :: Float

meEpoch :: Int

monitor :: MonadIO m => Consumer Monitor m r

A simple monitor that prints the training and evaluation losses to stdout.

runTwoLayerNetworkExample :: IO ()