Torch.GraduallyTyped.Examples.TwoLayerNetwork

Documentation

sinc :: Floating a => a -> a

Compute the sine cardinal (sinc) function, see https://mathworld.wolfram.com/SincFunction.html.

noisySinc :: (Floating a, Random a, RandomGen g) => a -> a -> g -> (a, g)

Compute the sine cardinal (sinc) function and add normally distributed noise of strength epsilon. We use the normal function from Normal which requires a random number generator with RandomGen instance.

data SincData

Datatype to represent a dataset of sine cardinal (sinc) inputs and outputs. The name field is used to identify the split of the dataset.

Constructors

SincData

Fields

name :: Text

unSincData :: [(Float, Float)]

mkSincData

Arguments

:: (RandomGen g, Monad m)

=> Text
name of the dataset

-> Int
number of samples

-> StateT g m SincData
dataset in the state monad over the random generator

Create a dataset of noisy sine cardinal (sinc) values of a desired size.

newtype TwoLayerNetwork fstLayer activation dropout sndLayer

Data type to represent a simple two-layer neural network. It is a product type of two layer types, fstLayer and sndLayer, an activation function, activation, and a dropout layer, dropout.

Constructors

TwoLayerNetwork (ModelStack '[fstLayer, activation, dropout, sndLayer])

type TwoLayerNetworkF gradient hasDropout device dataType inputDim outputDim hiddenDim = NamedModel (TwoLayerNetwork (NamedModel (GLinearF 'WithBias gradient device dataType inputDim hiddenDim)) Tanh (TLNDropoutF hasDropout) (NamedModel (GLinearF 'WithBias gradient device dataType hiddenDim outputDim)))

Specifies the type of the two-layer network.

type family TLNDropoutF hasDropout where ...

Specifies the type of the dropout layer

Equations

TLNDropoutF 'WithDropout = Dropout

TLNDropoutF 'WithoutDropout = ()

twoLayerNetworkSpec

Arguments

:: forall gradient hasDropout device dataType inputDim outputDim hiddenDim. SGradient gradient
whether or not to compute gradients for the parametrs

-> SHasDropout hasDropout
whether or not to use dropout

-> SDevice device
which device to use

-> SDataType dataType
which data type to use

-> SDim inputDim
input dimension

-> SDim outputDim
output dimension

-> SDim hiddenDim
hidden dimension

-> Double
dropout rate

-> ModelSpec (TwoLayerNetworkF gradient hasDropout device dataType inputDim outputDim hiddenDim)
specification for the network

Creates a value that specifies the parameters of a two-layer neural network.

data Monitor

Data type for monitoring the training and evaluation losses.

Constructors

TrainingMonitor
monitor for training loss

Fields

mtLoss :: Float

mtEpoch :: Int

EvaluationMonitor
monitor for evaluation loss

Fields

meLoss :: Float

meEpoch :: Int

monitor :: MonadIO m => Consumer Monitor m r

A simple monitor that prints the training and evaluation losses to stdout.

collate :: forall device r input target. (input ~ Tensor ('Gradient 'WithoutGradient) ('Layout 'Dense) device ('DataType 'Float) ('Shape '['Dim ('Name "*") 'UncheckedSize, 'Dim ('Name "*") ('Size 1)]), target ~ Tensor ('Gradient 'WithoutGradient) ('Layout 'Dense) device ('DataType 'Float) ('Shape '['Dim ('Name "*") 'UncheckedSize, 'Dim ('Name "*") ('Size 1)])) => SDevice device -> Int -> ListT IO (Float, Float) -> ContT r IO (ListT IO (input, target))

Collate a stream of examples into batches. The returned batches are of the form (input, target), where (input, target) is a pair of tensors. input and target are both of shape:

Dim (SName @"*") (UncheckedSize batchSize) :|: Dim (SName @"*") (SSize @1) :|: SNil

where batchSize is the number of examples in the batch.

runTwoLayerNetworkExample :: IO ()

Run the two-layer network training loop on a toy dataset.