Torch.GraduallyTyped.Layout

Documentation

data LayoutType

Data type that represents the memory layout of a tensor.

Constructors

Dense
The memory layout of the tensor is dense (strided).

Sparse
The memory layout of the tensor is sparse.

type DenseSym0 = 'Dense :: LayoutType

type SparseSym0 = 'Sparse :: LayoutType

data SLayoutType z where

Constructors

SDense :: SLayoutType ('Dense :: LayoutType)

SSparse :: SLayoutType ('Sparse :: LayoutType)

class KnownLayoutType (layoutType :: LayoutType) where

Methods

layoutTypeVal :: LayoutType

data Layout (layoutType :: Type) where

Data type to represent whether or not the tensor's memory layout is checked, that is, known to the compiler.

Constructors

UncheckedLayout :: forall layoutType. Layout layoutType
The tensor's memory layout is unknown to the compiler.

Layout :: forall layoutType. layoutType -> Layout layoutType
The tensor's memory layout is known to the compiler.

data SLayout (layout :: Layout LayoutType) where

Constructors

SUncheckedLayout :: LayoutType -> SLayout 'UncheckedLayout

SLayout :: forall layoutType. SLayoutType layoutType -> SLayout ('Layout layoutType)

class KnownLayout (layout :: Layout LayoutType) where

Methods

layoutVal :: Layout LayoutType

type family GetLayouts f where ...

Equations

GetLayouts (a :: Layout LayoutType) = '[a]

GetLayouts (f g) = Concat (GetLayouts f) (GetLayouts g)

GetLayouts _ = '[]