Torch.GraduallyTyped.NN.Class

Documentation

data NamedModel model

Constructors

NamedModel Text model

class HasForward model input generatorDevice output generatorOutputDevice | model input generatorDevice -> output, model input generatorDevice -> generatorOutputDevice where

Methods

forward

Arguments

:: forall m. MonadThrow m

=> model
model

-> input
model input, typically a tensor or a tuple of tensors

-> Generator generatorDevice
random generator

-> m (output, Generator generatorOutputDevice)
output of the model with an updated generator

forward m i g for a model m, an input i, and a generator g returns the tuple (o, g') where o is the output of the model applied to the input and g' is the updated generator. forward m i g may throw an exception if the input i or the generator g are not compatible with the model m.

type family ModelSpec model = (spec :: Type) | spec -> model

class HasInitialize (model :: Type) (generatorDevice :: Device (DeviceType Nat)) (output :: Type) (generatorOutputDevice :: Device (DeviceType Nat)) | model generatorDevice -> output, model generatorDevice -> generatorOutputDevice where

Methods

initialize :: forall m. MonadThrow m => ModelSpec model -> Generator generatorDevice -> m (output, Generator generatorOutputDevice)

data VectorSpec (n :: Nat) (a :: Type) where

Constructors

VectorSpec :: forall n a. SNat n -> Vector n (ModelSpec a) -> VectorSpec n a

type StateDictKey = Text

type StateDict = Map StateDictKey (ForeignPtr Tensor)

newtype FromStateDictError

Constructors

FromStateDictKeyNotFoundError

Fields

fsdeExpectedKey :: StateDictKey

newtype ToStateDictError

Constructors

ToStateDictKeyAlreadyInUseError

Fields

fsdeTakenKey :: StateDictKey

class HasStateDict model where

Methods

fromStateDict :: forall m. (MonadIO m, MonadThrow m, MonadState StateDict m) => ModelSpec model -> StateDictKey -> m model

toStateDict :: forall m. (MonadThrow m, MonadState StateDict m) => StateDictKey -> model -> m ()

stateDictFromFile :: FilePath -> IO StateDict

Load a state dictionary from a TorchScript file.

stateDictToFile :: StateDict -> FilePath -> IO ()

Save a state dictionary to a TorchScript file.