Torch.GraduallyTyped.NN.Class

Documentation

class HasForward model input generator output generatorOutput | model input generator -> output, model input generator -> generatorOutput where

Methods

forward :: forall m. MonadThrow m => model -> input -> generator -> m (output, generatorOutput)

forward m i g for a model m, an input i, and a generator g returns the tuple (o, g') where o is the output of the model applied to the input and g' is the updated generator. forward m i g may throw an exception if the input i or the generator g are not compatible with the model m.

class HasInitialize model input generator generatorOutput | model -> input, model generator -> generatorOutput where

Methods

initialize :: input -> generator -> (model, generatorOutput)

type StateDictKey = String

type StateDict = Map StateDictKey (ForeignPtr Tensor)

newtype FromStateDictError

Constructors

FromStateDictKeyNotFoundError

Fields

fsdeExpectedKey :: StateDictKey

newtype ToStateDictError

Constructors

ToStateDictKeyAlreadyInUseError

Fields

fsdeTakenKey :: StateDictKey

class HasStateDict model input | model -> input where

Methods

fromStateDict :: forall m. (MonadThrow m, MonadState StateDict m) => input -> StateDictKey -> m model

toStateDict :: forall m. (MonadThrow m, MonadState StateDict m) => StateDictKey -> model -> m ()

stateDictFromPretrained :: FilePath -> IO StateDict