Torch.GraduallyTyped.NN.Class

Documentation

data NamedModel model

Constructors

NamedModel Text model

pattern (::>) :: Text -> model -> NamedModel model

class HasForward model input generatorDevice output generatorOutputDevice | model input generatorDevice -> output, model input generatorDevice -> generatorOutputDevice where

Minimal complete definition

Nothing

Methods

forward

Arguments

:: forall m. MonadThrow m

=> model
model

-> input
model input, typically a tensor or a tuple of tensors

-> Generator generatorDevice
random generator

-> m (output, Generator generatorOutputDevice)
output of the model with an updated generator

forward m i g for a model m, an input i, and a generator g returns the tuple (o, g') where o is the output of the model applied to the input and g' is the updated generator. forward m i g may throw an exception if the input i or the generator g are not compatible with the model m.

default forward :: forall m. (MonadThrow m, Generic model, GHasForward (Rep model) input generatorDevice output generatorOutputDevice) => model -> input -> Generator generatorDevice -> m (output, Generator generatorOutputDevice)

class GHasForward gModel input generatorDevice output generatorOutputDevice | gModel input generatorDevice -> output, gModel input generatorDevice -> generatorOutputDevice where

Methods

gForward :: forall m c. MonadThrow m => gModel c -> input -> Generator generatorDevice -> m (output, Generator generatorOutputDevice)

newtype Wrap a

Constructors

Wrap a

type family ListToTuple xs = tuple | tuple -> xs where ...

Equations

ListToTuple '[] = ()

ListToTuple '[a] = Wrap a

ListToTuple '[a, b] = (a, b)

ListToTuple '[a, b, c] = (a, b, c)

ListToTuple '[a, b, c, d] = (a, b, c, d)

ListToTuple '[a, b, c, d, e] = (a, b, c, d, e)

ListToTuple '[a, b, c, d, e, f] = (a, b, c, d, e, f)

ListToTuple '[a, b, c, d, e, f, g] = (a, b, c, d, e, f, g)

newtype ModelStack models

Constructors

ModelStack (ListToTuple models)

type family ModelSpec model = spec | spec -> model

class HasInitialize model generatorDevice output generatorOutputDevice | model generatorDevice -> output, model generatorDevice -> generatorOutputDevice where

Minimal complete definition

Nothing

Methods

initialize :: forall m. MonadThrow m => ModelSpec model -> Generator generatorDevice -> m (output, Generator generatorOutputDevice)

default initialize :: forall m. (MonadThrow m, Generic (ModelSpec model), Generic output, GHasInitialize (Rep (ModelSpec model)) generatorDevice (Rep output) generatorOutputDevice) => ModelSpec model -> Generator generatorDevice -> m (output, Generator generatorOutputDevice)

class GHasInitialize gModelSpec generatorDevice gOutput generatorOutputDevice | gModelSpec generatorDevice -> gOutput, gModelSpec generatorDevice -> generatorOutputDevice where

Methods

gInitialize :: forall m c. MonadThrow m => gModelSpec c -> Generator generatorDevice -> m (gOutput c, Generator generatorOutputDevice)

data VectorSpec (n :: Nat) (a :: Type) where

Constructors

VectorSpec :: forall n a. SNat n -> Vector n (ModelSpec a) -> VectorSpec n a

type StateDictKey = Text

type StateDict = Map StateDictKey (ForeignPtr Tensor)

newtype FromStateDictError

Constructors

FromStateDictKeyNotFoundError

Fields

fsdeExpectedKey :: StateDictKey

newtype ToStateDictError

Constructors

ToStateDictKeyAlreadyInUseError

Fields

fsdeTakenKey :: StateDictKey

class HasStateDict model where

Minimal complete definition

Nothing

Methods

fromStateDict :: forall m. (MonadIO m, MonadThrow m, MonadState StateDict m) => ModelSpec model -> StateDictKey -> m model

default fromStateDict :: forall m. (MonadIO m, MonadThrow m, MonadState StateDict m, Generic model, Generic (ModelSpec model), GHasStateDict (Rep model) (Rep (ModelSpec model))) => ModelSpec model -> StateDictKey -> m model

toStateDict :: forall m. (MonadThrow m, MonadState StateDict m) => StateDictKey -> model -> m ()

default toStateDict :: forall m. (MonadThrow m, MonadState StateDict m, Generic model, GHasStateDict (Rep model) (Rep (ModelSpec model))) => StateDictKey -> model -> m ()

class GHasStateDict gModel gModelSpec | gModelSpec -> gModel, gModel -> gModelSpec where

Methods

gFromStateDict :: forall m c. (MonadIO m, MonadThrow m, MonadState StateDict m) => gModelSpec c -> StateDictKey -> m (gModel c)

gToStateDict :: forall m c. (MonadThrow m, MonadState StateDict m) => StateDictKey -> gModel c -> m ()

stateDictFromFile :: FilePath -> IO StateDict

Load a state dictionary from a TorchScript file.

stateDictToFile :: StateDict -> FilePath -> IO ()

Save a state dictionary to a TorchScript file.