Torch.GraduallyTyped.NN.Dropout

Documentation

newtype Dropout (p :: Type) where

Given a random generator, randomly zeroes some of the elements of the input tensor with probability p using samples from a Bernoulli distribution. Each channel will be zeroed out independently on every forward call.

Constructors

Dropout

Fields

:: forall p. p

probability of an element to be zeroed

-> Dropout p