Torch.GraduallyTyped.NN.Transformer.GEncoderDecoder

Documentation

data EncoderDecoderTransformerHasEmbedScaling

Data type that is used to represent whether the encoder-decoder transformer model has a scaled embedding.

Constructors

EncoderDecoderTransformerWithEmbedScaling

EncoderDecoderTransformerWithoutEmbedScaling

data GEncoderDecoderTransformer (inputEmbedDim :: Dim (Name Symbol) (Size Nat)) (encoder :: Type) (decoder :: Type) (sharedEmbedding :: Type) (head :: Type) where

Generic encoder-decoder transformer model. This is a model that can be used to encode and decode sequences of variable length.

inputEmbedDim: the dimension of the input embedding.

encoder: a transformer encoder.

decoder: a transformer decoder.

sharedEmbedding: a shared embedding layer.

head: a head layer for the output.

Constructors

GEncoderDecoderTransformer

Fields

:: forall inputEmbedDim encoder decoder sharedEmbedding head. { edtInputEmbedDim :: SDim inputEmbedDim

input embedding dim for scaling

, edtEncoder :: encoder

encoder

, edtDecoder :: decoder

decoder

, edtSharedEmbedding :: sharedEmbedding

embedding shared between encoder and decoder

, edtHead :: head

transformer head

, edtEmbedScaling :: EncoderDecoderTransformerHasEmbedScaling

embedding scaling

} -> GEncoderDecoderTransformer inputEmbedDim encoder decoder sharedEmbedding head

type family GEncoderDecoderTransformerF (style :: TransformerStyle) (transformerHead :: TransformerHead) (numEncoderLayers :: Nat) (numDecoderLayers :: Nat) (gradient :: Gradient RequiresGradient) (device :: Device (DeviceType Nat)) (dataType :: DataType DType) (headDim :: Dim (Name Symbol) (Size Nat)) (headEmbedDim :: Dim (Name Symbol) (Size Nat)) (embedDim :: Dim (Name Symbol) (Size Nat)) (inputEmbedDim :: Dim (Name Symbol) (Size Nat)) (ffnDim :: Dim (Name Symbol) (Size Nat)) (posEncDim :: Dim (Name Symbol) (Size Nat)) (vocabDim :: Dim (Name Symbol) (Size Nat)) (hasDropout :: HasDropout) :: Type where ...

Equations

GEncoderDecoderTransformerF style transformerHead numEncoderLayers numDecoderLayers gradient device dataType headDim headEmbedDim embedDim inputEmbedDim ffnDim posEncDim vocabDim hasDropout = GEncoderDecoderTransformer inputEmbedDim (EDTEncoderF style numEncoderLayers gradient device dataType headDim headEmbedDim embedDim inputEmbedDim ffnDim posEncDim hasDropout) (EDTDecoderF style numDecoderLayers gradient device dataType headDim headEmbedDim embedDim inputEmbedDim ffnDim posEncDim hasDropout) (EDTSharedEmbeddingF style gradient device dataType inputEmbedDim vocabDim) (EDTHeadF style transformerHead gradient device dataType inputEmbedDim vocabDim)

type family EDTEncoderF (style :: TransformerStyle) (numEncoderLayers :: Nat) (gradient :: Gradient RequiresGradient) (device :: Device (DeviceType Nat)) (dataType :: DataType DType) (headDim :: Dim (Name Symbol) (Size Nat)) (headEmbedDim :: Dim (Name Symbol) (Size Nat)) (embedDim :: Dim (Name Symbol) (Size Nat)) (inputEmbedDim :: Dim (Name Symbol) (Size Nat)) (ffnDim :: Dim (Name Symbol) (Size Nat)) (posEncDim :: Dim (Name Symbol) (Size Nat)) (hasDropout :: HasDropout) :: Type where ...

Specifies the encoder of the encoder-decoder transformer model.

Equations

EDTEncoderF style numEncoderLayers gradient device dataType headDim headEmbedDim embedDim inputEmbedDim ffnDim posEncDim hasDropout = NamedModel (TransformerEncoderF style numEncoderLayers gradient device dataType headDim headEmbedDim embedDim inputEmbedDim ffnDim posEncDim hasDropout)

type family EDTDecoderF (style :: TransformerStyle) (numDecoderLayers :: Nat) (gradient :: Gradient RequiresGradient) (device :: Device (DeviceType Nat)) (dataType :: DataType DType) (headDim :: Dim (Name Symbol) (Size Nat)) (headEmbedDim :: Dim (Name Symbol) (Size Nat)) (embedDim :: Dim (Name Symbol) (Size Nat)) (inputEmbedDim :: Dim (Name Symbol) (Size Nat)) (ffnDim :: Dim (Name Symbol) (Size Nat)) (posEncDim :: Dim (Name Symbol) (Size Nat)) (hasDropout :: HasDropout) :: Type where ...

Specifies the decoder of the encoder-decoder transformer model.

Equations

EDTDecoderF style numDecoderLayers gradient device dataType headDim headEmbedDim embedDim inputEmbedDim ffnDim posEncDim hasDropout = NamedModel (TransformerDecoderF style numDecoderLayers gradient device dataType headDim headEmbedDim embedDim inputEmbedDim inputEmbedDim ffnDim posEncDim hasDropout)

type family EDTSharedEmbeddingF (style :: TransformerStyle) (gradient :: Gradient RequiresGradient) (device :: Device (DeviceType Nat)) (dataType :: DataType DType) (inputEmbedDim :: Dim (Name Symbol) (Size Nat)) (vocabDim :: Dim (Name Symbol) (Size Nat)) where ...

Specifies the shared embedding layer of the encoder-decoder transformer model.

Equations

EDTSharedEmbeddingF _ gradient device dataType inputEmbedDim vocabDim = NamedModel (Embedding gradient ('Layout 'Dense) device dataType vocabDim inputEmbedDim 'Nothing)

type family EDTHeadF (style :: TransformerStyle) (transformerHead :: TransformerHead) (gradient :: Gradient RequiresGradient) (device :: Device (DeviceType Nat)) (dataType :: DataType DType) (inputEmbedDim :: Dim (Name Symbol) (Size Nat)) (vocabDim :: Dim (Name Symbol) (Size Nat)) :: Type where ...

Specifies the head of the encoder-decoder transformer model.

Equations

EDTHeadF style 'WithoutHead gradient device dataType inputEmbedDim vocabDim = ()

EDTHeadF style 'WithLMHead gradient device dataType inputEmbedDim vocabDim = NamedModel (GLMHeadF style gradient device dataType inputEmbedDim vocabDim)

encoderDecoderTransformerSpec

Arguments

:: forall style transformerHead numEncoderLayers numDecoderLayers gradient device dataType headDim headEmbedDim embedDim inputEmbedDim ffnDim posEncDim vocabDim hasDropout. STransformerStyle style
the style of the encoder-decoder transformer model, e.g. ST5, SBART, etc.

-> STransformerHead transformerHead
the head of the encoder-decoder transformer model.

-> SNat numEncoderLayers
the number of encoder layers of the encoder-decoder transformer model.

-> SNat numDecoderLayers
the number of decoder layers of the encoder-decoder transformer model.

-> SGradient gradient
whether or not to compute the gradient of the model parameters

-> SDevice device
the computational device on which the model is allocated.

-> SDataType dataType
the data type of the model parameters.

-> SDim headDim
the dimension of all transformer heads in the encoder-decoder transformer model.

-> SDim headEmbedDim
the dimension of the transformer head embeddings.

-> SDim embedDim
the dimension of the transformer embeddings.

-> SDim inputEmbedDim
the dimension of the input embeddings for both the encoder and the decoder.

-> SDim ffnDim
the dimension of the feed-forward network.

-> SDim posEncDim
the dimension of the positional embeddings.

-> SDim vocabDim
the dimension of the vocabulary.

-> SHasDropout hasDropout
whether or not to use dropout.

-> Double
the dropout rate.

-> Double
the epsilon value for numerical stability of the layer normalization.

-> ModelSpec (GEncoderDecoderTransformerF style transformerHead numEncoderLayers numDecoderLayers gradient device dataType headDim headEmbedDim embedDim inputEmbedDim ffnDim posEncDim vocabDim hasDropout)
the parameter specification of an encoder-decoder transformer model.

Specifies the parameters of an encoder-decoder transformer model.

data GSimplifiedEncoderDecoderTransformer (model :: Type) (mkPos :: Type) (mkDecoderPos :: Type) (mkPaddingMask :: Type) (mkAttentionMask :: Type) (mkCrossAttentionMask :: Type) (mkDecoderAttentionMask :: Type) where

Constructors

GSimplifiedEncoderDecoderTransformer

Fields

:: forall model mkPos mkDecoderPos mkPaddingMask mkAttentionMask mkCrossAttentionMask mkDecoderAttentionMask. { sedtModel :: model

encoder-decoder model

, sedtDecoderInputShift :: ShiftRight Int

shift for decoder input

, sedtPaddingMaskShift :: ShiftRight Int

shift for padding mask

, sedtMkPos :: mkPos

make encoder input positions

, sedtMkDecoderPos :: mkDecoderPos

make decoder input position

, sedtMkPaddingMask :: mkPaddingMask

make padding mask

, sedtMkAttentionMask :: mkAttentionMask

make attention mask

, sedtMkCrossAttentionMask :: mkCrossAttentionMask

make cross-attention mask

, sedtMkDecoderAttentionMask :: mkDecoderAttentionMask

make decoder attention mask

} -> GSimplifiedEncoderDecoderTransformer model mkPos mkDecoderPos mkPaddingMask mkAttentionMask mkCrossAttentionMask mkDecoderAttentionMask

data EncoderDecoderTransformerInput input decoderInput pos decoderPos attentionMask decoderAttentionMask crossAttentionMask where

Input data type for use with an encoder-decoder transformer. Use this for training.

Constructors

EncoderDecoderTransformerInput

Fields

:: forall input decoderInput pos decoderPos attentionMask decoderAttentionMask crossAttentionMask. { edtInput :: input

, edtDecoderInput :: decoderInput

, edtPos :: pos

, edtDecoderPos :: decoderPos

, edtAttentionMask :: attentionMask

, edtDecoderAttentionMask :: decoderAttentionMask

, edtCrossAttentionMask :: crossAttentionMask

} -> EncoderDecoderTransformerInput input decoderInput pos decoderPos attentionMask decoderAttentionMask crossAttentionMask

data EncoderDecoderTransformerInput' input pos attentionMask where

Constructors

EncoderDecoderTransformerInput'

Fields

:: forall input pos attentionMask. { edtInput' :: input

, edtPos' :: pos

, edtAttentionMask' :: attentionMask

} -> EncoderDecoderTransformerInput' input pos attentionMask

data SimplifiedEncoderDecoderTransformerInput input decoderInput where

Constructors

SimplifiedEncoderDecoderTransformerInput

Fields

:: forall input decoderInput. { sedtInput :: input

, sedtDecoderInput :: decoderInput

} -> SimplifiedEncoderDecoderTransformerInput input decoderInput

data SimplifiedEncoderDecoderTransformerInput' input where

Constructors

SimplifiedEncoderDecoderTransformerInput'

Fields

:: forall input. { sedtInput' :: input

} -> SimplifiedEncoderDecoderTransformerInput' input

data SimplifiedEncoderDecoderTransformerTrainingInput input target where

Constructors

SimplifiedEncoderDecoderTransformerTrainingInput

Fields

:: forall input target. { sedtTrainingInput :: input

, sedtTarget :: target

} -> SimplifiedEncoderDecoderTransformerTrainingInput input target

data EncoderDecoderTransformerOutput decoderOutput encoderOutput where

Output data type for use with an encoder-decoder transformer.

Constructors

EncoderDecoderTransformerOutput

Fields

:: forall decoderOutput encoderOutput. { edtDecoderOutput :: decoderOutput

, edtEncoderOutput :: encoderOutput

} -> EncoderDecoderTransformerOutput decoderOutput encoderOutput

data EncoderDecoderTransformerOutput' encoderOutput where

Constructors

EncoderDecoderTransformerOutput'

Fields

:: forall encoderOutput. { edtEncoderOutput' :: encoderOutput

} -> EncoderDecoderTransformerOutput' encoderOutput

data SimplifiedEncoderDecoderTransformerOutput decoderOutput encoderOutput decoderInput inputPaddingMask where

Constructors

SimplifiedEncoderDecoderTransformerOutput

Fields

:: forall decoderOutput encoderOutput decoderInput inputPaddingMask. { sedtDecoderOutput :: decoderOutput

, sedtEncoderOutput :: encoderOutput

, sedtOriginalDecoderInput :: decoderInput

, sedtInputPaddingMask :: inputPaddingMask

} -> SimplifiedEncoderDecoderTransformerOutput decoderOutput encoderOutput decoderInput inputPaddingMask

data SimplifiedEncoderDecoderTransformerOutput' encoderOutput inputPaddingMask where

Constructors

SimplifiedEncoderDecoderTransformerOutput'

Fields

:: forall encoderOutput inputPaddingMask. { sedtEncoderOutput' :: encoderOutput

, sedtInputPaddingMask' :: inputPaddingMask

} -> SimplifiedEncoderDecoderTransformerOutput' encoderOutput inputPaddingMask

data SimplifiedEncoderDecoderTransformerTrainingOutput loss where

Constructors

SimplifiedEncoderDecoderTransformerTrainingOutput

Fields

:: forall loss. { sedtLoss :: loss

} -> SimplifiedEncoderDecoderTransformerTrainingOutput loss

data EncoderDecoderTransformerGenerationInput decoderInput encoderOutput decoderPos decoderAttentionMask crossAttentionMask where

Input data type for use with an encoder-decoder transformer. Use this for inference.

Constructors

EncoderDecoderTransformerGenerationInput

Fields

:: forall decoderInput encoderOutput decoderPos decoderAttentionMask crossAttentionMask. { edtGenerationDecoderInput :: decoderInput

, edtGenerationEncoderOutput :: encoderOutput

, edtGenerationDecoderPos :: decoderPos

, edtGenerationDecoderAttentionMask :: decoderAttentionMask

, edtGenerationCrossAttentionMask :: crossAttentionMask

} -> EncoderDecoderTransformerGenerationInput decoderInput encoderOutput decoderPos decoderAttentionMask crossAttentionMask

data SimplifiedEncoderDecoderTransformerGenerationInput decoderInput encoderOutput inputPaddingMask where

Constructors

SimplifiedEncoderDecoderTransformerGenerationInput

Fields

:: forall decoderInput encoderOutput inputPaddingMask. { sedtGenerationDecoderInput :: decoderInput

, sedtGenerationEncoderOutput :: encoderOutput

, sedtGenerationInputPaddingMask :: inputPaddingMask

} -> SimplifiedEncoderDecoderTransformerGenerationInput decoderInput encoderOutput inputPaddingMask