Torch.GraduallyTyped.NN.Transformer.GEncoderOnly

Documentation

data EncoderOnlyTransformerHasEmbedScaling

Data type that is used to represent whether the encoder-only transformer model has a scaled embedding.

Constructors

EncoderOnlyTransformerWithEmbedScaling

EncoderOnlyTransformerWithoutEmbedScaling

data GEncoderOnlyTransformer (inputEmbedDim :: Dim (Name Symbol) (Size Nat)) (encoder :: Type) (encoderEmbedding :: Type) (encoderTypeEmbedding :: Type) (head :: Type) where

Generic encoder-only transformer model. This is a transformer model that only encodes the input, e.g. BERT.

inputEmbedDim: the dimension of the input embedding.

encoder: a transformer encoder.

encoderEmbedding: an embedding layer for the input.

encoderTypeEmbedding: an embedding layer for the type of the input.

head: a head layer for the output.

Constructors

GEncoderOnlyTransformer

Fields

:: forall inputEmbedDim encoder encoderEmbedding encoderTypeEmbedding head. { eotInputEmbedDim :: SDim inputEmbedDim

input embedding dim for scaling

, eotEncoder :: encoder

encoder

, eotEmbedding :: encoderEmbedding

encoder embedding

, eotTypeEmbedding :: encoderTypeEmbedding

encoder type embedding

, eotHead :: head

encoder head

, eotEmbedScaling :: EncoderOnlyTransformerHasEmbedScaling

encoder embedding scaling

} -> GEncoderOnlyTransformer inputEmbedDim encoder encoderEmbedding encoderTypeEmbedding head

type family GEncoderOnlyTransformerF (style :: TransformerStyle) (transformerHead :: TransformerHead) (numLayers :: Nat) (gradient :: Gradient RequiresGradient) (device :: Device (DeviceType Nat)) (dataType :: DataType DType) (headDim :: Dim (Name Symbol) (Size Nat)) (headEmbedDim :: Dim (Name Symbol) (Size Nat)) (embedDim :: Dim (Name Symbol) (Size Nat)) (inputEmbedDim :: Dim (Name Symbol) (Size Nat)) (ffnDim :: Dim (Name Symbol) (Size Nat)) (posEncDim :: Dim (Name Symbol) (Size Nat)) (vocabDim :: Dim (Name Symbol) (Size Nat)) (typeVocabDim :: Dim (Name Symbol) (Size Nat)) (hasDropout :: HasDropout) :: Type where ...

Equations

GEncoderOnlyTransformerF style transformerHead numLayers gradient device dataType headDim headEmbedDim embedDim inputEmbedDim ffnDim posEncDim vocabDim typeVocabDim hasDropout = GEncoderOnlyTransformer inputEmbedDim (NamedModel (TransformerEncoderF style numLayers gradient device dataType headDim headEmbedDim embedDim inputEmbedDim ffnDim posEncDim hasDropout)) (EOTEmbeddingF style gradient device dataType inputEmbedDim vocabDim) (EOTTypeEmbeddingF style gradient device dataType inputEmbedDim typeVocabDim) (EOTHeadF style transformerHead gradient device dataType inputEmbedDim vocabDim)

type family EOTEmbeddingF (style :: TransformerStyle) (gradient :: Gradient RequiresGradient) (device :: Device (DeviceType Nat)) (dataType :: DataType DType) (inputEmbedDim :: Dim (Name Symbol) (Size Nat)) (vocabDim :: Dim (Name Symbol) (Size Nat)) :: Type where ...

Specifies the embedding layer of the encoder-only transformer model.

Equations

EOTEmbeddingF _ gradient device dataType inputEmbedDim vocabDim = NamedModel (Embedding gradient ('Layout 'Dense) device dataType vocabDim inputEmbedDim 'Nothing)

type family EOTTypeEmbeddingF (style :: TransformerStyle) (gradient :: Gradient RequiresGradient) (device :: Device (DeviceType Nat)) (dataType :: DataType DType) (inputEmbedDim :: Dim (Name Symbol) (Size Nat)) (typeVocabDim :: Dim (Name Symbol) (Size Nat)) :: Type where ...

Specifies the type embedding layer of the encoder-only transformer model.

Equations

EOTTypeEmbeddingF 'BERT gradient device dataType inputEmbedDim typeVocabDim = NamedModel (Embedding gradient ('Layout 'Dense) device dataType typeVocabDim inputEmbedDim 'Nothing)

EOTTypeEmbeddingF 'RoBERTa gradient device dataType inputEmbedDim typeVocabDim = EOTTypeEmbeddingF 'BERT gradient device dataType inputEmbedDim typeVocabDim

type family EOTHeadF (style :: TransformerStyle) (transformerHead :: TransformerHead) (gradient :: Gradient RequiresGradient) (device :: Device (DeviceType Nat)) (dataType :: DataType DType) (inputEmbedDim :: Dim (Name Symbol) (Size Nat)) (vocabDim :: Dim (Name Symbol) (Size Nat)) :: Type where ...

Specifies the head layer of the encoder-only transformer model.

Equations

EOTHeadF _ 'WithoutHead _ _ _ _ _ = ()

EOTHeadF style 'WithLMHead gradient device dataType inputEmbedDim vocabDim = NamedModel (GLMHeadF style gradient device dataType inputEmbedDim vocabDim)

encoderOnlyTransformerSpec :: forall style transformerHead numLayers gradient device dataType headDim headEmbedDim embedDim inputEmbedDim ffnDim posEncDim vocabDim typeVocabDim hasDropout. STransformerStyle style -> STransformerHead transformerHead -> SNat numLayers -> SGradient gradient -> SDevice device -> SDataType dataType -> SDim headDim -> SDim headEmbedDim -> SDim embedDim -> SDim inputEmbedDim -> SDim ffnDim -> SDim posEncDim -> SDim vocabDim -> SDim typeVocabDim -> SHasDropout hasDropout -> Double -> Double -> ModelSpec (GEncoderOnlyTransformerF style transformerHead numLayers gradient device dataType headDim headEmbedDim embedDim inputEmbedDim ffnDim posEncDim vocabDim typeVocabDim hasDropout)

Specifies the parameters of an encoder-only transformer model.

style: the style of the encoder-only transformer model, e.g. SBERT, SRoBERTa, etc.

transformerHead: the head of the encoder-only transformer model.

numLayers: the number of layers of the encoder-only transformer model.

gradient: whether to compute the gradient of the model parameters

device: the computational device on which the model is allocated.

dataType: the data type of the model parameters.

headDim: the dimension of all transformer heads in the encoder-only transformer model.

headEmbedDim: the dimension of the transformer head embeddings.

embedDim: the dimension of the transformer embeddings.

inputEmbedDim: the dimension of the input embeddings.

ffnDim: the dimension of the feed-forward network.

posEncDim: the dimension of the positional embeddings.

vocabDim: the dimension of the vocabulary.

typeVocabDim: the dimension of the type vocabulary.

dropoutP: the dropout rate.

eps: the epsilon value for numerical stability of the layer normalization.

data GSimplifiedEncoderOnlyTransformer (model :: Type) (mkPos :: Type) (mkPaddingMask :: Type) (mkAttentionMask :: Type) where

Constructors

GSimplifiedEncoderOnlyTransformer

Fields

:: forall model mkPos mkPaddingMask mkAttentionMask. { seotModel :: model

encoder-only model

, seotMkPos :: mkPos

make input positions

, seotMkPaddingMask :: mkPaddingMask

make padding mask

, seotMkAttentionMask :: mkAttentionMask

make attention mask

} -> GSimplifiedEncoderOnlyTransformer model mkPos mkPaddingMask mkAttentionMask

data EncoderOnlyTransformerInput input inputType pos attentionMask where

Input data type for use with an encoder-only transformer.

Constructors

EncoderOnlyTransformerInput

Fields

:: forall input inputType pos attentionMask. { eotInput :: input

, eotInputType :: inputType

, eotPos :: pos

, eotAttentionMask :: attentionMask

} -> EncoderOnlyTransformerInput input inputType pos attentionMask

data SimplifiedEncoderOnlyTransformerInput input inputType where

Constructors

SimplifiedEncoderOnlyTransformerInput

Fields

:: forall input inputType. { seotInput :: input

, seotInputType :: inputType

} -> SimplifiedEncoderOnlyTransformerInput input inputType

data EncoderOnlyTransformerOutput output where

Output data type for use with an encoder-only transformer.

Constructors

EncoderOnlyTransformerOutput

Fields

:: forall output. { eotOutput :: output

} -> EncoderOnlyTransformerOutput output

data SimplifiedEncoderOnlyTransformerOutput output paddingMask where

Constructors

SimplifiedEncoderOnlyTransformerOutput

Fields

:: forall output paddingMask. { seotOutput :: output

, sedtPaddingMask :: paddingMask

} -> SimplifiedEncoderOnlyTransformerOutput output paddingMask