Torch.GraduallyTyped.NN.Transformer.GMultiHeadAttention

Documentation

data MultiHeadAttentionHasScaling

Data type for representing whether or not (and, if so, where) scaling is applied in the multi-headed attention layer.

Constructors

MultiHeadAttentionWithoutScaling
Scaling is not done.

MultiHeadAttentionWithQueryScaling
Scaling is applied to the query after in the in-projection.

MultiHeadAttentionWithWeightScaling
Scaling is applied to the attention weights.

data GMultiHeadAttention (headDim :: Dim (Name Symbol) (Size Nat)) (headEmbedDim :: Dim (Name Symbol) (Size Nat)) (embedDim :: Dim (Name Symbol) (Size Nat)) (qInProj :: Type) (kInProj :: Type) (vInProj :: Type) (outProj :: Type) (dropout :: Type) where

Generic multi-headed attention layer.

headDim is the dimension of the attention heads.

headEmbedDim is the dimension of the attention head embedding.

embedDim is the dimension of the embedding.

qInProj is the type of the query projection.

kInProj is the type of the key projection.

vInProj is the type of the value projection.

outProj is the type of the output projection.

dropout is the type of the dropout layer.

Constructors

GMultiHeadAttention

Fields

:: forall headDim headEmbedDim embedDim qInProj kInProj vInProj outProj dropout. { mhaHeadDim :: SDim headDim

head dim

, mhaHeadEmbedDim :: SDim headEmbedDim

head embed dim

, mhaEmbedDim :: SDim embedDim

embed dim

, mhaQInProj :: qInProj

in-projection for query

, mhaKInProj :: kInProj

in-projection for key

, mhaVInProj :: vInProj

in-projection for value

, mhaOutProj :: outProj

out-projection

, mhaDropout :: dropout

dropout

, mhaScaling :: MultiHeadAttentionHasScaling

scaling

} -> GMultiHeadAttention headDim headEmbedDim embedDim qInProj kInProj vInProj outProj dropout

type family GMultiHeadAttentionF (style :: TransformerStyle) (gradient :: Gradient RequiresGradient) (device :: Device (DeviceType Nat)) (dataType :: DataType DType) (headDim :: Dim (Name Symbol) (Size Nat)) (headEmbedDim :: Dim (Name Symbol) (Size Nat)) (embedDim :: Dim (Name Symbol) (Size Nat)) (queryEmbedDim :: Dim (Name Symbol) (Size Nat)) (keyEmbedDim :: Dim (Name Symbol) (Size Nat)) (valueEmbedDim :: Dim (Name Symbol) (Size Nat)) (hasDropout :: HasDropout) :: Type where ...

Equations

GMultiHeadAttentionF style gradient device dataType headDim headEmbedDim embedDim queryEmbedDim keyEmbedDim valueEmbedDim hasDropout = GMultiHeadAttention headDim headEmbedDim embedDim (QInProjF style gradient device dataType queryEmbedDim embedDim) (KInProjF style gradient device dataType keyEmbedDim embedDim) (VInProjF style gradient device dataType valueEmbedDim embedDim) (OutProjF style gradient device dataType embedDim queryEmbedDim) (DropoutF style hasDropout)

type family QInProjF (style :: TransformerStyle) (gradient :: Gradient RequiresGradient) (device :: Device (DeviceType Nat)) (dataType :: DataType DType) (queryEmbedDim :: Dim (Name Symbol) (Size Nat)) (embedDim :: Dim (Name Symbol) (Size Nat)) :: Type where ...

Specifies the linear transformation of the query.

Equations

QInProjF 'T5 gradient device dataType queryEmbedDim embedDim = NamedModel (GLinearF 'WithoutBias gradient device dataType queryEmbedDim embedDim)

QInProjF 'ByT5 gradient device dataType queryEmbedDim embedDim = QInProjF 'T5 gradient device dataType queryEmbedDim embedDim

QInProjF _ gradient device dataType queryEmbedDim embedDim = NamedModel (GLinearF 'WithBias gradient device dataType queryEmbedDim embedDim)

type family KInProjF (style :: TransformerStyle) (gradient :: Gradient RequiresGradient) (device :: Device (DeviceType Nat)) (dataType :: DataType DType) (keyEmbedDim :: Dim (Name Symbol) (Size Nat)) (embedDim :: Dim (Name Symbol) (Size Nat)) :: Type where ...

Specifies the linear transformation of the key.

Equations

KInProjF 'T5 gradient device dataType keyEmbedDim embedDim = NamedModel (GLinearF 'WithoutBias gradient device dataType keyEmbedDim embedDim)

KInProjF 'ByT5 gradient device dataType keyEmbedDim embedDim = KInProjF 'T5 gradient device dataType keyEmbedDim embedDim

KInProjF _ gradient device dataType keyEmbedDim embedDim = NamedModel (GLinearF 'WithBias gradient device dataType keyEmbedDim embedDim)

type family VInProjF (style :: TransformerStyle) (gradient :: Gradient RequiresGradient) (device :: Device (DeviceType Nat)) (dataType :: DataType DType) (valueEmbedDim :: Dim (Name Symbol) (Size Nat)) (embedDim :: Dim (Name Symbol) (Size Nat)) :: Type where ...

Specifies the linear transformation of the value.

Equations

VInProjF 'T5 gradient device dataType valueEmbedDim embedDim = NamedModel (GLinearF 'WithoutBias gradient device dataType valueEmbedDim embedDim)

VInProjF 'ByT5 gradient device dataType valueEmbedDim embedDim = VInProjF 'T5 gradient device dataType valueEmbedDim embedDim

VInProjF _ gradient device dataType valueEmbedDim embedDim = NamedModel (GLinearF 'WithBias gradient device dataType valueEmbedDim embedDim)

type family OutProjF (style :: TransformerStyle) (gradient :: Gradient RequiresGradient) (device :: Device (DeviceType Nat)) (dataType :: DataType DType) (embedDim :: Dim (Name Symbol) (Size Nat)) (queryEmbedDim :: Dim (Name Symbol) (Size Nat)) :: Type where ...

Specifies the type of the out-projection layer.

Equations

OutProjF 'T5 gradient device dataType embedDim queryEmbedDim = NamedModel (GLinearF 'WithoutBias gradient device dataType embedDim queryEmbedDim)

OutProjF 'ByT5 gradient device dataType embedDim queryEmbedDim = OutProjF 'T5 gradient device dataType embedDim queryEmbedDim

OutProjF _ gradient device dataType embedDim queryEmbedDim = NamedModel (GLinearF 'WithBias gradient device dataType embedDim queryEmbedDim)

type family DropoutF (style :: TransformerStyle) (hasDropout :: HasDropout) :: Type where ...

Specifies the type of the dropout layer.

Equations

DropoutF _ 'WithDropout = Dropout

DropoutF _ 'WithoutDropout = ()

multiHeadAttentionSpec :: forall style gradient device dataType headDim headEmbedDim embedDim queryEmbedDim keyEmbedDim valueEmbedDim hasDropout. STransformerStyle style -> SGradient gradient -> SDevice device -> SDataType dataType -> SDim headDim -> SDim headEmbedDim -> SDim embedDim -> SDim queryEmbedDim -> SDim keyEmbedDim -> SDim valueEmbedDim -> SHasDropout hasDropout -> Double -> ModelSpec (GMultiHeadAttentionF style gradient device dataType headDim headEmbedDim embedDim queryEmbedDim keyEmbedDim valueEmbedDim hasDropout)

Specifies the parameters of a multi-headed attention layer.

style: the style of the attention layer, e.g. ST5, ByT5, etc.

gradient: whether to compute the gradient of the attention layer.

device: the computational device on which to allocate the attention layer.

dataType: the data type of the attention layer.

headDim: the dimension of the attention heads.

headEmbedDim: the dimension of the attention head embeddings.

embedDim: the dimension of the input embeddings.

queryEmbedDim: the dimension of the query embeddings.

keyEmbedDim: the dimension of the key embeddings.

valueEmbedDim: the dimension of the value embeddings.

dropoutP: the dropout rate.

type BatchDim queryShape keyShape valueShape = (queryShape ! 0) <+> ((keyShape ! 0) <+> (valueShape ! 0))

getBatchDim :: forall m queryShape keyShape valueShape batchDim. (MonadThrow m, batchDim ~ BatchDim queryShape keyShape valueShape) => SShape queryShape -> SShape keyShape -> SShape valueShape -> m (SDim batchDim)

type QuerySeqDim queryShape = queryShape ! 1

getQuerySeqDim :: forall m queryShape querySeqDim. (MonadThrow m, querySeqDim ~ QuerySeqDim queryShape) => SShape queryShape -> m (SDim querySeqDim)

type KeySeqDim keyShape valueShape = (keyShape ! 1) <+> (valueShape ! 1)

getKeySeqDim :: forall m keyShape valueShape keySeqDim. (MonadThrow m, keySeqDim ~ KeySeqDim keyShape valueShape) => SShape keyShape -> SShape valueShape -> m (SDim keySeqDim)