Torch.GraduallyTyped.NN.Transformer.GStack

Documentation

newtype GTransformerStack (stack :: Type) where

Generic transformer stack.

stack is a stack of tranformer blocks.

Constructors

GTransformerStack :: forall stack. stack -> GTransformerStack stack

type family EncoderStackF (style :: TransformerStyle) (numLayers :: Nat) (gradient :: Gradient RequiresGradient) (device :: Device (DeviceType Nat)) (dataType :: DataType DType) (headDim :: Dim (Name Symbol) (Size Nat)) (headEmbedDim :: Dim (Name Symbol) (Size Nat)) (embedDim :: Dim (Name Symbol) (Size Nat)) (queryEmbedDim :: Dim (Name Symbol) (Size Nat)) (ffnDim :: Dim (Name Symbol) (Size Nat)) (hasDropout :: HasDropout) where ...

Equations

EncoderStackF style numLayers gradient device dataType headDim headEmbedDim embedDim queryEmbedDim ffnDim hasDropout = GTransformerStack (Vector numLayers (EncoderBlockF style gradient device dataType headDim headEmbedDim embedDim queryEmbedDim ffnDim hasDropout))

encoderStackSpec :: forall style numLayers gradient device dataType headDim headEmbedDim embedDim queryEmbedDim ffnDim hasDropout. STransformerStyle style -> SNat numLayers -> SGradient gradient -> SDevice device -> SDataType dataType -> SDim headDim -> SDim headEmbedDim -> SDim embedDim -> SDim queryEmbedDim -> SDim ffnDim -> SHasDropout hasDropout -> Double -> Double -> ModelSpec (EncoderStackF style numLayers gradient device dataType headDim headEmbedDim embedDim queryEmbedDim ffnDim hasDropout)

Specifies the parameters of a transformer stack in an encoder configuration.

style: the style of the transformer stack, e.g. ST5, SByT5, etc.

gradient: whether to compute the gradient of the stack's parameters.

device: the computational device on which the stack is allocated.

dataType: the data type of the stack's parameters.

headDim: the dimension of all transformer heads in the stack.

headEmbedDim: the dimension of the transformer head embeddings.

embedDim: the dimension of the transformer embeddings.

queryEmbedDim: the dimension of the transformer query embeddings.

ffnDim: the dimension of the feed-forward network.

dropoutP: the dropout rate.

eps: the epsilon value for numerical stability of the layer normalization.

type family DecoderStackF (style :: TransformerStyle) (numLayers :: Nat) (gradient :: Gradient RequiresGradient) (device :: Device (DeviceType Nat)) (dataType :: DataType DType) (headDim :: Dim (Name Symbol) (Size Nat)) (headEmbedDim :: Dim (Name Symbol) (Size Nat)) (embedDim :: Dim (Name Symbol) (Size Nat)) (queryEmbedDim :: Dim (Name Symbol) (Size Nat)) (keyEmbedDim :: Dim (Name Symbol) (Size Nat)) (ffnDim :: Dim (Name Symbol) (Size Nat)) (hasDropout :: HasDropout) where ...

Equations

DecoderStackF style numLayers gradient device dataType headDim headEmbedDim embedDim queryEmbedDim keyEmbedDim ffnDim hasDropout = GTransformerStack (Vector numLayers (DecoderBlockF style gradient device dataType headDim headEmbedDim embedDim queryEmbedDim keyEmbedDim ffnDim hasDropout))

decoderStackSpec :: forall style numLayers gradient device dataType headDim headEmbedDim embedDim queryEmbedDim keyEmbedDim ffnDim hasDropout. STransformerStyle style -> SNat numLayers -> SGradient gradient -> SDevice device -> SDataType dataType -> SDim headDim -> SDim headEmbedDim -> SDim embedDim -> SDim queryEmbedDim -> SDim keyEmbedDim -> SDim ffnDim -> SHasDropout hasDropout -> Double -> Double -> ModelSpec (DecoderStackF style numLayers gradient device dataType headDim headEmbedDim embedDim queryEmbedDim keyEmbedDim ffnDim hasDropout)

Specifies the parameters of a transformer stack in a decoder configuration.

style: the style of the transformer stack, e.g. ST5, SByT5, etc.

gradient: whether to compute the gradient of the stack's parameters.

device: the computational device on which the stack is allocated.

dataType: the data type of the stack's parameters.

headDim: the dimension of all transformer heads in the stack.

headEmbedDim: the dimension of the transformer head embeddings.

embedDim: the dimension of the transformer embeddings.

queryEmbedDim: the dimension of the transformer query embeddings.

keyEmbedDim: the dimension of the transformer key embeddings.

ffnDim: the dimension of the feed-forward network.

dropoutP: the dropout rate.

eps: the epsilon value for numerical stability of the layer normalization.