Torch.GraduallyTyped.NN.Transformer.T5.Generation

Documentation

data IsFinished

Constructors

Finished

Unfinished

data Beams a b where

Constructors

Beams

Fields

:: forall a b. { finished :: [Hypothesis 'Finished a b]

, unfinished :: [Hypothesis 'Unfinished a b]

} -> Beams a b

data Hypothesis (isFinished :: IsFinished) a b where

Constructors

InitialHypothesis :: forall a b. Hypothesis 'Unfinished a b

UnfinishedHypothesis

Fields

:: forall a b. { currentToken :: a

, currentScore :: Float

, previousHypothesis :: Hypothesis 'Unfinished a b

} -> Hypothesis 'Unfinished a b

FinishedHypothesis

Fields

:: forall a b. { finalToken :: a

, finalScore :: Float

, penultimateHypothesis :: Hypothesis 'Unfinished a b

, finalValue :: b

} -> Hypothesis 'Finished a b

getTokens :: forall a b. Hypothesis 'Unfinished a b -> [a]

getScore :: forall a b. Hypothesis 'Unfinished a b -> Float

data SomeHypothesis a b

Constructors

forall isFinished. SomeHypothesis

Fields

unSomeHypothesis :: Hypothesis isFinished a b

beamSearch :: forall a b m. Monad m => Int -> Int -> ([Hypothesis 'Unfinished a b] -> m [SomeHypothesis a b]) -> m [Beams a b]

beamSearchStep :: forall a b m. Functor m => ([Hypothesis 'Unfinished a b] -> m [SomeHypothesis a b]) -> Beams a b -> m (Beams a b)

runBeamSearch :: forall model input decoderInput encoderOutput encoderOutputShape encoderOutput' inputPaddingMask decoderOutput generatorDevice. (HasForward model (SimplifiedEncoderDecoderTransformerInput input decoderInput) generatorDevice (SimplifiedEncoderDecoderTransformerOutput decoderOutput encoderOutput decoderInput inputPaddingMask) generatorDevice, encoderOutput ~ Tensor ('Gradient 'WithGradient) ('Layout 'Dense) ('Device 'CPU) T5DataType encoderOutputShape, 'UncheckedShape ~ BroadcastShapesF encoderOutputShape 'UncheckedShape, SGetShape encoderOutputShape, HasForward model (SimplifiedEncoderDecoderTransformerGenerationInput decoderInput encoderOutput' inputPaddingMask) generatorDevice (SimplifiedEncoderDecoderTransformerOutput decoderOutput encoderOutput' decoderInput inputPaddingMask) generatorDevice, encoderOutput' ~ Tensor ('Gradient 'WithGradient) ('Layout 'Dense) ('Device 'CPU) T5DataType 'UncheckedShape, decoderInput ~ Tensor ('Gradient 'WithoutGradient) ('Layout 'Dense) ('Device 'CPU) ('DataType 'Int64) ('Shape '['Dim ('Name "*") 'UncheckedSize, 'Dim ('Name "*") 'UncheckedSize]), decoderOutput ~ Tensor ('Gradient 'WithGradient) ('Layout 'Dense) ('Device 'CPU) T5DataType 'UncheckedShape, generatorDevice ~ 'Device 'CPU) => Int -> Int -> model -> input -> Generator generatorDevice -> IO [Beams Int [Int]]

testBeamSearch :: IO ()

next :: forall t b i a. (i ~ Int, Show i, MonadTrans t, Monad (t (StateT [i] b)), Alternative (t (StateT [i] b)), Monad b, Foldable b) => t (StateT [i] b) i -> Parser (StateT [i] b) i a -> (Parser (StateT [i] b) i a -> t (StateT [i] b) a) -> t (StateT [i] b) a

notNull :: (Monad b, Foldable b) => Parser (StateT [i] b) i a -> StateT [i] b Bool

hasFree :: (Monad b, Foldable b) => Parser (StateT [i] b) i a -> StateT [i] b Bool

transParser :: MonadPlus b => Map Int String -> Parser b Char a -> Parser b Int a

transParser vocab p transforms a parser p over characters Char into a parser over token indices Int using the vocabulary vocab.

getIs :: forall model input generatorDevice b decoderInput encoderOutput decoderOutput inputPaddingMask s. (Alternative b, MonadThrow b, s ~ (Maybe (encoderOutput, inputPaddingMask), Generator generatorDevice), decoderInput ~ Tensor ('Gradient 'WithoutGradient) ('Layout 'Dense) ('Device 'CPU) ('DataType 'Int64) ('Shape '['Dim ('Name "*") ('Size 1), 'Dim ('Name "*") 'UncheckedSize]), decoderOutput ~ Tensor ('Gradient 'WithGradient) ('Layout 'Dense) ('Device 'CPU) T5DataType 'UncheckedShape, HasForward model (SimplifiedEncoderDecoderTransformerInput input decoderInput) generatorDevice (SimplifiedEncoderDecoderTransformerOutput decoderOutput encoderOutput decoderInput inputPaddingMask) generatorDevice, HasForward model (SimplifiedEncoderDecoderTransformerGenerationInput decoderInput encoderOutput inputPaddingMask) generatorDevice (SimplifiedEncoderDecoderTransformerOutput decoderOutput encoderOutput decoderInput inputPaddingMask) generatorDevice) => Int -> model -> input -> StateT s (StateT [Int] b) Int

Get continuations from model

runParser :: forall model input generatorDevice b a. _ => Int -> model -> input -> Generator generatorDevice -> Parser (StateT [Int] b) Int a -> b (a, [Int])

t5Text :: MonadPlus b => Parser b Char String

t5Text parses a Char sequence delimited by /s as a String.

t5Test :: MonadPlus b => Parser b Char String

t5Sql :: (TokenParsing (FreeT ((->) Char) b), MonadPlus b) => Parser b Char SpiderSQL

t5Sql parses a Char sequence starting with " and ending with " /s as SpiderSQL.

Orphan instances

(Alternative b, Foldable b, MonadPlus b) => CharParsing (FreeT ((->) Char) (StateT [Int] b))

Instance details

Methods

satisfy :: (Char -> Bool) -> FreeT ((->) Char) (StateT [Int] b) Char

char :: Char -> FreeT ((->) Char) (StateT [Int] b) Char

notChar :: Char -> FreeT ((->) Char) (StateT [Int] b) Char

anyChar :: FreeT ((->) Char) (StateT [Int] b) Char

string :: String -> FreeT ((->) Char) (StateT [Int] b) String

text :: Text -> FreeT ((->) Char) (StateT [Int] b) Text

(Alternative b, Foldable b, MonadPlus b) => Parsing (FreeT ((->) Char) (StateT [Int] b))

Instance details

Methods

try :: FreeT ((->) Char) (StateT [Int] b) a -> FreeT ((->) Char) (StateT [Int] b) a

(<?>) :: FreeT ((->) Char) (StateT [Int] b) a -> String -> FreeT ((->) Char) (StateT [Int] b) a

skipMany :: FreeT ((->) Char) (StateT [Int] b) a -> FreeT ((->) Char) (StateT [Int] b) ()

skipSome :: FreeT ((->) Char) (StateT [Int] b) a -> FreeT ((->) Char) (StateT [Int] b) ()

unexpected :: String -> FreeT ((->) Char) (StateT [Int] b) a

eof :: FreeT ((->) Char) (StateT [Int] b) ()

notFollowedBy :: Show a => FreeT ((->) Char) (StateT [Int] b) a -> FreeT ((->) Char) (StateT [Int] b) ()

(Alternative b, Foldable b, MonadPlus b) => TokenParsing (FreeT ((->) Char) (StateT [Int] b))

Instance details

Methods

someSpace :: FreeT ((->) Char) (StateT [Int] b) ()

nesting :: FreeT ((->) Char) (StateT [Int] b) a -> FreeT ((->) Char) (StateT [Int] b) a

semi :: FreeT ((->) Char) (StateT [Int] b) Char

highlight :: Highlight -> FreeT ((->) Char) (StateT [Int] b) a -> FreeT ((->) Char) (StateT [Int] b) a

token :: FreeT ((->) Char) (StateT [Int] b) a -> FreeT ((->) Char) (StateT [Int] b) a