Torch.GraduallyTyped.Prelude

Documentation

type Type = Type

The kind of types with lifted values. For example Int :: Type.

data Constraint

The kind of constraints, like Show a

data Proxy (t :: k)

Proxy is a type that holds no data, but has a phantom parameter of arbitrary type (or even kind). Its use is to provide type information, even though there is no value available of that type (or it may be too costly to create one).

Historically, Proxy :: Proxy a is a safer alternative to the undefined :: a idiom.

>>> Proxy :: Proxy (Void, Int -> Int)
Proxy

Proxy can even hold types of higher kinds,

>>> Proxy :: Proxy Either
Proxy

>>> Proxy :: Proxy Functor
Proxy

>>> Proxy :: Proxy complicatedStructure
Proxy

Constructors

Proxy

type family (a :: Bool) || (b :: Bool) :: Bool where ... infixr 2

Type-level "or"

Equations

'False || a = a

'True || a = 'True

a || 'False = a

a || 'True = 'True

a || a = a

type family If (cond :: Bool) (tru :: k) (fls :: k) :: k where ...

Type-level If. If True a b ==> a; If False a b ==> b

Equations

If 'True (tru :: k) (fls :: k) = tru

If 'False (tru :: k) (fls :: k) = fls

data Nat

(Kind) This is the kind of type-level natural numbers.

type family (a :: Nat) + (b :: Nat) :: Nat where ... infixl 6

Addition of type-level naturals.

Since: base-4.7.0.0

type family (a :: Nat) * (b :: Nat) :: Nat where ... infixl 7

Multiplication of type-level naturals.

Since: base-4.7.0.0

type family CmpNat (a :: Nat) (b :: Nat) :: Ordering where ...

Comparison of type-level naturals, as a function.

Since: base-4.7.0.0

type family TypeError (a :: ErrorMessage) :: b where ...

The type-level equivalent of error.

The polymorphic kind of this type allows it to be used in several settings. For instance, it can be used as a constraint, e.g. to provide a better error message for a non-existent instance,

-- in a context
instance TypeError (Text "Cannot Show functions." :$$:
                    Text "Perhaps there is a missing argument?")
      => Show (a -> b) where
    showsPrec = error "unreachable"

It can also be placed on the right-hand side of a type-level function to provide an error for an invalid case,

type family ByteSize x where
   ByteSize Word16   = 2
   ByteSize Word8    = 1
   ByteSize a        = TypeError (Text "The type " :<>: ShowType a :<>:
                                  Text " is not exportable.")

Since: base-4.9.0.0

data ErrorMessage

A description of a custom type error.

Constructors

Text Symbol
Show the text as is.

ShowType t
Pretty print the type. ShowType :: k -> ErrorMessage

ErrorMessage :<>: ErrorMessage infixl 6
Put two pieces of error message next to each other.

ErrorMessage :$$: ErrorMessage infixl 5
Stack two pieces of error message on top of each other.

data SList (a1 :: [a]) where

Constructors

SNil :: forall a. SList ('[] :: [a])

SCons :: forall a (n1 :: a) (n2 :: [a]). Sing n1 -> Sing n2 -> SList (n1 ': n2) infixr 5

module Torch.GraduallyTyped.Prelude.TypeLits

data IsChecked a

Constructors

Checked a

Unchecked a

pattern IsChecked :: a -> IsChecked a

pattern Demoted :: SingKind k => Demote k -> Sing (a :: k)

pattern Demoted' :: (SingKind k, Demote k ~ IsChecked t) => t -> Sing (a :: k)

forgetIsChecked :: IsChecked a -> a

pattern (:|:) :: forall a as. Sing a -> SList as -> SList (a ': as) infixr 8

type family All (c :: k -> Constraint) (xs :: [k]) :: Constraint where ...

Equations

All _ '[] = ()

All c (h ': t) = (c h, All c t)

class KnownElem k x where

Associated Types

type ElemValF k :: Type

Methods

elemVal :: ElemValF k

class KnownList k (xs :: [k]) where

Methods

listVal :: [ElemValF k]

type family Assert err a where ...

Can be used to report stuck type families, see https://kcsongor.github.io/report-stuck-families/. This family is able to check whether its argument a is stuck and report an error err in that case.

Equations

Assert _ T = Any

Assert _ a = a

type family Catch a where ...

Approximates a normal form on the type level. Catch forces its argument a and returns an empty Constraint if and only if the argument does not produce a TypeError.

The first equation will recursively force the kind of the argument until it reaches Type or a TypeError. In the former case, it falls over to the second equation which will produce the empty constraint. In the latter case, it gets stuck with 'Catch (TypeError ...)', and the compiler will report the error message.

Thanks to https://kcsongor.github.io/kcsongor for the suggestion.

Equations

Catch (f a) = (Catch f, Catch a)

Catch _ = ()

type family Seq (a :: k) (b :: k') :: k' where ...

Equations

Seq (f a) b = Seq (Seq f a) b

Seq _ b = b

type family Fst (t :: (k, k')) :: k where ...

Returns the first element of a type-level tuple with the kind (k, k') marked by a prefix quote.

>>> :kind! Fst '(Int, String)
Fst '(Int, String) :: *
= Int
>>> :kind! Fst '(Functor, String)
Fst '(Functor, String) :: (* -> *) -> Constraint
= Functor

Equations

Fst '(x, _) = x

type family Snd (t :: (k, k')) :: k' where ...

Returns the second element of a type-level tuple with the kind (k, k') marked by a prefix quote.

>>> :kind! Snd '(Int, String)
Snd '(Int, String) :: *
= [Char]
>>> :kind! Snd '(Int, Monad)
Snd '(Int, Monad) :: (* -> *) -> Constraint
= Monad

Equations

Snd '(_, y) = y

type family Elem (e :: t) (es :: [t]) :: Bool where ...

Check that a given type is an element of a type-level list:

>>> :kind! Elem String '[]
Elem String '[] :: Bool
= 'False
>>> :kind! Elem String '[Int, String]
Elem String '[Int, String] :: Bool
= 'True
>>> :kind! Elem String '[Int, Bool]
Elem String '[Int, Bool] :: Bool
= 'False

Equations

Elem _ '[] = 'False

Elem x (x ': _) = 'True

Elem x (_ ': xs) = Elem x xs

type family Head (xs :: [a]) :: Maybe a where ...

Equations

Head '[] = 'Nothing

Head (x ': _) = 'Just x

type family Tail (xs :: [a]) :: Maybe [a] where ...

Equations

Tail '[] = 'Nothing

Tail (_ ': xs) = 'Just xs

type family Length (xs :: [a]) :: Nat where ...

Equations

Length '[] = 0

Length (_ ': xs) = 1 + Length xs

type family Contains (f :: k) (a :: k') :: Bool where ...

Test whether or not a given type contains another:

>>> :kind! Contains (Either Int String) Int
Contains (Either Int String) Int :: Bool
= 'True
>>> :kind! Contains (Either Int String) Bool
Contains (Either Int String) Bool :: Bool
= 'False
>>> :kind! Contains (Either Int String) Either
Contains (Either Int String) Either :: Bool
= 'True

Equations

Contains a a = 'True

Contains (f g) a = Contains f a || Contains g a

Contains _ _ = 'False

type family Extract (f :: k) (a :: k') :: [k'] where ...

Extract all occurrences of a given type from another:

>>> :kind! Extract (Either Int String) Int
Extract (Either Int String) Int :: [*]
= '[Int]
>>> :kind! Extract (Either Int String) Bool
Extract (Either Int String) Bool :: [*]
= '[]
>>> :kind! Extract (Either Int String) Either
Extract (Either Int String) Either :: [* -> * -> *]
= '[Either]

Equations

Extract a a = '[a]

Extract (f g) a = Concat (Extract f a) (Extract g a)

Extract _ _ = '[]

type family FromMaybe (d :: k) (x :: Maybe k) :: k where ...

Equations

FromMaybe d 'Nothing = d

FromMaybe _ ('Just v) = v

type family MaybeF (d :: k') (f :: k -> k') (x :: Maybe k) :: k' where ...

Equations

MaybeF d _ 'Nothing = d

MaybeF _ f ('Just v) = f v

type family FstMaybe (t :: Maybe (k, k')) :: Maybe k where ...

Equations

FstMaybe 'Nothing = 'Nothing

FstMaybe ('Just '(x, _)) = 'Just x

type family SndMaybe (t :: Maybe (k, k')) :: Maybe k' where ...

Equations

SndMaybe 'Nothing = 'Nothing

SndMaybe ('Just '(_, y)) = 'Just y

type family PrependMaybe (h :: Maybe a) (t :: Maybe [a]) :: Maybe [a] where ...

Equations

PrependMaybe 'Nothing _ = 'Nothing

PrependMaybe _ 'Nothing = 'Nothing

PrependMaybe ('Just h) ('Just t) = 'Just (h ': t)

type family MapMaybe (f :: k -> k') (a :: Maybe k) :: Maybe k' where ...

Equations

MapMaybe _ 'Nothing = 'Nothing

MapMaybe f ('Just k) = 'Just (f k)

type family BindMaybe (f :: k -> Maybe k') (a :: Maybe k) :: Maybe k' where ...

Equations

BindMaybe _ 'Nothing = 'Nothing

BindMaybe f ('Just k) = f k

type family JoinMaybe (a :: Maybe (Maybe k)) :: Maybe k where ...

Equations

JoinMaybe 'Nothing = 'Nothing

JoinMaybe ('Just 'Nothing) = 'Nothing

JoinMaybe ('Just ('Just k)) = 'Just k

type family LiftM2Maybe (f :: k -> k' -> k'') (a :: Maybe k) (b :: Maybe k') :: Maybe k'' where ...

Equations

LiftM2Maybe _ 'Nothing _ = 'Nothing

LiftM2Maybe _ ('Just _) 'Nothing = 'Nothing

LiftM2Maybe f ('Just a) ('Just b) = 'Just (f a b)

type family LiftTimesMaybe (a :: Maybe Nat) (b :: Maybe Nat) :: Maybe Nat where ...

Equations

LiftTimesMaybe 'Nothing _ = 'Nothing

LiftTimesMaybe ('Just _) 'Nothing = 'Nothing

LiftTimesMaybe ('Just a) ('Just b) = 'Just (a * b)

type family LiftTypeEqMaybe (a :: Maybe k) (b :: Maybe k') :: Constraint where ...

Equations

LiftTypeEqMaybe 'Nothing _ = ()

LiftTypeEqMaybe ('Just _) 'Nothing = ()

LiftTypeEqMaybe ('Just a) ('Just b) = a ~ b

type family Concat (xs :: [k]) (ys :: [k]) :: [k] where ...

Equations

Concat '[] ys = ys

Concat (x ': xs) ys = x ': Concat xs ys

type Reverse xs = ReverseImplF xs '[]

type family (a :: Nat) <? (b :: Nat) where ...

Equations

x <? y = (x `CmpNat` y) == 'LT

type family When (cond :: Bool) (constraint :: Constraint) :: Constraint where ...

Equations

When 'True constraint = constraint

When 'False _ = ()

whenM :: Monad m => m Bool -> m () -> m ()

unlessM :: Monad m => m Bool -> m () -> m ()

ifM :: Monad m => m Bool -> m a -> m a -> m a

guardM :: MonadPlus m => m Bool -> m ()

(&&^) :: Monad m => m Bool -> m Bool -> m Bool infixr 3

The && operator lifted to a monad. If the first argument evaluates to False the second argument will not be evaluated.

(||^) :: Monad m => m Bool -> m Bool -> m Bool infixr 2

The || operator lifted to a monad. If the first argument evaluates to True the second argument will not be evaluated.

(<&&>) :: Applicative a => a Bool -> a Bool -> a Bool infixr 3

&& lifted to an Applicative. Unlike &&^ the operator is not short-circuiting.

(<||>) :: Applicative a => a Bool -> a Bool -> a Bool infixr 2

|| lifted to an Applicative. Unlike ||^ the operator is not short-circuiting.