Torch.GraduallyTyped.Tensor.IndexingSlicingJoining

Documentation

>>> import Data.Singletons.Prelude.List (SList (..))
>>> import Torch.GraduallyTyped

class HasCat (selectDim :: SelectDim (By Symbol Nat)) k (c :: k -> Type) (a :: k) where

Minimal complete definition

sCat

Associated Types

type CatF selectDim a c :: Type

Methods

sCat :: SSelectDim selectDim -> c a -> CatF selectDim a c

Concatenates the given sequence of seq tensors in the given dimension. All tensors must either have the same shape (except in the concatenating dimension) or be empty.

>>> t <- ones @('Gradient 'WithGradient) @('Layout 'Dense) @('Device 'CPU) @('DataType 'Float) @('Shape '[ 'Dim ('Name "batch") ('Size 32), 'Dim ('Name "feature") ('Size 8)])
>>> :type cat @('SelectDim ('ByName "feature")) [t]
cat @('SelectDim ('ByName "feature")) [t]
  :: Tensor
       ('Gradient 'WithGradient)
       ('Layout 'Dense)
       ('Device 'CPU)
       ('DataType 'Float)
       ('Shape
          '[ 'Dim ('Name "batch") ('Size 32),
             'Dim 'UncheckedName 'UncheckedSize])
>>> :type cat @('SelectDim ( 'ByIndex 0)) [t]
cat @('SelectDim ( 'ByIndex 0)) [t]
  :: Tensor
       ('Gradient 'WithGradient)
       ('Layout 'Dense)
       ('Device 'CPU)
       ('DataType 'Float)
       ('Shape
          '[ 'Dim 'UncheckedName 'UncheckedSize,
             'Dim ('Name "feature") ('Size 8)])
>>> :type sCat (SUncheckedSelectDim (ByIndex 0)) [t]
sCat (SUncheckedSelectDim (ByIndex 0)) [t]
  :: Tensor
       ('Gradient 'WithGradient)
       ('Layout 'Dense)
       ('Device 'CPU)
       ('DataType 'Float)
       'UncheckedShape

cat :: SingI selectDim => c a -> CatF selectDim a c

type family CatListImplF (selectDim :: SelectDim (By Symbol Nat)) (tensor :: Type) :: Maybe Type where ...

Equations

CatListImplF 'UncheckedSelectDim (Tensor gradient layout device dataType _) = 'Just (Tensor gradient layout device dataType 'UncheckedShape)

CatListImplF ('SelectDim _) (Tensor gradient layout device dataType 'UncheckedShape) = 'Just (Tensor gradient layout device dataType 'UncheckedShape)

CatListImplF ('SelectDim by) (Tensor gradient layout device dataType ('Shape dims)) = MapMaybe (Tensor gradient layout device dataType) (MapMaybe 'Shape (ReplaceDimImplF by dims ('Dim 'UncheckedName 'UncheckedSize)))

type CheckSpellingMessage = "Check the spelling of named dimensions, and make sure the number of dimensions is correct."

type family CatListCheckF (selectDim :: SelectDim (By Symbol Nat)) (tensor :: Type) (result :: Maybe Type) :: Type where ...

Equations

CatListCheckF selectDim (Tensor _ _ _ _ shape) 'Nothing = TypeError ("Cannot concatenate the dimension" % ("" % ((" " <> selectDim) % ("" % ("for tensors of shape" % ("" % (((" " <> shape) <> ".") % ("" % CheckSpellingMessage))))))))

CatListCheckF _ _ ('Just result) = result

type CatListF selectDim tensor = CatListCheckF selectDim tensor (CatListImplF selectDim tensor)

type family CatHListImplF (selectDim :: SelectDim (By Symbol Nat)) (tensors :: [Type]) (acc :: Maybe (Gradient RequiresGradient, Layout LayoutType, Device (DeviceType Nat), DataType DType, Shape [Dim (Name Symbol) (Size Nat)])) :: Type where ...

Equations

CatHListImplF _ '[] 'Nothing = TypeError (ToErrorMessage "Cannot concatenate an empty list of tensors.")

CatHListImplF _ '[] ('Just '(gradient, layout, device, dataType, shape)) = Tensor gradient layout device dataType shape

CatHListImplF selectDim (Tensor gradient layout device dataType shape ': tensors) 'Nothing = CatHListImplF selectDim tensors ('Just '(gradient, layout, device, dataType, shape))

CatHListImplF selectDim (Tensor gradient layout device dataType shape ': tensors) ('Just '(gradient', layout', device', dataType', shape')) = CatHListImplF selectDim tensors ('Just '(gradient <|> gradient', layout <+> layout', device <+> device', dataType <+> dataType', ReplaceDimF selectDim (shape <+> ReplaceDimF selectDim shape' (GetDimF selectDim shape)) (AddDimF (GetDimF selectDim shape) (GetDimF selectDim shape'))))

CatHListImplF _ (x ': _) _ = TypeError ("Cannot concatenate because" % ("" % (((" '" <> x) <> "'") % ("" % "is not a tensor type."))))

type CatHListF selectDim tensors = CatHListImplF selectDim tensors 'Nothing

type ReshapeNumelMismatchMessage (numel :: Nat) (numel' :: Nat) (shape :: Shape [Dim (Name Symbol) (Size Nat)]) (shape' :: Shape [Dim (Name Symbol) (Size Nat)]) = "Cannot reshape the tensor. The original shape," % ("" % (((" '" <> shape) <> "',") % ("" % ("and the new shape," % ("" % (((" '" <> shape') <> "',") % ("" % ("have different total numbers of elements," % ("" % (((((" '" <> numel) <> "' versus '") <> numel') <> "',") % ("" % "respectively.")))))))))))

type family ReshapeImplF (numel :: Maybe Nat) (numel' :: Maybe Nat) (shape :: Shape [Dim (Name Symbol) (Size Nat)]) (shape' :: Shape [Dim (Name Symbol) (Size Nat)]) :: Shape [Dim (Name Symbol) (Size Nat)] where ...

Equations

ReshapeImplF ('Just numel) ('Just numel) _ shape' = shape'

ReshapeImplF ('Just numel) ('Just numel') shape shape' = TypeError (ReshapeNumelMismatchMessage numel numel' shape shape')

ReshapeImplF 'Nothing _ _ _ = 'UncheckedShape

ReshapeImplF _ 'Nothing _ _ = 'UncheckedShape

ReshapeImplF _ _ 'UncheckedShape _ = 'UncheckedShape

ReshapeImplF _ _ _ 'UncheckedShape = 'UncheckedShape

type family ReshapeF (shape :: Shape [Dim (Name Symbol) (Size Nat)]) (shape' :: Shape [Dim (Name Symbol) (Size Nat)]) :: Shape [Dim (Name Symbol) (Size Nat)] where ...

Equations

ReshapeF shape shape' = ReshapeImplF (NumelF shape) (NumelF shape') shape shape'

sReshape :: forall m shape' gradient layout device dataType shape shape''. MonadThrow m => shape'' ~ ReshapeF shape shape' => SShape shape' -> Tensor gradient layout device dataType shape -> m (Tensor gradient layout device dataType shape'')

Returns a tensor with the same data and number of elements as the input tensor, but with the specified shape:

>>> g <- sMkGenerator (SDevice SCPU) 0
>>> (input, _) <- sRandn (TensorSpec (SGradient SWithGradient) (SLayout SDense) (SDevice SCPU) (SDataType SFloat) (SShape $ SName @"*" :&: SSize @4 :|: SNil)) g
>>> output <- sReshape (SShape $ SName @"*" :&: SSize @2 :|: SName @"*" :&: SSize @2 :|: SNil) input
>>> :type output
output
  :: Tensor
       ('Gradient 'WithGradient)
       ('Layout 'Dense)
       ('Device 'CPU)
       ('DataType 'Float)
       ('Shape '[ 'Dim ('Name "*") ('Size 2), 'Dim ('Name "*") ('Size 2)])

At the value level, a single dimension may be '-1', in which case it is inferred from the remaining dimensions and the number of elements in the input:

>>> output' <- sReshape (SShape $ SUncheckedName "*" :&: SUncheckedSize (-1) :|: SNil) output
>>> :type output'
output'
  :: Tensor
       ('Gradient 'WithGradient)
       ('Layout 'Dense)
       ('Device 'CPU)
       ('DataType 'Float)
       'UncheckedShape
>>> getDims output'
[Dim {dimName = "*", dimSize = 4}]

sSetShape :: forall m shape' gradient layout device dataType shape shape''. MonadThrow m => shape'' ~ ReshapeF shape shape' => SShape shape' -> Tensor gradient layout device dataType shape -> m (Tensor gradient layout device dataType shape'')

Returns a tensor with the same data and number of elements as the input tensor, but with the specified shape:

>>> g <- sMkGenerator (SDevice SCPU) 0
>>> (input, _) <- sRandn (TensorSpec (SGradient SWithGradient) (SLayout SDense) (SDevice SCPU) (SDataType SFloat) (SShape $ SName @"*" :&: SSize @4 :|: SNil)) g
>>> output <- sReshape (SShape $ SName @"*" :&: SSize @2 :|: SName @"*" :&: SSize @2 :|: SNil) input
>>> :type output
output
  :: Tensor
       ('Gradient 'WithGradient)
       ('Layout 'Dense)
       ('Device 'CPU)
       ('DataType 'Float)
       ('Shape '[ 'Dim ('Name "*") ('Size 2), 'Dim ('Name "*") ('Size 2)])

At the value level, a single dimension may be '-1', in which case it is inferred from the remaining dimensions and the number of elements in the input:

>>> output' <- sReshape (SShape $ SUncheckedName "*" :&: SUncheckedSize (-1) :|: SNil) output
>>> :type output'
output'
  :: Tensor
       ('Gradient 'WithGradient)
       ('Layout 'Dense)
       ('Device 'CPU)
       ('DataType 'Float)
       'UncheckedShape
>>> getDims output'
[Dim {dimName = "*", dimSize = 4}]

type family AllDimSizesChecked (shape :: Shape [Dim (Name Symbol) (Size Nat)]) :: Bool where ...

Equations

AllDimSizesChecked 'UncheckedShape = 'False

AllDimSizesChecked ('Shape '[]) = 'True

AllDimSizesChecked ('Shape ('Dim name ('Size size) ': xs)) = AllDimSizesChecked ('Shape xs)

reshape :: forall m shape' gradient layout device dataType shape shape''. (shape'' ~ ReshapeF shape shape', When (AllDimSizesChecked shape) (shape' ~ shape''), SingI shape', MonadThrow m) => Tensor gradient layout device dataType shape -> m (Tensor gradient layout device dataType shape'')

type TransposeBy0Message (by0 :: By Symbol Nat) (dims :: [Dim (Name Symbol) (Size Nat)]) = "Cannot transpose the tensor with the dimensions" % ("" % (((" '" <> dims) <> "'") % ("" % ("because the specified source dimension" % ("" % (((" '" <> by0) <> "'") % ("" % "could not be found.")))))))

type TransposeBy1Message (by1 :: By Symbol Nat) (dims :: [Dim (Name Symbol) (Size Nat)]) = "Cannot transpose the tensor with the dimensions" % ("" % (((" '" <> dims) <> "'") % ("" % ("because the specified target dimension" % ("" % (((" '" <> by1) <> "'") % ("" % "could not be found.")))))))

type family TransposeF (selectDim0 :: SelectDim (By Symbol Nat)) (selectDim1 :: SelectDim (By Symbol Nat)) (shape :: Shape [Dim (Name Symbol) (Size Nat)]) :: Shape [Dim (Name Symbol) (Size Nat)] where ...

Compute transposed shapes.

>>> type SelectBatch = 'SelectDim ('ByName "batch" :: By Symbol Nat)
>>> type SelectFeature = 'SelectDim ('ByName "feature" :: By Symbol Nat)
>>> type Dims = '[ 'Dim ('Name "batch") ('Size 10), 'Dim ('Name "feature") ('Size 8), 'Dim ('Name "anotherFeature") ('Size 12)]
>>> :kind! TransposeF SelectBatch SelectFeature ('Shape Dims)
TransposeF SelectBatch SelectFeature ('Shape Dims) :: Shape
                                                        [Dim (Name Symbol) (Size Nat)]
= 'Shape
    '[ 'Dim ('Name "feature") ('Size 8),
       'Dim ('Name "batch") ('Size 10),
       'Dim ('Name "anotherFeature") ('Size 12)]
>>> :kind! TransposeF SelectFeature SelectBatch ('Shape Dims)
TransposeF SelectFeature SelectBatch ('Shape Dims) :: Shape
                                                        [Dim (Name Symbol) (Size Nat)]
= 'Shape
    '[ 'Dim ('Name "feature") ('Size 8),
       'Dim ('Name "batch") ('Size 10),
       'Dim ('Name "anotherFeature") ('Size 12)]

Equations

TransposeF _ _ 'UncheckedShape = 'UncheckedShape

TransposeF _ 'UncheckedSelectDim _ = 'UncheckedShape

TransposeF 'UncheckedSelectDim _ _ = 'UncheckedShape

TransposeF ('SelectDim ('ByName name0)) ('SelectDim ('ByName name1)) ('Shape dims) = 'Shape (TransposeIndexIndexDimsF (FromMaybe (TypeError (TransposeBy0Message ('ByName name0) dims)) (GetIndexByNameF name0 dims)) (FromMaybe (TypeError (TransposeBy1Message ('ByName name1) dims)) (GetIndexByNameF name1 dims)) dims)

TransposeF ('SelectDim ('ByIndex index0)) ('SelectDim ('ByIndex index1)) ('Shape dims) = 'Shape (TransposeIndexIndexDimsF index0 index1 dims)

TransposeF ('SelectDim by0) ('SelectDim by1) _ = TypeError ("Cannot transpose the tensor. " % ("" % ("The source and target dimensions must be selected either both by name or both by index, " % ("but mixed selectors were found: " % ("" % (((((" '" <> 'SelectDim by0) <> "' and '") <> 'SelectDim by1) <> "'.") % ""))))))

type family TransposeIndexIndexDimsF (index0 :: Nat) (index1 :: Nat) (dims :: [Dim (Name Symbol) (Size Nat)]) :: [Dim (Name Symbol) (Size Nat)] where ...

Equations

TransposeIndexIndexDimsF index0 index1 dims = FromMaybe (TypeError (TransposeBy1Message ('ByIndex index1) dims)) (ReplaceDimImplF ('ByIndex index1) (FromMaybe (TypeError (TransposeBy0Message ('ByIndex index0) dims)) (ReplaceDimImplF ('ByIndex index0) dims (FromMaybe (TypeError (TransposeBy1Message ('ByIndex index1) dims)) (GetDimImplF ('ByIndex index1) dims)))) (FromMaybe (TypeError (TransposeBy0Message ('ByIndex index0) dims)) (GetDimImplF ('ByIndex index0) dims)))

sTranspose :: forall selectDim0 selectDim1 gradient layout device dataType shape shape' m. (shape' ~ TransposeF selectDim0 selectDim1 shape, MonadThrow m) => SSelectDim selectDim0 -> SSelectDim selectDim1 -> Tensor gradient layout device dataType shape -> m (Tensor gradient layout device dataType shape')

Returns a tensor that is a transposed version of input. The selected dimensions selectDim0 and selectDim1 are swapped.

>>> g <- sMkGenerator (SDevice SCPU) 0
>>> (input, _) <- sRandn (TensorSpec (SGradient SWithGradient) (SLayout SDense) (SDevice SCPU) (SDataType SFloat) (SShape $ SName @"batch" :&: SSize @10 :|: SName @"feature" :&: SSize @5 :|: SNil)) g
>>> output <- sTranspose (SSelectDim (SByName @"batch")) (SSelectDim (SByName @"feature")) input
>>> :type output
output
  :: Tensor
       ('Gradient 'WithGradient)
       ('Layout 'Dense)
       ('Device 'CPU)
       ('DataType 'Float)
       ('Shape
          '[ 'Dim ('Name "feature") ('Size 5),
             'Dim ('Name "batch") ('Size 10)])
>>> output <- sTranspose (SUncheckedSelectDim (ByIndex 0)) (SSelectDim (SByIndex @1)) input
>>> :type output
output
  :: Tensor
       ('Gradient 'WithGradient)
       ('Layout 'Dense)
       ('Device 'CPU)
       ('DataType 'Float)
       'UncheckedShape
>>> getDims output
[Dim {dimName = "feature", dimSize = 5},Dim {dimName = "batch", dimSize = 10}]

data TransposeError

Constructors

TransposeMixedSelectorsError

Fields

teBy0 :: By String Integer

teBy1 :: By String Integer

transpose :: forall selectDim0 selectDim1 gradient layout device dataType shape shape' m. (shape' ~ TransposeF selectDim0 selectDim1 shape, SingI selectDim0, SingI selectDim1, MonadThrow m) => Tensor gradient layout device dataType shape -> m (Tensor gradient layout device dataType shape')

type UnsqueezeByMessage (by :: By Symbol Nat) (dims :: [Dim (Name Symbol) (Size Nat)]) = "Cannot unsqueeze the tensor with the dimensions" % ("" % (((" '" <> dims) <> "'") % ("" % ("because the specified source dimension" % ("" % (((" '" <> by) <> "'") % ("" % "could not be found.")))))))

type family UnsqueezeF (selectDim :: SelectDim (By Symbol Nat)) (shape :: Shape [Dim (Name Symbol) (Size Nat)]) :: Shape [Dim (Name Symbol) (Size Nat)] where ...

Equations

UnsqueezeF _ 'UncheckedShape = 'UncheckedShape

UnsqueezeF 'UncheckedSelectDim _ = 'UncheckedShape

UnsqueezeF ('SelectDim ('ByName name)) ('Shape dims) = 'Shape (UnsqueezeIndexDimsF (FromMaybe (TypeError (UnsqueezeByMessage ('ByName name) dims)) (GetIndexByNameF name dims)) dims)

UnsqueezeF ('SelectDim ('ByIndex index)) ('Shape dims) = 'Shape (UnsqueezeIndexDimsF index dims)

type family UnsqueezeIndexDimsF (index :: Nat) (dims :: [Dim (Name Symbol) (Size Nat)]) :: [Dim (Name Symbol) (Size Nat)] where ...

Equations

UnsqueezeIndexDimsF index dims = FromMaybe (TypeError (UnsqueezeByMessage ('ByIndex index) dims)) (InsertDimImplF ('ByIndex index) dims ('Dim ('Name "*") ('Size 1)))

sUnsqueeze :: forall selectDim gradient layout device dataType shape shape'. shape' ~ UnsqueezeF selectDim shape => SSelectDim selectDim -> Tensor gradient layout device dataType shape -> Tensor gradient layout device dataType shape'

Unsqueezes a tensor with the specified dimension.

unsqueeze :: forall selectDim gradient layout device dataType shape shape'. (shape' ~ UnsqueezeF selectDim shape, SingI selectDim) => Tensor gradient layout device dataType shape -> Tensor gradient layout device dataType shape'

Unsqueezes a tensor with the specified dimension.

sExpand :: forall shape' gradient layout device dataType shape input output. (input ~ Tensor gradient layout device dataType shape, output ~ Tensor gradient layout device dataType (BroadcastShapesF shape shape')) => SShape shape' -> input -> output

Expands a tensor to the specified shape.

expand :: forall shape' gradient layout device dataType shape input output. (SingI shape', input ~ Tensor gradient layout device dataType shape, output ~ Tensor gradient layout device dataType (BroadcastShapesF shape shape')) => input -> output

Expands a tensor to the specified shape.

sSelect :: forall selectDim index gradient layout device dataType shapeIn shapeOut m. (index `InRangeF` GetDimF selectDim shapeIn, shapeOut ~ RemoveDimF selectDim shapeIn, SGetShape shapeIn, MonadThrow m) => SSelectDim selectDim -> SIndex index -> Tensor gradient layout device dataType shapeIn -> m (Tensor gradient layout device dataType shapeOut)

Slices the self tensor along the selected dimension at the given index. This function returns a view of the original tensor with the given dimension removed.

>>> nats <- sArangeNaturals (SGradient SWithoutGradient) (SLayout SDense) (SDevice SCPU) (SDataType SInt32) (SSize @8)
>>> input <- sReshape (SShape $ SName @"*" :&: SSize @4 :|: SName @"*" :&: SSize @2 :|: SNil) nats
>>> input
Tensor Int32 [4,2] [[ 0,  1],
                    [ 2,  3],
                    [ 4,  5],
                    [ 6,  7]]

index can be provided at compile-time:

>>> sSelect (SSelectDim (SByIndex @0)) (SIndex @1) input
Tensor Int32 [2] [ 2,  3]

index can also be provided at runtime:

>>> sSelect (SSelectDim (SByIndex @0)) (SUncheckedIndex 1) input
Tensor Int32 [2] [ 2,  3]

It produces a runtime error if the index is too large:

>>> sSelect (SSelectDim (SByIndex @0)) (SUncheckedIndex 10) input
*** Exception: IndexOutOfBoundError {ioobeIndex = 10, ioobeDim = Dim {dimName = "*", dimSize = 4}}

data IndexOutOfBoundError

Constructors

IndexOutOfBoundError

Fields

ioobeIndex :: Integer

ioobeDim :: Dim String Integer

select :: forall selectDim index gradient layout device dataType shapeIn shapeOut m. (SingI selectDim, SingI index, index `InRangeF` GetDimF selectDim shapeIn, shapeOut ~ RemoveDimF selectDim shapeIn, SGetShape shapeIn, MonadThrow m) => Tensor gradient layout device dataType shapeIn -> m (Tensor gradient layout device dataType shapeOut)