Torch.GraduallyTyped.Tensor.Type

Documentation

>>> import Data.Singletons.Prelude.List (SList (..))
>>> import Torch.GraduallyTyped

newtype Tensor (gradient :: Gradient RequiresGradient) (layout :: Layout LayoutType) (device :: Device (DeviceType Nat)) (dataType :: DataType DType) (shape :: Shape [Dim (Name Symbol) (Size Nat)]) where

A gradually typed tensor.

                         ┌─► Compute device, e.g. `'Device CPU
                         │
                         │               ┌─► List of dimensions, e.g. `'Shape '[ 'Dim 'UncheckedName ('Size 8), 'Dim 'UncheckedName ('Size 1) ]`
                         │               │
Tensor gradient layout device dataType shape
          │       │              │
          │       │              └─► Data type, e.g. `'DataType Float
          │       │
          │       └─► Memory layout, e.g. `'Layout Dense
          │
          └─► Whether or not the tensor requires a gradient, e.g. `'Gradient WithGradient for one that does

Constructors

UnsafeTensor :: forall gradient layout device dataType shape. ForeignPtr Tensor -> Tensor gradient layout device dataType shape
Unsafe constructor for tensors. Do not call this constructor directly, use smart constructors like ones or randn instead.

data TensorSpec (gradient :: Gradient RequiresGradient) (layout :: Layout LayoutType) (device :: Device (DeviceType Nat)) (dataType :: DataType DType) (shape :: Shape [Dim (Name Symbol) (Size Nat)]) where

Constructors

TensorSpec

Fields

:: forall gradient layout device dataType shape. { tsGradient :: SGradient gradient

, tsLayout :: SLayout layout

, tsDevice :: SDevice device

, tsDataType :: SDataType dataType

, tsShape :: SShape shape

} -> TensorSpec gradient layout device dataType shape

type UncheckedTensor = Tensor 'UncheckedGradient 'UncheckedLayout 'UncheckedDevice 'UncheckedDataType 'UncheckedShape

Alias for an untyped tensor without gradients.

type UncheckedParameter = Tensor ('Gradient 'WithGradient) 'UncheckedLayout 'UncheckedDevice 'UncheckedDataType 'UncheckedShape

Alias for an untyped tensor with gradients.

type CPUTensor = Tensor ('Gradient 'WithoutGradient) ('Layout 'Dense) ('Device 'CPU)

Alias for a tensor on CPU memory without gradients.

type CPUParameter = Tensor ('Gradient 'WithGradient) ('Layout 'Dense) ('Device 'CPU)

Alias for a tensor on CPU memory with gradients.

type SparseCPUTensor = Tensor ('Gradient 'WithoutGradient) ('Layout 'Sparse) ('Device 'CPU)

Alias for a sparse tensor on CPU memory without gradients.

type SparseCPUParameter = Tensor ('Gradient 'WithGradient) ('Layout 'Sparse) ('Device 'CPU)

Alias for a sparse tensor on CPU memory with gradients.

type CUDATensor deviceId = Tensor ('Gradient 'WithoutGradient) ('Layout 'Dense) ('Device ('CUDA deviceId))

Alias for a tensor on CUDA memory without gradients.

type CUDAParameter deviceId = Tensor ('Gradient 'WithGradient) ('Layout 'Dense) ('Device ('CUDA deviceId))

Alias for a tensor on CUDA memory with gradients.

type SparseCUDATensor deviceId = Tensor ('Gradient 'WithoutGradient) ('Layout 'Sparse) ('Device ('CUDA deviceId))

Alias for a sparse tensor on CUDA memory without gradients.

type SparseCUDAParameter deviceId = Tensor ('Gradient 'WithGradient) ('Layout 'Sparse) ('Device ('CUDA deviceId))

Alias for a sparse tensor on CUDA memory with gradients.

withoutGradient

Arguments

:: forall gradient layout device dataType shape. Tensor gradient layout device dataType shape
input tensor

-> IO (Tensor ('Gradient 'WithoutGradient) layout device dataType shape)
copy of the input tensor without gradient computations turned off.

Takes a tensor that may or may not require gradient computations and returns a copy that does not require them.

withGradient

Arguments

:: forall gradient layout device dataType shape. Tensor gradient layout device dataType shape
input tensor

-> IO (Tensor ('Gradient 'WithGradient) layout device dataType shape)
copy of the input tensor with gradient computations turned on.

Takes a tensor that does not requires gradient computations and returns a copy that requires them.

sSetGradient :: forall gradient gradient' layout device dataType shape. SGradient gradient -> Tensor gradient' layout device dataType shape -> IO (Tensor gradient layout device dataType shape)

Turn gradient computations off or on for a tensor.

class SGetGradient (gradient :: Gradient RequiresGradient) where

Minimal complete definition

sGetGradient

Methods

sGetGradient

Arguments

:: forall layout device dataType shape. Tensor gradient layout device dataType shape
input tensor

-> SGradient gradient
information about whether or not gradient computations are required

Returns the gradually typed information for whether or not gradient computations for the tensor are turned on.

>>> sOnes' gradient = sOnes $ TensorSpec gradient (SLayout SDense) (SDevice SCPU) (SDataType SFloat) (SShape $ SName @"batch" :&: SSize @32 :|: SName @"feature" :&: SSize @8 :|: SNil)
>>> t = sOnes' $ SGradient SWithGradient
>>> sGetGradient t
SGradient SWithGradient
>>> t = sOnes' $ SUncheckedGradient WithoutGradient
>>> sGetGradient t
SUncheckedGradient WithoutGradient

getRequiresGradient

Arguments

:: forall layout device dataType shape. Tensor gradient layout device dataType shape
input tensor

-> RequiresGradient
information about whether or not gradient computations are required

Returns the untyped memory layout of the input tensor.

>>> sOnes' gradient = sOnes $ TensorSpec gradient (SLayout SDense) (SDevice SCPU) (SDataType SFloat) (SShape $ SName @"batch" :&: SSize @32 :|: SName @"feature" :&: SSize @8 :|: SNil)
>>> t = sOnes' $ SGradient SWithGradient
>>> getRequiresGradient t
WithGradient
>>> t = sOnes' $ SUncheckedGradient WithoutGradient
>>> getRequiresGradient t
WithoutGradient

data GradientError

Constructors

GradientError

Fields

geExpected :: RequiresGradient

geActual :: RequiresGradient

sCheckedGradient

Arguments

:: forall gradient' m gradient layout device dataType shape. (SGetGradient gradient, MonadThrow m)

=> SGradient gradient'
layout

-> Tensor gradient layout device dataType shape
input tensor

-> m (Tensor (Seq (gradient <+> gradient') gradient') layout device dataType shape)
annotated output tensor wrapped in m

Checks whether or not the input tensor has the memory layout layout and returns a statically annotated copy of it wrapped in a MonadThrow m.

For instance, if m is Maybe, then the result will be wrapped in Just if and only if the tensor has indeed the memory layout layout. If it does not have it, then the result will be Nothing.

In the REPL, m will default to IO: >>> t = sOnes $ TensorSpec (SGradient SWithGradient) (SUncheckedLayout Dense) (SDevice SCPU) (SDataType SFloat) (SShape $ SName "batch" :&: SSize 32 :|: SName "feature" :&: SSize 8 :|: SNil) >>> t' <- sCheckedLayout (SLayout SDense) t >>> :type t' t' :: Tensor ('Gradient 'WithGradient) ('Layout 'Dense) ('Device 'CPU) ('DataType 'Float) ('Shape '[ 'Dim ('Name "batch") ('Size 32), 'Dim ('Name "feature") ('Size 8)]) >>> t' <- sCheckedLayout (SLayout SSparse) t *** Exception: LayoutError {leExpected = Sparse, leActual = Dense}

checkedGradient

Arguments

:: forall gradient' m gradient layout device dataType shape. (SingI gradient', SGetGradient gradient, MonadThrow m)

=> Tensor gradient layout device dataType shape
input tensor

-> m (Tensor (Seq (gradient <+> gradient') gradient') layout device dataType shape)
annotated output tensor wrapped in m

uncheckedGradient

Arguments

:: forall gradient layout device dataType shape. Tensor gradient layout device dataType shape
input tensor

-> Tensor 'UncheckedGradient layout device dataType shape
tensor without checked layout

Returns the input tensor but with UncheckedLayout as memory layout type annotation. Any static information about the tensor's memory layout is thus erased. However, the tensor's underlying data structure is not changed.

>>> t = ones @('Gradient 'WithGradient) @('Layout 'Dense) @('Device 'CPU) @('DataType 'Float) @('Shape '[ 'Dim ('Name "batch") ('Size 32), 'Dim ('Name "feature") ('Size 8)])
>>> :type uncheckedLayout t
uncheckedLayout t
  :: Tensor
       ('Gradient 'WithGradient)
       'UncheckedLayout
       ('Device 'CPU)
       ('DataType 'Float)
       ('Shape
          '[ 'Dim ('Name "batch") ('Size 32),
             'Dim ('Name "feature") ('Size 8)])

toDense

Arguments

:: forall m gradient layout device dataType shape. MonadThrow m

=> Tensor gradient layout device dataType shape
input

-> m (Tensor gradient ('Layout 'Dense) device dataType shape)
dense copy

Returns a dense copy of the tensor.

toSparse

Arguments

:: forall m gradient layout device dataType shape. MonadThrow m

=> Tensor gradient layout device dataType shape
input

-> m (Tensor gradient ('Layout 'Sparse) device dataType shape)
sparse copy

Returns a sparse copy of the tensor.

sSetLayout :: forall m gradient layout layout' device dataType shape. MonadThrow m => SLayout layout -> Tensor gradient layout' device dataType shape -> m (Tensor gradient layout device dataType shape)

Set the memory layout of a tensor to a given layout.

class SGetLayout (layout :: Layout LayoutType) where

Minimal complete definition

sGetLayout

Methods

sGetLayout

Arguments

:: forall gradient device dataType shape. Tensor gradient layout device dataType shape
input

-> SLayout layout
memory layout

Returns the gradually typed memory layout of the input tensor.

>>> sOnes' layout = sOnes $ TensorSpec (SGradient SWithGradient) layout (SDevice SCPU) (SDataType SFloat) (SShape $ SName @"batch" :&: SSize @32 :|: SName @"feature" :&: SSize @8 :|: SNil)
>>> t = sOnes' $ SLayout SDense
>>> sGetLayout t
SLayout SDense
>>> t = sOnes' $ SUncheckedLayout Dense
>>> sGetLayout t
SUncheckedLayout Dense

getLayoutType

Arguments

:: forall gradient device dataType shape. Tensor gradient layout device dataType shape
input

-> LayoutType
memory layout

Returns the untyped memory layout of the input tensor.

>>> sOnes' layout = sOnes $ TensorSpec (SGradient SWithGradient) layout (SDevice SCPU) (SDataType SFloat) (SShape $ SName @"batch" :&: SSize @32 :|: SName @"feature" :&: SSize @8 :|: SNil)
>>> t = sOnes' $ SLayout SDense
>>> getLayoutType t
Dense
>>> t = sOnes' $ SUncheckedLayout Dense
>>> getLayoutType t
Dense

data LayoutError

Constructors

LayoutError

Fields

leExpected :: LayoutType

leActual :: LayoutType

sCheckedLayout

Arguments

:: forall layout' m gradient layout device dataType shape. (SGetLayout layout, MonadThrow m)

=> SLayout layout'
layout

-> Tensor gradient layout device dataType shape
input tensor

-> m (Tensor gradient (Seq (layout <+> layout') layout') device dataType shape)
annotated output tensor wrapped in m

Checks whether or not the input tensor has the memory layout layout and returns a statically annotated copy of it wrapped in a MonadThrow m.

For instance, if m is Maybe, then the result will be wrapped in Just if and only if the tensor has indeed the memory layout layout. If it does not have it, then the result will be Nothing.

In the REPL, m will default to IO: >>> t = sOnes $ TensorSpec (SGradient SWithGradient) (SUncheckedLayout Dense) (SDevice SCPU) (SDataType SFloat) (SShape $ SName "batch" :&: SSize 32 :|: SName "feature" :&: SSize 8 :|: SNil) >>> t' <- sCheckedLayout (SLayout SDense) t >>> :type t' t' :: Tensor ('Gradient 'WithGradient) ('Layout 'Dense) ('Device 'CPU) ('DataType 'Float) ('Shape '[ 'Dim ('Name "batch") ('Size 32), 'Dim ('Name "feature") ('Size 8)]) >>> t' <- sCheckedLayout (SLayout SSparse) t *** Exception: LayoutError {leExpected = Sparse, leActual = Dense}

checkedLayout

Arguments

:: forall layout' m gradient layout device dataType shape. (SingI layout', SGetLayout layout, MonadThrow m)

=> Tensor gradient layout device dataType shape
input tensor

-> m (Tensor gradient (Seq (layout <+> layout') layout') device dataType shape)
annotated output tensor wrapped in m

uncheckedLayout

Arguments

:: forall gradient layout device dataType shape. Tensor gradient layout device dataType shape
input tensor

-> Tensor gradient 'UncheckedLayout device dataType shape
tensor without checked layout

Returns the input tensor but with UncheckedLayout as memory layout type annotation. Any static information about the tensor's memory layout is thus erased. However, the tensor's underlying data structure is not changed.

>>> t = ones @('Gradient 'WithGradient) @('Layout 'Dense) @('Device 'CPU) @('DataType 'Float) @('Shape '[ 'Dim ('Name "batch") ('Size 32), 'Dim ('Name "feature") ('Size 8)])
>>> :type uncheckedLayout t
uncheckedLayout t
  :: Tensor
       ('Gradient 'WithGradient)
       'UncheckedLayout
       ('Device 'CPU)
       ('DataType 'Float)
       ('Shape
          '[ 'Dim ('Name "batch") ('Size 32),
             'Dim ('Name "feature") ('Size 8)])

cpu

Arguments

:: forall m gradient layout device dataType shape. MonadThrow m

=> Tensor gradient layout device dataType shape
input

-> m (Tensor gradient layout ('Device 'CPU) dataType shape)
copy in CPU memory

Returns a copy of the tensor in CPU memory.

cuda

Arguments

:: forall m gradient layout device dataType shape. MonadThrow m

=> Tensor gradient layout device dataType shape
input

-> m (Tensor gradient layout ('Device ('CUDA 0)) dataType shape)
copy in CUDA memory

Returns a copy of the tensor in CUDA memory.

sSetDevice :: forall m gradient layout device device' dataType shape. MonadThrow m => SDevice device -> Tensor gradient layout device' dataType shape -> m (Tensor gradient layout device dataType shape)

Reallocates a tensor on the specified device.

class SGetDevice (device :: Device (DeviceType Nat)) where

Minimal complete definition

sGetDevice

Methods

sGetDevice

Arguments

:: forall gradient layout dataType shape. Tensor gradient layout device dataType shape
input

-> SDevice device
compute device of the input tensor

Returns the gradually typed compute device of the input tensor.

>>> ones' device = sOnes $ TensorSpec (SGradient SWithGradient) (SLayout SDense) device (SDataType SFloat) (SShape $ SName @"batch" :&: SSize @32 :|: SName @"feature" :&: SSize @8 :|: SNil)
>>> t = ones' $ SDevice SCPU
>>> sGetDevice t
SDevice SCPU
>>> t = ones' $ SUncheckedDevice CPU
>>> sGetDevice t
SUncheckedDevice CPU

getDeviceType

Arguments

:: forall gradient layout dataType shape. Tensor gradient layout device dataType shape
input

-> DeviceType Int16
compute device of the input tensor

Returns the untyped compute device of the input tensor.

>>> ones' device = sOnes $ TensorSpec (SGradient SWithGradient) (SLayout SDense) device (SDataType SFloat) (SShape $ SName @"batch" :&: SSize @32 :|: SName @"feature" :&: SSize @8 :|: SNil)
>>> t = ones' $ SDevice SCPU
>>> getDeviceType t
CPU
>>> t = ones' $ SUncheckedDevice CPU
>>> getDeviceType t
CPU

data DeviceError

Constructors

DeviceError

Fields

deExpected :: DeviceType Int16

deActual :: DeviceType Int16

sCheckedDevice

Arguments

:: forall device' m gradient layout device dataType shape. (SGetDevice device, MonadThrow m)

=> SDevice device'
device

-> Tensor gradient layout device dataType shape
input tensor

-> m (Tensor gradient layout (Seq (device <+> device') device') dataType shape)
annotated output tensor wrapped in m

Checks whether or not the input tensor is in the memory of device and returns a statically annotated copy of it wrapped in a MonadThrow m.

For instance, if m is Maybe, then the result will be wrapped in Just if and only if the tensor is indeed on device. If it is not, then the result will be Nothing.

In the REPL, m will default to IO: >>> t = sOnes $ TensorSpec (SGradient SWithGradient) (SLayout SDense) (SUncheckedDevice CPU) (SDataType SFloat) (SShape $ SName "batch" :&: SSize 32 :|: SName "feature" :&: SSize 8 :|: SNil) >>> t' <- sCheckedDevice (SDevice SCPU) t >>> :type t' t' :: Tensor ('Gradient 'WithGradient) ('Layout 'Dense) ('Device 'CPU) ('DataType 'Float) ('Shape '[ 'Dim ('Name "batch") ('Size 32), 'Dim ('Name "feature") ('Size 8)]) >>> t' <- sCheckedDevice (SDevice (SCUDA @0)) t *** Exception: DeviceError {deExpected = CUDA 0, deActual = CPU}

checkedDevice

Arguments

:: forall device' m gradient layout device dataType shape. (SingI device', SGetDevice device, MonadThrow m)

=> Tensor gradient layout device dataType shape
input tensor

-> m (Tensor gradient layout (Seq (device <+> device') device') dataType shape)
annotated output tensor wrapped in m

uncheckedDevice

Arguments

:: forall gradient layout device dataType shape. Tensor gradient layout device dataType shape
input tensor

-> Tensor gradient layout 'UncheckedDevice dataType shape
tensor without checked device

Returns the input tensor but with UncheckedDevice as device type annotation. Any static information about the tensor's device is thus erased. However, the tensor's underlying data structure is not changed.

>>> t = ones @('Gradient 'WithGradient) @('Layout 'Dense) @('Device 'CPU) @('DataType 'Float) @('Shape '[ 'Dim ('Name "batch") ('Size 32), 'Dim ('Name "feature") ('Size 8)])
>>> :type uncheckedDevice t
uncheckedDevice t
  :: Tensor
       ('Gradient 'WithGradient)
       ('Layout 'Dense)
       'UncheckedDevice
       ('DataType 'Float)
       ('Shape
          '[ 'Dim ('Name "batch") ('Size 32),
             'Dim ('Name "feature") ('Size 8)])

bool

Arguments

:: forall m gradient layout device dataType shape. MonadThrow m

=> Tensor gradient layout device dataType shape
input

-> m (Tensor ('Gradient 'WithoutGradient) layout device ('DataType 'Bool) shape)
Bool copy

Returns a copy of the tensor converted to Bool.

byte

Arguments

:: forall m gradient layout device dataType shape. MonadThrow m

=> Tensor gradient layout device dataType shape
input

-> m (Tensor ('Gradient 'WithoutGradient) layout device ('DataType 'UInt8) shape)
UInt8 copy

Returns a copy of the tensor converted to UInt8.

char

Arguments

:: forall m gradient layout device dataType shape. MonadThrow m

=> Tensor gradient layout device dataType shape
input

-> m (Tensor ('Gradient 'WithoutGradient) layout device ('DataType 'Int8) shape)
Int8 copy

Returns a copy of the tensor converted to Int8.

short

Arguments

:: forall m gradient layout device dataType shape. MonadThrow m

=> Tensor gradient layout device dataType shape
input

-> m (Tensor ('Gradient 'WithoutGradient) layout device ('DataType 'Int16) shape)
Int16 copy

Returns a copy of the tensor converted to Int16.

int

Arguments

:: forall m gradient layout device dataType shape. MonadThrow m

=> Tensor gradient layout device dataType shape
input

-> m (Tensor ('Gradient 'WithoutGradient) layout device ('DataType 'Int32) shape)
Int32 copy

Returns a copy of the tensor converted to Int32.

long

Arguments

:: forall m gradient layout device dataType shape. MonadThrow m

=> Tensor gradient layout device dataType shape
input

-> m (Tensor ('Gradient 'WithoutGradient) layout device ('DataType 'Int64) shape)
Int64 copy

Returns a copy of the tensor converted to Int64.

half

Arguments

:: forall m gradient layout device dataType shape. MonadThrow m

=> Tensor gradient layout device dataType shape
input

-> m (Tensor gradient layout device ('DataType 'Half) shape)
Half copy

Returns a copy of the tensor converted to the 16-bit floating point format Half.

float

Arguments

:: forall m gradient layout device dataType shape. MonadThrow m

=> Tensor gradient layout device dataType shape
input

-> m (Tensor gradient layout device ('DataType 'Float) shape)
Float copy

Returns a copy of the tensor converted to the 32-bit floating point format Float.

double

Arguments

:: forall m gradient layout device dataType shape. MonadThrow m

=> Tensor gradient layout device dataType shape
input

-> m (Tensor gradient layout device ('DataType 'Double) shape)
Double copy

Returns a copy of the tensor converted to the 32-bit floating point format Double.

sSetDataType :: forall m gradient layout device dataType dataType' shape. MonadThrow m => SDataType dataType -> Tensor gradient layout device dataType' shape -> m (Tensor gradient layout device dataType shape)

Set the data type of a tensor to the specified data type.

class SGetDataType (dataType :: DataType DType) where

Minimal complete definition

sGetDataType

Methods

sGetDataType

Arguments

:: forall gradient layout device shape. Tensor gradient layout device dataType shape
input

-> SDataType dataType
data type of the input tensor

Returns the gradually typed compute data type of the input tensor.

>>> sOnes' dataType = sOnes $ TensorSpec (SGradient SWithGradient) (SLayout SDense) (SDevice SCPU) dataType (SShape $ SName @"batch" :&: SSize @32 :|: SName @"feature" :&: SSize @8 :|: SNil)
>>> t = sOnes' $ SDataType SFloat
>>> sGetDataType t
SDataType SFloat
>>> t = sOnes' $ SUncheckedDataType Float
>>> sGetDataType t
SUncheckedDataType Float

getDType

Arguments

:: forall gradient layout device shape. Tensor gradient layout device dataType shape
input

-> DType
data type of the input tensor

Returns the untyped compute data type of the input tensor.

>>> sOnes' dataType = sOnes $ TensorSpec (SGradient SWithGradient) (SLayout SDense) (SDevice SCPU) dataType (SShape $ SName @"batch" :&: SSize @32 :|: SName @"feature" :&: SSize @8 :|: SNil)
>>> t = sOnes' $ SDataType SFloat
>>> dType t
Float
>>> t = sOnes' $ SUncheckedDataType Float
>>> dType t
Float

data DataTypeError

Constructors

DataTypeError

Fields

dtExpected :: DType

dtActual :: DType

sCheckedDataType

Arguments

:: forall dataType' m gradient layout device dataType shape. (SGetDataType dataType, MonadThrow m)

=> SDataType dataType'
data type

-> Tensor gradient layout device dataType shape
input tensor

-> m (Tensor gradient layout device (Seq (dataType <+> dataType') dataType') shape)
annotated output tensor wrapped in m

Checks whether or not the input tensor has the data type dataType and returns a statically annotated copy of it wrapped in a MonadThrow m.

For instance, if m is Maybe, then the result will be wrapped in Just if and only if the tensor has indeed the data type dataType. If it does not have it, then the result will be Nothing.

In the REPL, m will default to IO: >>> t = sOnes $ TensorSpec (SGradient SWithGradient) (SLayout SDense) (SDevice SCPU) (SUncheckedDataType Float) (SShape $ SName "batch" :&: SSize 32 :|: SName "feature" :&: SSize 8 :|: SNil) >>> t' <- checkedDataType ('DataType 'Float) t >>> :type t' t' :: Tensor ('Gradient 'WithGradient) ('Layout 'Dense) ('Device 'CPU) ('DataType 'Float) ('Shape '[ 'Dim ('Name "batch") ('Size 32), 'Dim ('Name "feature") ('Size 8)]) >>> t' <- checkedDataType ('DataType 'Double) t *** Exception: DataTypeError {dtExpected = Double, dtActual = Float}

checkedDataType

Arguments

:: forall dataType' m gradient layout device dataType shape. (SingI dataType', SGetDataType dataType, MonadThrow m)

=> Tensor gradient layout device dataType shape
input tensor

-> m (Tensor gradient layout device (Seq (dataType <+> dataType') dataType') shape)
annotated output tensor wrapped in m

uncheckedDataType

Arguments

:: forall gradient layout device dataType shape. Tensor gradient layout device dataType shape
input tensor

-> Tensor gradient layout device 'UncheckedDataType shape
tensor without checked data type

Returns the input tensor but with UncheckedDataType as data-type type annotation. Any static information about the tensor's data type is thus erased. However, the tensor's underlying data structure is not changed.

>>> t = ones @('Gradient 'WithGradient) @('Layout 'Dense) @('Device 'CPU) @('DataType 'Float) @('Shape '[ 'Dim ('Name "batch") ('Size 32), 'Dim ('Name "feature") ('Size 8)])
>>> :type uncheckedDataType t
uncheckedDataType t
  :: Tensor
       ('Gradient 'WithGradient)
       ('Layout 'Dense)
       ('Device 'CPU)
       'UncheckedDataType
       ('Shape
          '[ 'Dim ('Name "batch") ('Size 32),
             'Dim ('Name "feature") ('Size 8)])

class SGetShape (shape :: Shape [Dim (Name Symbol) (Size Nat)]) where

Minimal complete definition

sGetShape

Methods

sGetShape :: forall gradient layout device dataType. Tensor gradient layout device dataType shape -> SShape shape

Returns the gradually typed shape of the input tensor.

>>> sOnes' = sOnes . TensorSpec (SGradient SWithGradient) (SLayout SDense) (SDevice SCPU) (SDataType SFloat)
>>> t = sOnes' . SShape $ SName @"batch" :&: SSize @32 :|: SName @"feature" :&: SSize @8 :|: SNil
>>> sGetShape t
SShape (SCons (SDim {sDimName = SName, sDimSize = SSize}) (SCons (SDim {sDimName = SName, sDimSize = SSize}) SNil))
>>> t = sOnes' . SUncheckedShape $ [Dim "batch" 32, Dim "feature" 8]
>>> sGetShape t
SUncheckedShape [Dim {dimName = "batch", dimSize = 32},Dim {dimName = "feature", dimSize = 8}]
>>> t = sOnes' . SShape $ SUncheckedName "batch" :&: SUncheckedSize 32 :|: SUncheckedName "feature" :&: SSize @32 :|: SNil
>>> sGetShape t
SShape (SCons (SDim {sDimName = SUncheckedName "batch", sDimSize = SUncheckedSize 32}) (SCons (SDim {sDimName = SUncheckedName "feature", sDimSize = SSize}) SNil))
>>> t = sOnes' . SShape $ SName @"batch" :&: SUncheckedSize 32 :|: SName @"feature" :&: SUncheckedSize 8 :|: SNil
>>> sGetShape t
SShape (SCons (SDim {sDimName = SName, sDimSize = SUncheckedSize 32}) (SCons (SDim {sDimName = SName, sDimSize = SUncheckedSize 8}) SNil))

getDims :: forall gradient layout device dataType. Tensor gradient layout device dataType shape -> [Dim String Integer]

Returns the untyped shape of the input tensor.

>>> sOnes' = sOnes . TensorSpec (SGradient SWithGradient) (SLayout SDense) (SDevice SCPU) (SDataType SFloat)
>>> t = sOnes' . SShape $ SName @"batch" :&: SSize @32 :|: SName @"feature" :&: SSize @8 :|: SNil
>>> getDims t
[Dim {dimName = "batch", dimSize = 32},Dim {dimName = "feature", dimSize = 8}]
>>> t = sOnes' . SUncheckedShape $ [Dim "batch" 32, Dim "feature" 8]
>>> getDims t
[Dim {dimName = "batch", dimSize = 32},Dim {dimName = "feature", dimSize = 8}]
>>> t = sOnes' . SShape $ SUncheckedName "batch" :&: SUncheckedSize 32 :|: SUncheckedName "feature" :&: SSize @32 :|: SNil
>>> getDims t
[Dim {dimName = "batch", dimSize = 32},Dim {dimName = "feature", dimSize = 32}]
>>> t = sOnes' . SShape $ SName @"batch" :&: SUncheckedSize 32 :|: SName @"feature" :&: SUncheckedSize 8 :|: SNil
>>> getDims t
[Dim {dimName = "batch", dimSize = 32},Dim {dimName = "feature", dimSize = 8}]

class SGetDims (dims :: [Dim (Name Symbol) (Size Nat)]) where

Methods

sGetDims :: [String] -> [Integer] -> SList dims

dimsError :: forall a. a

dimNameError :: forall a. String -> String -> a

dimSizeError :: forall a b. Show a => a -> a -> b

dimNameSizeError :: forall a b. Show a => String -> String -> a -> a -> b

class SGetDim (dim :: Dim (Name Symbol) (Size Nat)) where

Methods

sGetDim :: String -> Integer -> SDim dim

data ShapeError

Constructors

ShapeError

Fields

seExpected :: [Dim String Integer]

seActual :: [Dim String Integer]

sCheckedShape

Arguments

:: forall shape' m gradient layout device dataType shape. (SGetShape shape, MonadThrow m)

=> SShape shape'
shape

-> Tensor gradient layout device dataType shape
input tensor

-> m (Tensor gradient layout device dataType (Seq (shape <+> shape') shape'))
annotated output tensor wrapped in m

Checks whether or not the input tensor has the shape shape and returns a statically annotated copy of it wrapped in a MonadThrow m.

For instance, if m is Maybe, then the result will be wrapped in Just if and only if the tensor has indeed the shape shape. If it is not, then the result will be Nothing.

In the REPL, m will default to IO: >>> t = sOnes $ TensorSpec (SGradient SWithGradient) (SLayout SDense) (SDevice SCPU) (SDataType SFloat) (SUncheckedShape [Dim "batch" 32, Dim "feature" 8]) >>> t' <- sCheckedShape (SShape $ SName "batch" :&: SSize 32 :|: SName "feature" :&: SSize 8 :|: SNil) t >>> :type t' t' :: Tensor ('Gradient 'WithGradient) ('Layout 'Dense) ('Device 'CPU) ('DataType 'Float) ('Shape '[ 'Dim ('Name "batch") ('Size 32), 'Dim ('Name "feature") ('Size 8)]) >>> t' <- sCheckedShape (SShape $ SUncheckedName "batch" :&: SSize 32 :|: SName "feature" :&: SUncheckedSize 8 :|: SNil) t >>> :type t' t' :: Tensor ('Gradient 'WithGradient) ('Layout 'Dense) ('Device 'CPU) ('DataType 'Float) ('Shape '[ 'Dim 'UncheckedName ('Size 32), 'Dim ('Name "feature") 'UncheckedSize]) >>> t' <- sCheckedShape (SShape $ SName "batch" :&: SSize 32 :|: SName @"feature" :&: SUncheckedSize 32 :|: SNil) t *** Exception: ShapeError {seExpected = [Dim {dimName = "batch", dimSize = 32},Dim {dimName = "feature", dimSize = 32}], seActual = [Dim {dimName = "batch", dimSize = 32},Dim {dimName = "feature", dimSize = 8}]}

checkedShape

Arguments

:: forall shape' m gradient layout device dataType shape. (SingI shape', SGetShape shape, MonadThrow m)

=> Tensor gradient layout device dataType shape
input tensor

-> m (Tensor gradient layout device dataType (Seq (shape <+> shape') shape'))
annotated output tensor wrapped in m

uncheckedDim

Arguments

:: forall selectDim gradient layout device dataType shape. Tensor gradient layout device dataType shape
input tensor

-> Tensor gradient layout device dataType (ReplaceDimF selectDim shape ('Dim 'UncheckedName 'UncheckedSize))
tensor with the selected dimensions unchecked

Returns the input tensor but with the selected dimension replaces with UncheckedDim as dimension type annotation. The static information about the selected tensor dimension is thus erased. However, the tensor's underlying data structure is not changed.

>>> t = ones @('Gradient 'WithGradient) @('Layout 'Dense) @('Device 'CPU) @('DataType 'Float) @('Shape '[ 'Dim ('Name "batch") ('Size 32), 'Dim ('Name "feature") ('Size 8)])
>>> :type uncheckedDim @('SelectDim ('ByName "batch")) t
uncheckedDim @('SelectDim ('ByName "batch")) t
  :: Tensor
       ('Gradient 'WithGradient)
       ('Layout 'Dense)
       ('Device 'CPU)
       ('DataType 'Float)
       ('Shape
          '[ 'Dim 'UncheckedName 'UncheckedSize,
             'Dim ('Name "feature") ('Size 8)])
>>> t = ones @('Gradient 'WithGradient) @('Layout 'Dense) @('Device 'CPU) @('DataType 'Float) @('Shape '[ 'Dim ('Name "batch") ('Size 32), 'Dim ('Name "feature") ('Size 8)])
>>> :type uncheckedDim @('SelectDim ('ByIndex 1)) t
uncheckedDim @('SelectDim ('ByIndex 1)) t
  :: Tensor
       ('Gradient 'WithGradient)
       ('Layout 'Dense)
       ('Device 'CPU)
       ('DataType 'Float)
       ('Shape
          '[ 'Dim ('Name "batch") ('Size 32),
             'Dim 'UncheckedName 'UncheckedSize])

uncheckedShape

Arguments

:: forall gradient layout device dataType shape. Tensor gradient layout device dataType shape
input tensor

-> Tensor gradient layout device dataType 'UncheckedShape
tensor without checked shape

Returns the input tensor but with UncheckedShape as shape type annotation. Any static information about the tensor's shape is thus erased. However, the tensor's underlying data structure is not changed.

>>> t = ones @('Gradient 'WithGradient) @('Layout 'Dense) @('Device 'CPU) @('DataType 'Float) @('Shape '[ 'Dim ('Name "batch") ('Size 32), 'Dim ('Name "feature") ('Size 8)])
>>> :type uncheckedShape t
uncheckedShape t
  :: Tensor
       ('Gradient 'WithGradient)
       ('Layout 'Dense)
       ('Device 'CPU)
       ('DataType 'Float)
       'UncheckedShape

gitHubErrorMsg :: String

isContiguous :: Tensor gradient layout device dataType shape -> Bool

contiguous :: Tensor gradient layout device dataType shape -> Tensor gradient layout device dataType shape

withTensor :: Tensor gradient layout device dataType shape -> (Ptr () -> IO a) -> IO a

class TensorLikeRaw a where

Methods

guessDim

Arguments

:: Maybe a
value Nothing if the data type wrapping a is empty.

-> Maybe Int
dimension Nothing if a is a scalar.

Guesses outer dim.

>>> guessDim @[[Int]] $ pure [[1, 2], [3, 4], [5, 6]]
Just 3

guessInnerDims

Arguments

:: MonadThrow m

=> Maybe a
value Nothing if the data type wrapping a is empty.

-> m [Int]
inner dimensions

Guesses inner dims.

>>> guessInnerDims @[[Int]] $ pure [[1, 2], [3, 4], [5, 6]]
[2]

tensorPeekElemOff

Arguments

:: Ptr ()
pointer to tensor

-> Int
offset

-> [Int]
tensor dimensions

-> IO a
value

Reads a value from a tensor.

tensorPokeElemOff

Arguments

:: Ptr ()
pointer to tensor

-> Int
offset

-> [Int]
tensor dimensions

-> a
value

-> IO ()

Writes a value to a tensor.

guessDims :: forall a m. (TensorLikeRaw a, MonadThrow m) => Maybe a -> m [Int]

Guesses dims: concatenates guessDim with guessInnerDims.

>>> guessDims @[[Int]] $ pure [[1, 2], [3, 4], [5, 6]]
[3,2]

unexpectedDimsError :: forall a m b. (TensorLikeRaw a, MonadThrow m) => [Int] -> Maybe a -> m b

class TensorLike a (dType :: DType) (dims :: [Dim (Name Symbol) (Size Nat)]) | a -> dims, a -> dType where

Methods

sToTensor :: forall gradient layout device m. MonadThrow m => SGradient gradient -> SLayout layout -> SDevice device -> a -> m (Tensor gradient layout device ('DataType dType) ('Shape dims))

Creates a tensor from a TensorLike value.

>>> t <- sToTensor (SGradient SWithoutGradient) (SLayout SDense) (SDevice SCPU) ([(1, 2), (3, 4), (5, 6)] :: [(Int, Int)])
>>> t
Tensor Int64 [3,2] [[ 1,  2],
                    [ 3,  4],
                    [ 5,  6]]
>>> :type t
t :: Tensor
       ('Gradient 'WithoutGradient)
       ('Layout 'Dense)
       ('Device 'CPU)
       ('DataType 'Int64)
       ('Shape
          '[ 'Dim ('Name "*") 'UncheckedSize, 'Dim ('Name "*") ('Size 2)])

fromTensor :: forall gradient layout device. Tensor gradient layout device ('DataType dType) ('Shape dims) -> a

Creates a TensorLike from a tensor.

toTensor :: forall gradient layout device a dType dims m. (TensorLike a dType dims, SingI gradient, SingI layout, SingI device, MonadThrow m) => a -> m (Tensor gradient layout device ('DataType dType) ('Shape dims))

Non-singleton version of sToTensor.

sToTensorRaw :: forall gradient layout device a dType dims m. (TensorLike a dType dims, TensorLikeRaw a, SingI dType, MonadThrow m) => SGradient gradient -> SLayout layout -> SDevice device -> a -> m (Tensor gradient layout device ('DataType dType) ('Shape dims))

fromTensorRaw :: forall gradient layout device a dType dims. (TensorLike a dType dims, TensorLikeRaw a, SGetDims dims) => Tensor gradient layout device ('DataType dType) ('Shape dims) -> a

data DimMismatchError

Constructors

DimMismatchError

Fields

dmeFirst :: [Int]

dmeOther :: [Int]

checkDims :: MonadThrow m => [Int] -> [Int] -> m ()

sSetTensorOptions :: forall gradient layout device dataType gradientFrom layoutFrom deviceFrom dataTypeFrom shape. SGradient gradient -> SLayout layout -> SDevice device -> SDataType dataType -> Tensor gradientFrom layoutFrom deviceFrom dataTypeFrom shape -> IO (Tensor gradient layout device dataType shape)

setTensorOptions :: forall gradient layout device dataType gradientFrom layoutFrom deviceFrom dataTypeFrom shape. (SingI gradient, SingI layout, SingI device, SingI dataType) => Tensor gradientFrom layoutFrom deviceFrom dataTypeFrom shape -> IO (Tensor gradient layout device dataType shape)

Orphan instances

Castable (HList l) [ForeignPtr Tensor] => Castable (HList l) (ForeignPtr TensorList)

Instance details

Methods

cast :: HList l -> (ForeignPtr TensorList -> IO r) -> IO r

uncast :: ForeignPtr TensorList -> (HList l -> IO r) -> IO r

Castable (HList ('[] :: [k])) [ForeignPtr Tensor]

Instance details

Methods

cast :: HList '[] -> ([ForeignPtr Tensor] -> IO r) -> IO r

uncast :: [ForeignPtr Tensor] -> (HList '[] -> IO r) -> IO r